PNML

Documentation for PNML.

PNML.INCLUDEXML

PNML.default_pntd_map

PNML.pnmltype_map

PNML.tagmap

PNML.AbstractHLCore

PNML.AbstractPnmlCore

PNML.Document

PNML.HLCore

PNML.Maybe

PNML.MissingIDException

PNML.PTNet

PNML.PnmlCore

PNML.PnmlType

PNML.SimpleNet

PNML.SimpleNet

PNML.add_label!

PNML.add_nettype!

PNML.add_tool!

PNML.allchildren

PNML.attribute_attributes

PNML.attribute_content

PNML.attribute_elem

PNML.extract_pnml

PNML.find_nets

PNML.firstchild

PNML.has_place

PNML.includexml!

PNML.node_summary

PNML.parse_and

PNML.parse_arbitraryoperator

PNML.parse_arbitrarysort

PNML.parse_bool

PNML.parse_booleanconstant

PNML.parse_condition

PNML.parse_declaration

PNML.parse_declarations

PNML.parse_doc

PNML.parse_equality

PNML.parse_file

PNML.parse_graphics

PNML.parse_graphics_coordinate

PNML.parse_imply

PNML.parse_inequality

PNML.parse_label

PNML.parse_mulitsetsort

PNML.parse_name

PNML.parse_namedoperator

PNML.parse_net

PNML.parse_node

PNML.parse_node_pair

PNML.parse_not

PNML.parse_or

PNML.parse_page

PNML.parse_pnml

PNML.parse_pnml_common!

PNML.parse_pnml_label_common!

PNML.parse_pnml_node_common!

PNML.parse_productsort

PNML.parse_str

PNML.parse_structure

PNML.parse_term

PNML.parse_text

PNML.parse_tokengraphics

PNML.parse_tokenposition

PNML.parse_toolspecific

PNML.parse_tuple

PNML.parse_unparsed

PNML.parse_useroperator

PNML.parse_usersort

PNML.parse_variable

PNML.parse_variabledecl

PNML.place

PNML.pnml_common_defaults

PNML.pnml_label_defaults

PNML.pnml_node_defaults

PNML.to_net_type

PNML.to_net_type_sym

PNML.validate

PNML.validate_node

PNML.validate_pnml

PNML.@pnml_str

PNML.@xml_str

PNML.INCLUDEXML — Constant

Include the XML as part of data.

PNML.Maybe — Type

maybe of type T or nothing

PNML.default_pntd_map — Constant

default_pntd_map

Map from Petri Net Type Definition (pntd) URI to Symbol. The URI used can be any string regardless of any violation of the PNML Specification. There is a companion map pnmltype_map that takes the symbol to a type object. The 'pntd symbol' should match the name used in the URI with inconvinient characters removed or replaced. For example, '-' is replaced by '_'.

PNML.pnmltype_map — Constant

pnmltype_map

The keys are the supported kinds of Petri Nets.

Provides a place to abstract relationship of pntd name and implementation type. Allows multiple strings to map to the same parser implementation. Is a point at which different parser implmentations may be introduced.

PNML.tagmap — Constant

Map XML tag names to parser functions.

PNML.AbstractHLCore — Type

Base of High Level Petri Net pntds.

PNML.AbstractPnmlCore — Type

Most minimal Petri Net type that is the foundation of all pntd.

PNML.Document — Type

Wrap the collection of PNML nets from a single XML tree.

PNML.HLCore — Type

High-Level Petri Nets add large extensions to core.

PNML.MissingIDException — Type

Use exception to allow dispatch and additional data presentation to user.

PNML.PTNet — Type

Place-Transition Petri Nets add small extensions to core.

PNML.PnmlCore — Type

All Petri Nets support core.

PNML.PnmlType — Type

abstract type PnmlType

Abstract root of a dispatch type based on Petri Net Type Definition (pntd).

Each Petri Net Markup Language (PNML) network element will have a single pntd URI as a required type XML attribute. That URI could/should refer to a RelaxNG schema defining the syntax and semantics of the XML model.

See (`pnmltype_map)[@ref] for the map from type string to a dispatch singleton.

Within PNML.jl no schema-level validation is done. Nor is any use made of the schema within the code. Schemas, UML, ISO Specification and papers used to inform the design. See https://www.pnml.org/ for details.

In is allowed by the PNML specification to omit validation with the presumption that some specialized, external tool can be applied, thus allowing the file format to be used for inter-tool communication with lower overhead.

Some pnml files exist that do not use a valid type URI. However it is done, an appropriate subtype of PnmlType must be chosen. Refer to to_net_type and pnmltype_map for how to get from the URI string to a Julia type.

PNML.SimpleNet — Type

SimpleNet wraps the 'place', 'transition' & "arc" collections of a single page.

Omits the page level of the pnml-defined hierarchy, and all labels at the merged net/page level of pnml.

TODO: Support labels? Some, all, non-standard?

A multi-page net can be collpsed by removing referenceTransitions & referencePlaces, and merging labels of net and all pages.

PNML.SimpleNet — Method

Given a PnmlDict of a pnml 'net' element, assume there is only one page.

PNML.add_label! — Method

Add node tod[:labels]. Return updated d[:labels].

PNML.add_nettype! — Method

Add or replace mapping from symbol s to nettype dispatch singleton t.

PNML.add_tool! — Method

Add node tod[:tools]. Return updated d[:tools].

PNML.allchildren — Function

Return vector of 'elelement's immediate children withtag`.

PNML.attribute_attributes — Method

Return vector of pairs.

PNML.attribute_content — Method

Return PnmlDict with values that are vectors when there are multiple instances of a tag in 'nv' and scalar otherwise.

PNML.attribute_elem — Method

attribute_elem(node)

Return PnmlDict after debug print of nodename. If element node has any children, each is placed in the dictonary with the tag name symbol as the key, repeated tags produce a vector as the value. Any XML attributes found are added as as key,value. to the tuple returned.

Note that this will recursivly decend the well-formed XML, transforming the the children into vector NamedTuples & Dicts.

Note the assumption that children and content are mutually exclusive. Content is always a leaf element. However XML attributes can be anywhere in the hiearchy.

PNML.extract_pnml — Function

extract_pnml()

given a filename, EzXML.Document, or EzXML.Node returns all of the Pnml nodes.

PNML.find_nets — Function

Return nets of 'd' matching the given pntd 'type'.

PNML.firstchild — Function

Return up to 1 immediatechild of elementelthat is atag`.

PNML.has_place — Method

Is there any place with 'id' in net 's'?

PNML.includexml! — Method

Set value of key :xml based on global variable.

PNML.node_summary — Function

node_summary([io::IO], node; n=5, pp=EzXML.prettyprint)

Pretty print the first n lines of the XML node. If io is not supplied, prints to the default output stream stdout. pp can be any pretty print method that takes (io::IO, node).

PNML.parse_and — Method

PNML.parse_arbitraryoperator — Method

PNML.parse_arbitrarysort — Method

PNML.parse_bool — Method

PNML.parse_booleanconstant — Method

PNML.parse_condition — Method

Annotation label of transition nodes. Meaning it can have text, graphics, et al.

PNML.parse_declaration — Method

Attribute label of 'net' and 'page' nodes

PNML.parse_declarations — Method

 parse_declarations(node)

Return NamedTuple with :contents holding a vector of parsed child elements.

PNML.parse_doc — Method

Start descent from the root XML element node of doc. A well formed PNML XML document has a single root node: 'pnml'.

PNML.parse_equality — Method

PNML.parse_file — Method

Build pnml from a file.

PNML.parse_graphics — Method

parse_graphics

Arcs, Annotations and Nodes (places, transitions, pages) have different graphics semantics. Return a dictonary with the union of possibilities.

PNML.parse_graphics_coordinate — Method

Coordinates x, y are in points.

PNML.parse_imply — Method

PNML.parse_inequality — Method

PNML.parse_label — Method

Should not often have a 'label' tag, this will bark if one is found. Return named tuple (tag,node), used to defer parsing the xml while matching usage of PnmlDict that has at least the :tag and :xml keys.

PNML.parse_mulitsetsort — Method

PNML.parse_name — Method

Return named tuple with pnml name text and optional tool & GUI information.

PNML.parse_namedoperator — Method

PNML.parse_net — Method

parse_net(node)

Return a dictonary of the pnml net with keys matching their XML tag names.

PNML.parse_node — Method

parse_node(node;verbose=true)

Take a node and parse it by calling the method matching node.name from tagmap if mapping exists, otherwise call attribute_elem. verbose is a boolean controlling debug logging.

PNML.parse_node_pair — Method

PNML.parse_not — Method

PNML.parse_or — Method

PNML.parse_page — Method

parse_page(node)

PNML requires at least on page.

PNML.parse_pnml — Method

parse_pnml(node)

Start parse from the pnml root node of the well formed XML document. Return a a named tuple containing vector of pnml petri nets.

PNML.parse_pnml_common! — Method

parse_pnml_common(s, node)

Update d with graphics, tools, label children of pnml node and label elements. Used by parsepnmlnodecommonlabel ! & parsepnmllabelcommon!. Adds, graphics, tools, labels. Note that "lables" are the everything else option and this should be called after parsing any elements that has an expected tags.

PNML.parse_pnml_label_common! — Method

parse_pnml_label_common!(d, node)

Update d with 'text' and 'structure' children of node, defering other tags to parse_pnml_common!.

PNML.parse_pnml_node_common! — Method

parse_pnml_node_common!(d, node)

Update d with name children, defering other tags to parse_pnml_common!.

PNML.parse_productsort — Method

PNML.parse_str — Method

Build pnml from a string.

PNML.parse_structure — Method

Return dictonary including a vector of child content elements. A pnml structure can possibly hold any well formed XML. Structure will vary based on parent element and petri net type definition of the net. #TODO: Specialized structure parsers are needed. 2nd pass parser?

PNML.parse_term — Method

parse_term(n)

PNML.parse_text — Method

Return the striped string of text child's nodecontent in a named tuple.

PNML.parse_tokengraphics — Method

High-level place-transition nets have a toolspecific structure defined for token graphics.

PNML.parse_tokenposition — Method

Position is relative to containing element. Units are points.

PNML.parse_toolspecific — Method

Return tuple with tag name, tool & version attributes and xml node. Anyone that can parse the nodecontents may specialize on tool & version.

PNML.parse_tuple — Method

PNML.parse_unparsed — Method

PNML.parse_useroperator — Method

PNML.parse_usersort — Method

PNML.parse_variable — Method

PNML.parse_variabledecl — Method

PNML.place — Method

Return the place with 'id' in net 's'.

PNML.pnml_common_defaults — Method

Return Dict of tags common to both pnml nodes and pnml labels.

PNML.pnml_label_defaults — Method

Merge xs into dictonary with default pnml label tags. Used on pnml tags below a pnmlnode tag. Label level tags include: name, inscription, initialMarking. Notable differences from [`pnmlnode_defaults`](@ref): text, structure, no name tag.

PNML.pnml_node_defaults — Method

Merge xs into dictonary with default pnml node tags. Used on: net, page ,place, transition, arc. Usually default value will be nothing or empty vector.

PNML.to_net_type — Function

to_net_type(uri)
to_net_type(symbol)

Map either a text string or a symbol to a dispatch type singlton.

While that string may be a URI for a pntd, we treat it as a simple string without parsing. The pnmltypemap and pntdmap are both assumed to be correct here.

Unknown or empty uri will map to symbol :pnmlcore as part of the logic. Unknown symbol returns nothing.

PNML.to_net_type_sym — Method

to_net_type_sym(uri)

We map uri to a symbol using a dictionary like default_pntd_map. Return symbol that is a valid pnmltype_map key. Defaults to :pnmlcore.

PNML.validate — Method

Log a warning if s is not a known Petri Net Markup Language schema/pntd.

PNML.validate_node — Method

PNML.validate_pnml — Method

Check the <pnml> element against TODO.

Return TODO

PNML.@pnml_str — Macro

@pnml_str(s)

utility macro for parsing xml strings into symbolics

PNML.@xml_str — Macro

@xml_str(s)

utility macro for parsing xml strings into node