PnmlTypes

PnmlTypes

Types

Constants

Functions

Macros

PNML.PnmlTypes — Module

Kind of Petri Net. Petri Net Type Definition (pntd) URI mapped to PnmlType subtype singleton.

Imports

Base

Core

DocStringExtensions

PNML

Exports

AbstractContinuousNet

AbstractHLCore

AbstractPnmlCore

ContinuousNet

HLCore

HLPNG

OpenNet

PTNet

PT_HLPNG

PnmlCore

PnmlType

StochasticNet

SymmetricNet

TimedNet

pnmltype

pntd_symbol

Types

PNML.PnmlTypes.AbstractContinuousNet — Type

abstract type AbstractContinuousNet <: PnmlType

Uses floating point numbers for markings, inscriptions, and conditions.

PNML.PnmlTypes.AbstractHLCore — Type

Base of High Level Petri Net pntds. See SymmetricNet, PT_HLPNG and others.

abstract type AbstractHLCore <: PnmlType

PNML.PnmlTypes.AbstractPnmlCore — Type

Base of token/integer-based Petri Net pntds. See PnmlCore, PTNet and others.

abstract type AbstractPnmlCore <: PnmlType

PNML.PnmlTypes.ContinuousNet — Type

TODO: Continuous Petri Net Concrete type.

struct ContinuousNet <: AbstractContinuousNet

PNML.PnmlTypes.HLCore — Type

High-Level Petri Net Graphs (HLPNGs) add large extensions to core. HLCore can be used for generic high-level nets. We try to implement and test all function at PnmlCore level, but expect to find use for a concrete type at this level.

struct HLCore <: AbstractHLCore

PNML.PnmlTypes.HLPNG — Type

HLPNG is the most intricate High-Level Petri Net schema. It extends SymmetricNet, including with

declarations for sorts and functions (ArbitraryDeclarations)

sorts for Integer, String, and List

struct HLPNG <: AbstractHLCore

PNML.PnmlTypes.OpenNet — Type

TODO: Open Petri Net

struct OpenNet <: AbstractContinuousNet

PNML.PnmlTypes.PTNet — Type

Place-Transition Petri Nets add small extensions to core. The grammer file is ptnet.pnml so we name it PTNet. Note that 'PT' is often the prefix for XML tags specialized for this net type.

struct PTNet <: AbstractPnmlCore

PNML.PnmlTypes.PT_HLPNG — Type

Place-Transition Net in HLCore notation (HLPNG=High-Level Petri Net Graph).

struct PT_HLPNG <: AbstractHLCore

PNML.PnmlTypes.PnmlCore — Type

The most minimal concrete Petri Net. Used to implement and test the complete Petri Net Graph infrastructure. Labels of the graph is where meaning is attached. Much of the Label infrastructure for High Level Petri Net Graphs is tested at this level. Subtypes of PnmlType should be used to specialize Labels for expressiveness and optimization.

struct PnmlCore <: AbstractPnmlCore

PNML.PnmlTypes.PnmlType — Type

Abstract root of a dispatch type based on Petri Net Type Definition (pntd).

Each Petri Net Markup Language (PNML) network element will have a single pntd URI as a required 'type' XML attribute. That URI should refer to a RelaxNG schema defining the syntax and semantics of the XML model.

Selected abbreviations, URIs that do not resolve to a valid schema file, are also allowed.

Refer to pntd_symbol and pnmltype for how to get from the URI to a singleton.

abstract type PnmlType

PNML.PnmlTypes.StochasticNet — Type

TODO: Stochastic Petri Net

struct StochasticNet <: AbstractHLCore

PNML.PnmlTypes.SymmetricNet — Type

Symmetric Petri Net is the best-worked use case in the primer and ISO specification part 2.

struct SymmetricNet <: AbstractHLCore

PNML.PnmlTypes.TimedNet — Type

TODO: Timed Petri Net

struct TimedNet <: AbstractHLCore

Constants

Docstrings for constants. This includes type aliases.

PNML.PnmlTypes.default_pntd_map — Constant

Map from Petri Net Type Definition (pntd) URI to Symbol. Allows multiple strings to map to the same pntd.

There is a companion map pnmltype_map that takes the symbol to a type object.

The URI is a string and may be the full URL of a pntd schema, just the schema file name, or a placeholder for a future schema.

Examples

The 'pntd symbol' should match the name used in the URI with inconvinient characters removed or replaced. For example, '-' is replaced by '_'.

PNML.PnmlTypes.pnmltype_map — Constant

The key Symbols are the supported kinds of Petri Nets.

Functions

Docstrings for functions.

PNML.PnmlTypes.add_nettype! — Method

add_nettype!(
    dict::AbstractDict,
    s::Symbol,
    pntd::PnmlType
) -> PnmlType

Add or replace mapping from symbol s to nettype dispatch singleton t.

PNML.PnmlTypes.pnmltype — Function

pnmltype(pntd::T; kw...)
pnmltype(uri::AbstractString; kw...)
function pnmltype(s::Symbol; pnmltype_map=pnmltype_map, kw...)

Map either a text string or a symbol to a dispatch type singlton.

While that string may be a URI for a pntd, we treat it as a simple string without parsing. The PnmlTypes.pnmltype_map and PnmlTypes.default_pntd_map are both assumed to be correct here.

Unknown or empty uri will map to symbol :pnmlcore. Unknown symbol throws a PNML.MalformedException

Examples

julia> pnmltype(PnmlCore())
PnmlCore()

julia> pnmltype("nonstandard")
PnmlCore()

julia> pnmltype(:symmetric)
SymmetricNet()

PNML.PnmlTypes.pntd_symbol — Method

pntd_symbol(s::String) -> Symbol

Map string s to a pntd symbol using default_pntd_map. Any unknown s is mapped to :pnmlcore. Returned symbol is suitable for pnmltype to use to index into pnmltype_map.

Examples

julia> PnmlTypes.pntd_symbol("foo")
:pnmlcore

Macros

Docstrings for macros.