Library Reference

PNML.PNML — Module

PNML

!!! PNML.jl is still a work-in-progress.

Petri Net Markup Language, is an XML-based format. PNML.jl reads a pnml model and emits an intermediate representation (IR).

Petri Net Type Definitions (pntd) are defined using RelaxNG XML Schema files. It is possibly to create a non-standard pntd. Since validation is not a goal of PNML.jl, non-standard pntds can be used.

Note that ISO is working on part 3 of the PNML standard covering pntd (as of October 2021). It is for interchange of pnml models between different tools.

The intermediate representation (IR) represents the XML tree via named tuples, dictionaries, LabelledArrays (there is an interface duck here). The tags of the XML are used as keys as much as possible.

What is accepted as values is often a superset of what the pntd specifies. This can be thought of as duck-typing. Conforming to the pntd is not the role of the IR.

SimpleNet

Created to be a end-to-end use case. And explore implementing something-that-works while building upon and improving the IR. Does not try to conform to any standard. Much of the complexity possible with pnml is ignored.

The first use is to recreate the lotka-volterra model from Petri.jl examples. Find it in the examples folder. This is a stochastic Petri Net.

Liberties are taken with pnml, remember that standards-checking is not a goal. A less-simple consumer of the IR can impose standards-checking.

TODO

Features that have not been started:

Write pnml file

Update pnml model

Create pnml model

Features that are not complete:

pnml high-level marking, inscription, condition

pntd schemas

  - specialize where? 
  - do not highjack tags, only use standard semantics
* parsing of sort, terms, declarations
* Symbolics support (is this where PnmlType is useful?)

toolspecific usage example

*

Acknowledgments

MathML.jl

Its function map architecture was adopted and per(permute|perverse|use)d in PNML.j.

PNML.INCLUDEXML — Constant

Include the XML as part of data.

PNML.Maybe — Type

Maybe of type T or nothing.

PNML.applymap — Constant

PNML.default_pntd_map — Constant

Map from Petri Net Type Definition (pntd) URI to Symbol. The URI used can be any string regardless of any violation of the PNML Specification. There is a companion map pnmltype_map that takes the symbol to a type object. The 'pntd symbol' should match the name used in the URI with inconvinient characters removed or replaced. For example, '-' is replaced by '_'.

PNML.pnmltype_map — Constant

The keys are the supported kinds of Petri Nets.

Provides a place to abstract relationship of pntd name and implementation type. Allows multiple strings to map to the same parser implementation. Is a point at which different parser implmentations may be introduced.

PNML.tagmap — Constant

Map XML tag names to parser functions.

PNML.verifymap — Constant

Map tag names to verification methods.

PNML.AbstractHLCore — Type

abstract type AbstractHLCore <: PNML.AbstractPnmlCore

Base of High Level Petri Net pntds.

PNML.AbstractPnmlCore — Type

abstract type AbstractPnmlCore <: PNML.PnmlType

Most minimal Petri Net type that is the foundation of all pntd.

PNML.Document — Type

struct Document{N, X}

nets::Any

xml::Any

reg::PNML.IDRegistry

Wrap the collection of PNML nets from a single XML tree.

PNML.HLCore — Type

struct HLCore <: PNML.AbstractHLCore

High-Level Petri Nets add large extensions to core. HLCore can be used for generic high-leve nets.

PNML.HLNet — Type

struct HLNet <: PNML.AbstractHLCore

HLNet is the most intricate High-Level Petri Net schema

PNML.HLPetriNet — Type

struct HLPetriNet{T} <: PNML.PetriNet

id::Symbol

net::Dict{Symbol, Any}

PNML.HLPetriNet — Method

HLPetriNet(net::Dict{Symbol, Any}) -> PNML.HLPetriNet{_A} where _A

PNML.IDRegistry — Type

mutable struct IDRegistry

ids::Set{Symbol}

lk::ReentrantLock

Holds a set of pnml id symbols and a lock to allow safe reentrancy.

PNML.MalformedException — Type

struct MalformedException <: PnmlException

msg::String

node::EzXML.Node

PNML.MissingIDCounter — Type

mutable struct MissingIDCounter

i::Int64

lk::ReentrantLock

PNML.MissingIDException — Type

struct MissingIDException <: PnmlException

msg::String

node::EzXML.Node

Use exception to allow dispatch and additional data presentation to user.

PNML.OpenNet — Type

struct OpenNet <: PNML.AbstractHLCore

Open Petri Net

PNML.PTNet — Type

struct PTNet <: PNML.AbstractPnmlCore

Place-Transition Petri Nets add small extensions to core.

PNML.PT_HLPNG — Type

struct PT_HLPNG <: PNML.AbstractHLCore

Place-Transition High-Level Petri Net Graph

PNML.PetriNet — Type

abstract type PetriNet

Abstract type providing 2nd-level parsing of the intermediate representation of a single network in a PNML.Document.

Extended

Subtypes should map directly and simply to PnmlType.

Since a PNML.Document can contain multiple networks it is possible that a higher-level will create multiple PNML.PetriNet instances, each a different subtype.

Pages are used for visual layout for humans. They can be merged into one page without loosing any Petri Net semantics. Often we will only work with merged pages.

Interface

#TODO define type for network IR: Wrap a single tag net's PnmlDict? We start a description of the net IR here.

XML <net> tags are pnml nodes. These nodes are parsed into PnmlDict with keys

key	value description

tag	XML tag name is standard in the IR

id	unique ID

name	text name, optional

tools	set of tool specific - possibly empty

labels	set of generic "pnml labels" - possible empty

type	PnmlType defines schema the XML should meet

declarations	defines high-level semantics of a net

pages	set of pages - not empty

See pnml_common_defaults, pnml_node_defaults and parse_net for more detail.

XML <page> tags are also parsed into PnmlDict

key	value description

tag	XML tag name is standard in the IR

id	unique ID

name	text name, optional

tools	set of tool specific - possibly empty

labels	set of generic "pnml labels" - possible empty

places	set of places

trans	set of transitions

arcs	set of arcs

refP	references to place on different page

refT	references to transition on different page

declarations	only net & page tags have declarations

See parse_page, parse_net

PNML.PnmlCfg — Type

struct PnmlCfg

Contains configuration data. #TODO add something

PNML.PnmlCore — Type

struct PnmlCore <: PNML.AbstractPnmlCore

PnmlCore is the most minimal concrete Petri Net.

PNML.PnmlDict — Type

mutable struct Dict{Symbol, Any} <: AbstractDict{Symbol, Any}

Alias for Dict with expected value types for PNML intermediate representation.

PNML.PnmlException — Type

abstract type PnmlException <: Exception

PNML.PnmlType — Type

abstract type PnmlType

Abstract root of a dispatch type based on Petri Net Type Definition (pntd).

Each Petri Net Markup Language (PNML) network element will have a single pntd URI as a required type XML attribute. That URI could/should refer to a RelaxNG schema defining the syntax and semantics of the XML model.

See (`pnmltype_map)[@ref] for the map from type string to a dispatch singleton.

Within PNML.jl no schema-level validation is done. Nor is any use made of the schema within the code. Schemas, UML, ISO Specification and papers used to inform the design. See https://www.pnml.org/ for details.

In is allowed by the PNML specification to omit validation with the presumption that some specialized, external tool can be applied, thus allowing the file format to be used for inter-tool communication with lower overhead.

Some pnml files exist that do not use a valid type URI. However it is done, an appropriate subtype of PnmlType must be chosen. Refer to to_net_type and pnmltype_map for how to get from the URI string to a Julia type.

PNML.SimpleNet — Type

struct SimpleNet{P, T, A, RT, RP, L, TI} <: PNML.PetriNet

id::Symbol

Same as the XML attribute of the same name.

place::Any

transition::Any

arc::Any

refT::Any

refP::Any

label::Any

tools::Any

SimpleNet wraps the place, transition & arc collections of a single page of one net.

Omits the page level of the pnml-defined hierarchy by collapsing down to one page. A multi-page net can be collpsed by removing referenceTransitions & referencePlaces, and merging pages into the first page. Only selected fields are merged.

PNML.StochasticNet — Type

struct StochasticNet <: PNML.AbstractHLCore

Stochastic Petri Net

PNML.SymmetricNet — Type

struct SymmetricNet <: PNML.AbstractHLCore

Symmetric Petri Net

PNML.TimedNet — Type

struct TimedNet <: PNML.AbstractHLCore

Timed Petri Net

Base.isempty — Method

isempty(reg::PNML.IDRegistry) -> Bool

PNML.add_label! — Method

add_label!(d::Dict{Symbol, Any}, node; kwargs...) -> Vector{Dict{Symbol, Any}}

Add node tod[:labels]. Return updated d[:labels].

PNML.add_nettype! — Method

add_nettype!(d::AbstractDict, s::Symbol, t::PNML.PnmlType) -> PNML.PnmlType

Add or replace mapping from symbol s to nettype dispatch singleton t.

PNML.add_tool! — Method

add_tool!(d::Dict{Symbol, Any}, node; kwargs...) -> Vector{Dict{Symbol, Any}}

Add node tod[:tools]. Return updated d[:tools].

PNML.all_arcs — Method

all_arcs(s::PNML.SimpleNet, id::Symbol) -> Any

Return vector of arcs that have a source or target of transition id.

PNML.allchildren — Function

allchildren(tag, el) -> Vector{EzXML.Node}
allchildren(tag, el, ns) -> Vector{EzXML.Node}

Return vector of 'elelement's immediate children withtag`.

PNML.arc — Method

arc(s::PNML.SimpleNet, id::Symbol) -> Any

PNML.arc_ids — Method

arc_ids(s::PNML.SimpleNet) -> Any

PNML.arcs — Method

arcs(s::PNML.SimpleNet) -> Any

PNML.attribute_content — Method

attribute_content(nv::Vector{EzXML.Node}; kwargs...) -> Dict{Symbol, Any}

Return PnmlDict with values that are vectors when there are multiple instances of a tag in 'nv' and scalar otherwise.

PNML.attribute_elem — Method

attribute_elem(node; kwargs...)

Return PnmlDict after debug print of nodename. If element node has any children, each is placed in the dictonary with the tag name symbol as the key, repeated tags produce a vector as the value. Any XML attributes found are added as as key,value. to the tuple returned.

Note that this will recursivly decend the well-formed XML.

Note the assumption that children and content are mutually exclusive. Content is always a leaf element. However XML attributes can be anywhere in the hiearchy.

Example

julia> using PNML, EzXML

julia> node = parse_node(xml"""<aaa id="FOO">BAR</aaa>"""; reg=PNML.IDRegistry());

PNML.collapse_pages! — Function

Merge page content into the 1st page of each pnml net. Note that refrence nodes are still present. They can be removed later with deref!.

PNML.compress — Function

Copy PnmlDict with keys removed when paired with value of nothing.

PNML.compress! — Method

compress!(d::Array{T<:AbstractDict, 1}) -> Vector{T} where T<:AbstractDict

In-place PnmlDict keys removed when they are paired with value of nothing.

PNML.condition — Function

Return condition value of transition.

PNML.conditions — Function

Return a vector of condition values for net s.

PNML.deref! — Function

Remove reference nodes from arcs. Design intent expects collapse_pages! to have been applied so that everything is on one page.

Examples

Axioms

All ids in a PNML.Document are unique in that they only have one instance in the XML.

A chain of reference Places or Transitions always ends at a Place or Transition.

All ids are valid.

PNML.deref_place — Method

deref_place(s::PNML.SimpleNet, id::Symbol) -> Any

Return id of referenced place.

PNML.deref_transition — Method

deref_transition(s::PNML.SimpleNet, id::Symbol) -> Any

Return id of referenced transition.

PNML.duplicate_id_action — Method

duplicate_id_action(id::Symbol)

Use a global configuration to choose what to do when a duplicated pnml node id has been detected. Default is to do nothing.

PNML.find_nets — Function

Return nets of d matching the given pntd type.

PNML.first_net — Method

first_net(d::PNML.Document) -> Any

Return first net contained by d.

PNML.firstchild — Function

firstchild(tag, el) -> Union{Nothing, EzXML.Node}
firstchild(tag, el, ns) -> Union{Nothing, EzXML.Node}

Return up to 1 immediatechild of elementelthat is atag`.

PNML.has_arc — Method

has_arc(s::PNML.SimpleNet, id::Symbol) -> Any

PNML.has_place — Method

has_place(s::PNML.SimpleNet, id::Symbol) -> Any

Is there any place with id in net s?

PNML.has_refP — Method

has_refP(s::PNML.SimpleNet, id::Symbol) -> Any

PNML.has_refT — Method

has_refT(s::PNML.SimpleNet, id::Symbol) -> Any

PNML.has_transition — Method

has_transition(s::PNML.SimpleNet, id::Symbol) -> Any

PNML.id — Method

id(node::Dict{Symbol, Any}) -> Symbol

PNML.id — Method

id(s::PNML.PetriNet) -> Any

Return id field of s.

PNML.in_out — Method

in_out(s::PNML.SimpleNet, t::Symbol) -> Tuple{LabelledArrays.LArray, LabelledArrays.LArray}

Return tuple of input, output labelled vectors with key of place ids and value of arc inscription's value.

PNML.includexml — Method

includexml(node)

Set value of key :xml based on global configuration/control variable.

PNML.inscription — Function

Return incription value of arc.

PNML.isregistered — Method

isregistered(reg::PNML.IDRegistry, s::AbstractString) -> Bool

PNML.marking — Function

Return marking value of a place p.

Examples

julia> using PNML

julia> p = Dict(:marking => Dict(:value=>nothing));

julia> PNML.marking(p)
0

julia> p = Dict(:marking => Dict(:value=>12.34));

julia> PNML.marking(p)
12.34

PNML.nets — Method

nets(d::PNML.Document) -> Any

Return all nets of d.

PNML.next_missing_id — Method

next_missing_id(c::PNML.MissingIDCounter) -> Int64

PNML.node_summary — Function

Pretty print the first n lines of the XML node. If io is not supplied, prints to the default output stream stdout. pp can be any pretty print method that takes (io::IO, node).

PNML.number_value — Method

Parse string as a number. First try integer then float.

PNML.parse_and — Method

parse_and(node; kwargs...)

PNML.parse_arbitraryoperator — Method

parse_arbitraryoperator(node; kwargs...)

PNML.parse_arbitrarysort — Method

parse_arbitrarysort(node; kwargs...)

PNML.parse_arc — Method

parse_arc(node; kwargs...)

PNML.parse_bool — Method

parse_bool(node; kwargs...)

PNML.parse_booleanconstant — Method

parse_booleanconstant(node; kwargs...)

PNML.parse_condition — Method

parse_condition(node; kwargs...) -> Dict{Symbol, Any}

Annotation label of transition nodes. Meaning it can have text, graphics, et al.

PNML.parse_declaration — Method

parse_declaration(node; kwargs...)

Attribute label of 'net' and 'page' nodes.

PNML.parse_declarations — Method

parse_declarations(node; kwargs...)

PNML.parse_doc — Method

parse_doc(doc::EzXML.Document) -> PNML.Document{_A, _B} where {_A, _B}

Return a PNML.Document built from an XML Doncuent. A well formed PNML XML document has a single root node: 'pnml'.

PNML.parse_equality — Method

parse_equality(node; kwargs...)

PNML.parse_file — Method

parse_file(fn) -> PNML.Document{_A, _B} where {_A, _B}

Build pnml from a file.

PNML.parse_graphics — Method

parse_graphics(node; kwargs...) -> Dict{Symbol, Any}

Arcs, Annotations and Nodes (places, transitions, pages) have different graphics semantics. Return a dictonary with the union of possibilities.

PNML.parse_graphics_coordinate — Method

parse_graphics_coordinate(node; kwargs...) -> Dict{Symbol, Any}

Coordinates x, y are in points.

PNML.parse_graphics_fill — Method

parse_graphics_fill(node; kwargs...) -> Dict{Symbol, Any}

PNML.parse_graphics_font — Method

parse_graphics_font(node; kwargs...) -> Dict{Symbol, Any}

PNML.parse_graphics_line — Method

parse_graphics_line(node; kwargs...) -> Dict{Symbol, Any}

PNML.parse_hlinitialMarking — Method

parse_hlinitialMarking(node; kwargs...) -> Dict{Symbol, Any}

PNML.parse_hlinscription — Method

parse_hlinscription(node; kwargs...) -> Dict{Symbol, Any}

PNML.parse_imply — Method

parse_imply(node; kwargs...)

PNML.parse_inequality — Method

parse_inequality(node; kwargs...)

PNML.parse_initialMarking — Method

parse_initialMarking(node; kwargs...) -> Dict{Symbol, Any}

PNML.parse_inscription — Method

parse_inscription(node; kwargs...) -> Dict{Symbol, Any}

PNML.parse_label — Method

parse_label(node; kwargs...) -> Dict{Symbol, Any}

Should not often have a 'label' tag, this will bark if one is found. Return minimal PnmlDict holding (tag,node), to defer parsing the xml.

PNML.parse_mulitsetsort — Method

parse_mulitsetsort(node; kwargs...)

PNML.parse_name — Method

Return PnmlDict pnml name text and optional tool & GUI information.

PNML.parse_namedoperator — Method

parse_namedoperator(node; kwargs...)

PNML.parse_net — Method

parse_net(node; kwargs...)

Return a dictonary of the pnml net with keys matching their XML tag names.

PNML.parse_node — Method

parse_node(node; verbose, kwargs...)

Take a node and parse it by calling the method matching node.name from tagmap if that mapping exists, otherwise call attribute_elem. verbose is a boolean controlling debug logging.

PNML.parse_not — Method

parse_not(node; kwargs...)

PNML.parse_or — Method

parse_or(node; kwargs...)

PNML.parse_page — Method

parse_page(node; kwargs...)

PNML requires at least on page.

PNML.parse_place — Method

parse_place(node; kwargs...)

PNML.parse_pnml — Method

parse_pnml(node; kwargs...)

Start parse from the pnml root node of the well formed XML document. Return a a named tuple containing vector of pnml petri nets.

PNML.parse_pnml_common! — Method

parse_pnml_common!(d::Dict{Symbol, Any}, node; kwargs...)

Update d WITH graphics, tools, label children of pnml node and label elements. Used by parse_pnml_node_common! & parse_pnml_label_common!. Adds, graphics, tools, labels. Note that "lables" are the everything else option and this should be called after parsing any elements that has an expected tags.

PNML.parse_pnml_label_common! — Method

parse_pnml_label_common!(d::Dict{Symbol, Any}, node; kwargs...)

Update d with 'text' and 'structure' children of node, defering other tags to parse_pnml_common!.

PNML.parse_pnml_node_common! — Method

parse_pnml_node_common!(d::Dict{Symbol, Any}, node; kwargs...)

Update d with name children, defering other tags to parse_pnml_common!.

PNML.parse_productsort — Method

parse_productsort(node; kwargs...)

PNML.parse_refPlace — Method

parse_refPlace(node; kwargs...)

PNML.parse_refTransition — Method

parse_refTransition(node; kwargs...)

PNML.parse_sort — Method

parse_sort(node; kwargs...)

PNML.parse_str — Method

parse_str(str) -> PNML.Document{_A, _B} where {_A, _B}

Build pnml from a string.

PNML.parse_structure — Method

parse_structure(node; kwargs...)

A pnml structure node can hold any well formed XML. Structure semantics will vary based on parent element and petri net type definition of the net. #TODO: Specialized structure parsers are needed. 2nd pass parser?

PNML.parse_term — Method

parse_term(node; kwargs...)

PNML.parse_text — Method

parse_text(node; kwargs...) -> Dict{Symbol, Any}

"Return the stripped string of text child's nodecontent as :content key of PnmlDict.

PNML.parse_tokengraphics — Method

parse_tokengraphics(node; kwargs...) -> Dict{Symbol, Any}

High-level place-transition nets (HL-PTNet) have a toolspecific structure defined for token graphics. Contains <tokenposition> tags.

PNML.parse_tokenposition — Method

parse_tokenposition(node; kwargs...) -> Dict{Symbol, Any}

Position is a coordinate relative to containing element. Units are points.

PNML.parse_toolspecific — Method

parse_toolspecific(node; kwargs...) -> Dict{Symbol, Any}

Return PnmlDict with tag name, tool & version attributes and content parsed. Anyone that can parse the :content may specialize on tool & version.

The content can be any well-formed xml. We use our usual parsing mechanism, which can be enhanced if someone makes a good case.

PNML.parse_transition — Method

parse_transition(node; kwargs...)

PNML.parse_tuple — Method

parse_tuple(node; kwargs...)

PNML.parse_type — Method

parse_type(node; kwargs...)

These type of a place is different from net type or pntd. Used to define the "sort" of tokens held by the place and semantics of the marking.

PNML.parse_unparsed — Method

parse_unparsed(node; kwargs...)

PNML.parse_useroperator — Method

parse_useroperator(node; kwargs...)

PNML.parse_usersort — Method

parse_usersort(node; kwargs...)

PNML.parse_variable — Method

parse_variable(node; kwargs...)

PNML.parse_variabledecl — Method

parse_variabledecl(node; kwargs...)

PNML.place — Method

place(s::PNML.SimpleNet, id::Symbol) -> Any

Return the place with id in net s.

PNML.place_ids — Method

place_ids(s::PNML.SimpleNet) -> Any

Return vector of place ids in s.

PNML.places — Method

places(s::PNML.SimpleNet) -> Any

PNML.pnml_common_defaults — Method

pnml_common_defaults(node) -> Dict{Symbol, Any}

Return Dict of tags common to both pnml nodes and pnml labels. See pnml_label_defaults and pnml_node_defaults.

PNML.pnml_label_defaults — Method

pnml_label_defaults(node, xs...) -> Dict{Symbol, Any}

Merge xs into dictonary with default pnml label tags. Used on pnml tags below a pnmlnode tag. Label level tags include: name, inscription, initialMarking. Notable differences from [`pnmlnodedefaults](@ref): text, structure, no name tag. See [pnmlcommon_defaults`](@ref).

PNML.pnml_node_defaults — Method

pnml_node_defaults(node, xs...) -> Any

Merge xs into dictonary with default pnml node tags. Used on: net, page ,place, transition, arc. Usually default value will be nothing or empty vector. See pnml_label_defaults and pnml_common_defaults.

PNML.pntd — Method

pntd(s::AbstractString) -> Symbol

Map s to a pntd symbol. Any unknown s is mapped to pnmlcore.

PNML.refplace — Method

refplace(s::PNML.SimpleNet, id::Symbol) -> Any

PNML.refplace_ids — Method

refplace_ids(s::PNML.SimpleNet) -> Any

PNML.refplaces — Method

refplaces(s::PNML.SimpleNet) -> Any

PNML.reftransition — Method

reftransition(s::PNML.SimpleNet, id::Symbol) -> Any

PNML.reftransition_ids — Method

reftransition_ids(s::PNML.SimpleNet) -> Any

PNML.reftransitions — Method

reftransitions(s::PNML.SimpleNet) -> Any

PNML.register_id! — Method

register_id!(reg::PNML.IDRegistry, s::AbstractString) -> Symbol

Register id symbol and return the symbol.

PNML.reset_registry! — Method

reset_registry!(reg::PNML.IDRegistry) -> Set{Symbol}

Empty the set of id symbols. Use case is unit tests. In normal use it should never be needed.

PNML.source — Method

source(arc) -> Symbol

PNML.src_arcs — Method

src_arcs(s::PNML.SimpleNet, id::Symbol) -> Any

Return vector of arcs that have a source of transition id.

PNML.target — Method

target(arc) -> Symbol

PNML.tgt_arcs — Method

tgt_arcs(s::PNML.SimpleNet, id::Symbol) -> Any

Return vector of arcs that have a target of transition id.

PNML.to_net_type — Function

Map either a text string or a symbol to a dispatch type singlton.

While that string may be a URI for a pntd, we treat it as a simple string without parsing. The pnmltypemap and pntdmap are both assumed to be correct here.

Unknown or empty uri will map to symbol :pnmlcore as part of the logic. Unknown symbol returns nothing.

PNML.to_net_type_sym — Method

to_net_type_sym(uri::AbstractString; pntd_map) -> Symbol

We map uri to a symbol using a dictionary like default_pntd_map. Return symbol that is a valid pnmltype_map key. Defaults to :pnmlcore.

PNML.transition — Method

transition(s::PNML.SimpleNet, id::Symbol) -> Any

PNML.transition_function — Function

Transition function of a Petri Net. Each transition has an input vector and an output vector. Each labelled vector is indexed by the place on the other end of the arc. Values are inscriptions.

PNML.transition_ids — Method

transition_ids(s::PNML.SimpleNet) -> Any

PNML.transitions — Method

transitions(s::PNML.SimpleNet) -> Any

PNML.validate — Method

validate(s::Symbol; pnmltype_map) -> Symbol

Log a warning if s is not a known Petri Net Markup Language schema/pntd.

PNML.validate_arc — Method

validate_arc(node)

PNML.validate_net — Method

validate_net(node)

PNML.validate_node — Method

validate_node(node) -> Any

PNML.validate_page — Method

validate_page(node)

PNML.validate_place — Method

validate_place(node)

PNML.validate_pnml — Method

validate_pnml(node)

Check the <pnml> element against TODO. Return TODO

PNML.validate_transition — Method

validate_transition(node)

PNML.@xml_str — Macro

Utility macro for parsing xml strings into node.