PNML Reference

Types

Constants

Functions

Macros

PNML Reference

PNML.PNML — Module

PNML

Petri Net Markup Language, is an XML-based format. PNML.jl reads a pnml model and emits an intermediate representation (IR).

The intermediate representation (IR) represents the XML tree via julia data structures: dictionaries, NamedTuples, LabelledArrays, strings, numbers, objects, vectors. The exact mixture changes as the project continues.

The tags of the XML are used as keys and names as much as possible.

What is accepted as values is often a superset of what the pntd schema specifies. This can be thought of as duck-typing. Conforming to the pntd is not the role of the IR.

On top of the IR is (will be) implemented Petri Net adaptions and interpertations. This is the level that pntd conformance can be imposed. Adaption to julia packages for graphs, agents, and composing into the greater hive-mind.

TODO

Features that have not been started:

Write pnml file

Update pnml model

Create pnml model

Graphs.jl

Features that are not complete:

pnml high-level marking, inscription, condition

pntd schemas- specialize where?

* Symbolics support (is this where PnmlType is useful?)

toolspecific usage example

Imports

AbstractTrees

AutoHashEquals

Base

Core

DocStringExtensions

EzXML

LabelledArrays

PNML.PnmlTypes

PrettyPrinting

Reexport

Exports

AbstractContinuousCore

AbstractHLCore

ContinuousNet

HLCore

HLNet

MalformedException

MissingIDException

OpenNet

PTNet

PT_HLPNG

PnmlCore

PnmlDict

PnmlException

PnmlType

PnmlTypes

StochasticNet

SymmetricNet

TimedNet

parse_file

parse_node

parse_pnml

parse_str

pnmltype

pntd_symbol

@xml_str

Types

PNML.AbstractDeclaration — Type

Declarations are the core of high-level Petri Net. They define objects/names that are used for conditions, inscriptions, markings. They are attached to PNML nets and pages.

abstract type AbstractDeclaration <: HLAnnotation

PNML.AbstractLabel — Type

abstract type AbstractLabel

Labels are attached to the Petri Net Graph object subtypes. See PnmlObject.

PNML.AbstractPnmlTool — Type

abstract type AbstractPnmlTool

Tool specific objects can be attached to PnmlObjects and AbstractLabels subtypes.

PNML.AbstractTerm — Type

abstract type AbstractTerm

Terms are part of the multi-sorted algebra that is part of High-Level Petri Net markings, inscriptions and conditions.

See also AbstractDeclaration.

PNML.Annotation — Type

abstract type Annotation <: AbstractLabel

Label that may be displayed. It differs from an Attribute Label by possibly having a Graphics field.

PNML.AnyElement — Type

struct AnyElement

tag::Symbol

dict::Dict{Symbol, Any}

xml::EzXML.Node

Wrap PnmlDict holding well-formed XML. See ToolInfo and PnmlLabel.

PNML.Arc — Type

Edge of a Petri Net Markup Language graph that connects place and transition.

mutable struct Arc{PNTD<:PnmlType, ITYPE} <: PnmlObject

pntd::PnmlType

id::Symbol

source::Symbol

target::Symbol

inscription::Any

name::Union{Nothing, Name}

com::ObjectCommon

PNML.BuiltInSort — Type

Wrap a AnyElement. Use until specialized/cooked.

From the 'primer': built-in sorts of Symmetric Nets are the following: Booleans, range of integers, finite enumerations, cyclic enumerations and dots

PNML.Condition — Type

Label of a Transition that determines when the transition fires.

struct Condition{T} <: HLAnnotation

text::Union{Nothing, String}

term::Any

com::ObjectCommon

Examples

julia> c = Condition()
Condition(nothing, true, )

julia> c()
true

julia> c = Condition(false)
Condition(nothing, false, )

julia> c()
false

julia> c = Condition("xx", false)
Condition("xx", false, )

julia> c()
false

PNML.Condition — Method

Evaluate a Condition instance.

PNML.Coordinate — Type

Cartesian Coordinate.

struct Coordinate{T<:Number}

x::Number

y::Number

PNML.Declaration — Type

Label of a net or page that holds zero or more AbstractDeclaration.

struct Declaration <: HLAnnotation

declarations::Vector{Any}

com::ObjectCommon

PNML.Fill — Type

Fill attributes as strings.

struct Fill

color::Union{Nothing, String}

image::Union{Nothing, String}

gradient_color::Union{Nothing, String}

gradient_rotation::Union{Nothing, String}

PNML.Font — Type

Font attributes as strings.

struct Font

family::Union{Nothing, String}

style::Union{Nothing, String}

weight::Union{Nothing, String}

size::Union{Nothing, String}

align::Union{Nothing, String}

rotation::Union{Nothing, String}

decoration::Union{Nothing, String}

PNML.Graphics — Type

PNML Graphics can be attached to many parts of PNML models.

struct Graphics

dimension::Union{Nothing, Coordinate}

fill::Union{Nothing, Fill}

font::Union{Nothing, Font}

line::Union{Nothing, Line}

offset::Union{Nothing, Coordinate}

position::Union{Nothing, Vector{Coordinate}}

PNML.HLAnnotation — Type

abstract type HLAnnotation <: AbstractLabel

Annotation label that uses <text> and <structure>.

PNML.HLInscription — Type

struct HLInscription{TermType<:PNML.AbstractTerm} <: HLAnnotation

text::Union{Nothing, String}

term::Union{Nothing, TermType} where TermType<:PNML.AbstractTerm

Any <structure> must be a many-sorted algebra term for a <hlinscription> annotation label.

com::ObjectCommon

Labels an Arc.

Examples

julia> i = HLInscription()
HLInscription(nothing, Term(:empty, Dict()), )

julia> i()
1

julia> i = HLInscription(Term(:term, PnmlDict(:value=>3)))
HLInscription(nothing, Term(:term, Dict(:value => 3)), )

julia> i()
3

julia> i = HLInscription("text", Term())
HLInscription("text", Term(:empty, Dict()), )

julia> i()
1

julia> i = HLInscription("text", Term(:term, PnmlDict(:value=>3)))
HLInscription("text", Term(:term, Dict(:value => 3)), )

julia> i()
3

PNML.HLInscription — Method

Evaluate a HLInscription. Returns a value of the same sort as TBD.

PNML.HLLabel — Type

struct HLLabel <: HLAnnotation

text::Union{Nothing, String}

structure::Union{Nothing, PNML.Structure}

com::ObjectCommon

xml::EzXML.Node

High-level pnml labels are expected to have <text> and <structure> elements. This concrete type is for "unclaimed" labels in a high-level petri net. Some "claimed" HLAnnotation labels are Condition, Declaration, HLMarking, HLInscription.

PNML.HLMarking — Type

Label of a Place in a High-level Petri Net Graph. See AbstractHLCore, Term. #TODO Term as wrapper of PnmlDict should be replaced (someday).

Is a functor that evaluates the term.

struct HLMarking{TermType} <: HLAnnotation

text::Union{Nothing, String}

term::Union{Nothing, TermType} where TermType

Any <structure> must be a many-sorted algebra term for a <hlmarking> annotation label.

com::ObjectCommon

Examples

julia> m = HLMarking("the text", Term(:term, PnmlDict(:value=>3)))
HLMarking("the text", Term(:term, Dict(:value => 3)), )

julia> m()
3

PNML.HLMarking — Method

Evaluate a HLMarking instance. Returns a value of the same sort as its Place. #TODO How to ensure sort type?

PNML.HLPetriNet — Type

Wrap a single pnml net.

struct HLPetriNet{PNTD} <: PetriNet{PNTD}

net::PnmlNet{PNTD} where PNTD

Details

PNML.HLPetriNet — Method

Construct from string of valid pnml XML using the first network

PNML.IDRegistry — Type

Holds a set of pnml id symbols and a lock to allow safe reentrancy.

struct IDRegistry

ids::Set{Symbol}

lk::ReentrantLock

PNML.Line — Type

Line attributes as strings.

struct Line

color::Union{Nothing, String}

shape::Union{Nothing, String}

style::Union{Nothing, String}

width::Union{Nothing, String}

PNML.MalformedException — Type

struct MalformedException <: PnmlException

msg::String

node::EzXML.Node

PNML.MissingIDException — Type

struct MissingIDException <: PnmlException

msg::String

node::EzXML.Node

Use exception to allow dispatch and additional data presentation to user.

PNML.MultisetSort — Type

Wrap a AnyElement. Use until specialized/cooked.

PNML.Name — Type

struct Name <: AbstractLabel

text::String

graphics::Union{Nothing, Graphics}

tools::Union{Nothing, Vector{ToolInfo}}

Name is for display, possibly in a tool specific way.

PNML.NamedOperator — Type

struct NamedOperator{V, T} <: PNML.OperatorDeclaration

id::Symbol

name::String

parameter::Vector

def::Any

PNML.NamedSort — Type

struct NamedSort{S} <: PNML.SortDeclaration

id::Symbol

name::String

def::Any

PNML.ObjectCommon — Type

struct ObjectCommon

graphics::Union{Nothing, Graphics}

tools::Union{Nothing, Vector{ToolInfo}}

labels::Union{Nothing, Vector{PnmlLabel}}

Common infrastructure shared by PNML objects and labels. Some optional incidental bits are shared by most PNML objects are also collected here.

PNML.PTInscription — Type

struct PTInscription{T<:Number} <: PNML.Annotation

value::Number

com::ObjectCommon

Labels an Arc.

Examples

julia> i = PTInscription()
PTInscription(1, )

julia> i()
1

julia> i = PTInscription(3)
PTInscription(3, )

julia> i()
3

PNML.PTInscription — Method

Evaluate a PTInscription.

PNML.PTMarking — Type

Number-valued label of Place. See PTNet, ContinuousNet.

Is a functor that returns the value.

struct PTMarking{N<:Number} <: PNML.Annotation

value::Number

com::ObjectCommon

Examples

julia> m = PTMarking()
PTMarking(0, )

julia> m()
0

julia> m = PTMarking(1)
PTMarking(1, )

julia> m()
1

julia> m = PTMarking(12.34)
PTMarking(12.34, )

julia> m()
12.34

PNML.PTMarking — Method

Evaluate a PTMarking instance.

PNML.Page — Type

struct Page{PNTD<:PnmlType, D} <: PnmlObject

id::Symbol

places::Vector{Place}

refPlaces::Vector{RefPlace}

transitions::Vector{Transition}

refTransitions::Vector{RefTransition}

arcs::Vector{Arc}

declaration::Any

subpages::Union{Nothing, Vector{Page}}

name::Union{Nothing, Name}

com::ObjectCommon

Contain all places, transitions & arcs. Pages are for visual presentation. There must be at least 1 Page for a valid pnml model.

PNML.Partition — Type

struct Partition{S, PE} <: PNML.SortDeclaration

id::Symbol

name::String

def::Any

element::Any

User-declared sort.

PNML.PetriNet — Type

Provides 2nd-level parsing of the intermediate representation of a single network in a pnml model. See PnmlModel.

abstract type PetriNet{T<:PnmlType}

Extended

The type parameter of a nets should map directly and simply to subtypes of PnmlType.

Additional constrants can be imposed. We want to run under the motto: "syntax is not semantics, quack".

Since a PNML.Document can contain multiple networks it is possible that a higher-level will create multiple PetriNet instances, each a different subtype.

Multiple Page can (are permitted) be merged into one page by flatten_pages! without losing any Petri Net semantics. Initial concrete PetriNets are constructed by flattening to a single Page.

PNML.Place — Type

Place node of a Petri Net Markup Language graph.

struct Place{PNTD<:PnmlType, MarkingType, SortType} <: PnmlNode

pntd::PnmlType

id::Symbol

marking::Any

initialMarking::Any

sorttype::Any

name::Union{Nothing, Name}

com::ObjectCommon

PNML.PnmlCfg — Type

Contains configuration data. #TODO add something

struct PnmlCfg

PNML.PnmlDict — Type

Alias for Dict with Symbol as key.

mutable struct Dict{Symbol, Any} <: AbstractDict{Symbol, Any}

PNML.PnmlException — Type

abstract type PnmlException <: Exception

PNML.PnmlLabel — Type

struct PnmlLabel <: PNML.Annotation

tag::Symbol

dict::Dict{Symbol, Any}

xml::EzXML.Node

Wrap a PnmlDict that may be the root of an XML-tree.

Used for labels that do not have, or we choose not to use, a dedicated parse method. Claimed labels will have a type defined to make use of the structure defined by the pntd schema. See Name, the only label defined in pnmlcore and HLLabel for similat treatment of "unclaimed" high-level labels.

PNML.PnmlModel — Type

struct PnmlModel

nets::Vector{PnmlNet}

namespace::String

reg::IDRegistry

xml::EzXML.Node

One or more Petri Nets and an ID Registry shared by all nets.

PNML.PnmlNet — Type

struct PnmlNet{PNTD<:PnmlType, D}

type::PnmlType

id::Symbol

pages::Vector{Page}

declaration::Any

name::Union{Nothing, Name}

com::ObjectCommon

xml::EzXML.Node

One Petri Net of a PNML model.

PNML.PnmlNode — Type

abstract type PnmlNode <: PnmlObject

Petri Net Graph nodes are places, transitions.

PNML.PnmlObject — Type

abstract type PnmlObject

Objects of a Petri Net Graph are pages, arcs, nodes.

PNML.ProductSort — Type

Wrap a AnyElement. Use until specialized/cooked. Should contain an ordered collection of sorts.

PNML.RefPlace — Type

Reference Place node of a Petri Net Markup Language graph. For connections between pages.

struct RefPlace{PNTD<:PnmlType} <: PNML.ReferenceNode

pntd::PnmlType

id::Symbol

ref::Symbol

name::Union{Nothing, Name}

com::ObjectCommon

PNML.RefTransition — Type

Refrence Transition node of a Petri Net Markup Language graph. For connections between pages.

struct RefTransition{PNTD<:PnmlType} <: PNML.ReferenceNode

pntd::PnmlType

id::Symbol

ref::Symbol

name::Union{Nothing, Name}

com::ObjectCommon

PNML.ReferenceNode — Type

For common behavior shared by RefPlace, RefTransition.

PNML.SimpleNet — Type

struct SimpleNet{PNTD} <: PetriNet{PNTD}

id::Symbol

net::PnmlNet{PNTD} where PNTD

TODO: Rename SimpleNet to TBD SimpleNet wraps one net. This is a PetriNet wrapping a PnmlNet.

Omits the page level of the pnml-defined hierarchy by flattening pages. A multi-page net can be flattened by removing referenceTransitions & referencePlaces, and merging pages into the first page.

Only selected fields are merged.

PNML.SimpleNet — Method

Construct from string of valid pnml XML using the first network

PNML.Sort — Type

struct Sort

dict::Dict{Symbol, Any}

Part of the many-sorted algebra attached to nodes on a Petri Net Graph. Is content of a <structure> element of a High-Level label.

PNML.Structure — Type

struct Structure{T}

tag::Symbol

dict::Any

xml::EzXML.Node

Used by/in unclaimed high-level pnml labels. Is an abstract syntax tree (ast) expressed in XML. Note the structural similarity to PnmlLabel and AnyElement

Extra

There are various defined structure ast variants:

Sort of a Place type [builtin, multi, product, user]

Term of Place HLMarking [variable, operator]

Term of Transition Condition [variable, operator]

Term of Arc Inscription [variable, operator]

Declarations of Declaration * [sort, variable, operator]

PNML.Term — Type

struct Term{T<:AbstractDict} <: PNML.AbstractTerm

tag::Symbol

dict::AbstractDict

Part of the many-sorted algebra attached to nodes on a Petri Net Graph.

ast variants:

variable

operator

PNML.Term — Method

Evaluate a term by returning the ':value' in dict. Assumes that dict is an AbstractDictionary.

PNML.TokenGraphics — Type

struct TokenGraphics <: AbstractPnmlTool

positions::Vector{Coordinate}

TokenGraphics is <toolspecific> content. Combines the <tokengraphics> and <tokenposition> elements.

PNML.ToolInfo — Type

struct ToolInfo

toolname::String

version::String

infos::Vector{AnyElement}

xml::EzXML.Node

ToolInfo holds a <toolspecific> tag.

It wraps a vector of well formed elements parsed into AnyElements. for use by anything that understands toolname, version toolspecifics.

PNML.Transition — Type

Transition node of a Petri Net Markup Language graph.

struct Transition{PNTD<:PnmlType, C} <: PnmlNode

pntd::PnmlType

id::Symbol

condition::Any

name::Union{Nothing, Name}

com::ObjectCommon

PNML.UnknownDeclaration — Type

struct UnknownDeclaration <: AbstractDeclaration

id::Symbol

name::String

nodename::String

content::Vector{AnyElement}

PNML.UserOperator — Type

struct UserOperator <: AbstractOperator

declaration::Symbol

Identity of operators's declaration.

PNML.UserSort — Type

Wrap a AnyElement. Use until specialized/cooked.

PNML.Variable — Type

struct Variable <: PNML.AbstractTerm

variableDecl::Symbol

Example input: <variable refvariable="varx"/>

PNML.VariableDeclaration — Type

struct VariableDeclaration{S} <: AbstractDeclaration

id::Symbol

name::String

sort::Any

PNML.XMLNode — Type

Alias for EzXML.Node

Constants

Docstrings for constants. This includes type aliases.

PNML.Maybe — Type

Alias for union of type T or nothing.

PNML.indent_width — Constant

Indention increment.

PNML.pnml_ns — Constant

Namespace expected for pnml XML.

PNML.tagmap — Constant

Map XML tag names to parser functions.

Functions

Docstrings for functions.

PNML._anyelement_content — Method

_anyelement_content(nodes::Vector{EzXML.Node}, pntd::PnmlType, parser; kw...) -> Dict{Symbol, Any}

Apply parser to each node in nodes. Return PnmlDict with values that are vectors when there are multiple instances of a tag in nodes and scalar otherwise.

PNML._evaluate — Function

Inscriptions, Markings, Conditions evaluate a value that may be a scalar or a Term functor.

Examples

julia> _evaluate(1)
1

julia> _evaluate(true)
true

julia> _evaluate(Term(:term, Dict(:value => 3)))
3

PNML._harvest_any! — Method

_harvest_any!(node::EzXML.Node, pntd::PnmlType, parser; kw...)

Return PnmlDict holding a well-formed XML node.

If element node has any children, each is placed in the dictonary with the child's tag name symbol as the key, repeated tags produce a vector as the value. Any XML attributes found are added as as key,value pairs.

Descend the well-formed XML using parser on child nodes.

Note the assumption that "children" and "content" are mutually exclusive. Content is always a leaf element. However XML attributes can be anywhere in the hierarchy. And neither children nor content nor attribute may be present.

PNML._match — Function

Match toolname and version.

PNML.add_label! — Method

add_label!(d::Dict{Symbol, Any}, node::EzXML.Node, pntd; kw...)

Add node to d[:labels], a vector of PnmlLabel. Return updated d[:labels].

PNML.add_toolinfo! — Method

add_toolinfo!(d::Dict{Symbol, Any}, node, pntd; kw...) -> Vector{ToolInfo}

Add ToolInfo node to d[:tools]. Return updated d[:tools].

The UML from the pnml primer (and schemas) use <toolspecific> as the tag name for instances of the type ToolInfo.

PNML.all_arcs — Function

Return vector of arcs that have a source or target of transition id.

See also src_arcs, tgt_arcs.

PNML.allchildren — Function

allchildren(tag, el) -> Vector{EzXML.Node}
allchildren(tag, el, ns) -> Vector{EzXML.Node}

Return vector of el element's immediate children with tag.

PNML.anyelement — Function

Return AnyElement wraping a tag symbol and PnmlDict holding a well-formed XML node.

See ToolInfo for one intended use-case and unclaimed_label for when a pnml label is expected but does not have a parser or the tag appears in an unexpected place.

PNML.append_page! — Method

Append selected fields of r to fields of l. NB: subpages are omitted from append_page! See flatten_pages!. Names and xml are omitted because they are scalar values, not collections.

PNML.arc — Function

Return arc with id if found, otherwise nothing.

PNML.arc_ids — Function

Return vector of arc ids.

PNML.arcs — Function

Return vector of arcs.

PNML.condition — Function

Return condition value of transition.

PNML.conditions — Function

Return a labelled vector of condition values for net s. Key is transition id.

PNML.default_condition — Function

Return default condition based on PNTD. Has meaning of true or always.

Examples

julia> m = default_condition(PnmlCore())
Condition(nothing, true, )

julia> m = default_condition(ContinuousNet())
Condition(nothing, true, )

julia> m = default_condition(HLCore())
Condition(nothing, true, )

PNML.default_inscription — Function

Return default inscription value based on PNTD. Has meaning of unity, as in one.

Examples

julia> i = default_inscription(PnmlCore())
1

julia> i = default_inscription(ContinuousNet())
1.0

julia> i = default_inscription(HLCore())
Term(:empty, Dict(:value => 1))

julia> i()
1

PNML.default_marking — Function

Return default marking value based on PNTD. Has meaning of empty, as in zero.

Examples

julia> m = default_marking(pnmltype(PnmlCore()))
PTMarking(0, )

julia> m()
0

julia> m = default_marking(typeof(pnmltype(PnmlCore())))
PTMarking(0, )

julia> m()
0

julia> m = default_marking(pnmltype(HLCore()))
HLMarking(nothing, Term(:empty, Dict(:value => 0)), )

julia> m()
0

PNML.default_sort — Function

Return default based on PNTD. Has meaning of empty, as in zero.

PNML.default_term — Function

Return default marking value based on PNTD. Has meaning of empty, as in zero.

Examples

julia> m = default_one_term(HLCore())
Term(:empty, Dict(:value => 1))

julia> m()
1

julia> m = default_zero_term(HLCore())
Term(:empty, Dict(:value => 0))

julia> m()
0

PNML.deref! — Function

Remove reference nodes from arcs. Expects flatten_pages! to have been applied so that everything is on one page (default is first page).

Axioms

All ids in a network are unique in that they only have one instance in the XML.

A chain of reference Places or Transitions always ends at a Place or Transition.

All ids are valid.

No cycles.

PNML.deref_place — Function

Return id of referenced place.

PNML.deref_transition — Function

Return id of referenced transition.

PNML.duplicate_id_action — Method

duplicate_id_action(id::Symbol; action)

Duplicated pnml node id has been detected. Default action is to issue a warning.

PNML.find_net — Function

Return PnmlNet having id or nothing`.

PNML.find_nets — Function

Return nets matching pntd type given as string or symbol. See PnmlTypes.pntd_symbol, PnmlTypes.pnmltype.

PNML.first_net — Method

first_net(model) -> PnmlNet

Return first net contained by doc.

PNML.firstchild — Function

firstchild(tag, el::EzXML.Node) -> Union{Nothing, EzXML.Node}
firstchild(tag, el::EzXML.Node, ns) -> Union{Nothing, EzXML.Node}

Return up to 1 immediate child of element el that is a tag.

PNML.firstpage — Method

Usually the only interesting page.

PNML.flatten_pages! — Function

Merge page content into the 1st page of each pnml net.

Note that refrence nodes are still present. They can be removed later with deref!.

PNML.flatten_pages! — Method

After appending r to l, recursivly flatten r into l, then empty r.

PNML.flatten_pages! — Method

flatten_pages!(net::PnmlNet) -> PnmlNet

Collect keys from all pages and move to first page.

PNML.get_label — Function

get_label(x, tag::Symbol) -> PnmlLabel

Return first label with a matching tagvalue.

PNML.get_toolinfo — Function

Return first toolinfo having a matching toolname and version.

PNML.getfirst — Method

Return first true f of v or nothing.

PNML.has_arc — Function

Return true if any arc has id.

PNML.has_graphics — Method

Return true if has a graphics element.

PNML.has_label — Function

has_label(x, tag::Symbol) -> Bool

Does any label have a matching tagvalue.

PNML.has_labels — Function

has_labels(x) -> Bool

Does x have any labels.

PNML.has_labels — Method

Return true if there is a labels element.

PNML.has_place — Function

Is there any place with id?

PNML.has_refP — Function

PNML.has_refT — Function

PNML.has_structure — Method

Return true if label has a structure field.

PNML.has_text — Method

Return true if label has text field.

PNML.has_toolinfo — Function

Does any toolinfo attached to d have a matching toolname.

PNML.has_tools — Method

Return true if has a tools element.

PNML.has_transition — Function

Is there a transition with id?

PNML.has_xml — Function

has_xml(x) -> Bool

Return true if has XML attached. Defaults to false.

PNML.in_out — Function

Return tuple of input, output labelled vectors with key of place ids and value of arc inscription's value for use as a transition function. #TODO When do these get called "pre" and "post"?

PNML.inc_indent — Method

Increment the :indent value by indent_width.

PNML.indent — Method

Return string of current indent size in io.

PNML.initialMarking — Function

Return a labelled vector with key of place id and value of marking.

PNML.ins — Method

ins(p, transition_id::Symbol) -> LabelledArrays.LArray

Return arcs of p that have transition_id as the target.

PNML.inscription — Function

Return incription value of arc.

PNML.ispid — Method

ispid(x) -> Base.Fix2{typeof(===)}

Return function to be used like: any(ispid(sym), iteraterwithpid)

PNML.isregistered — Method

isregistered(reg::IDRegistry, s::AbstractString) -> Bool

Return true if s is registered in reg.

PNML.marking — Function

Return marking value of a place p.

PNML.name — Function

name(x) -> String

Return name String. Default to empty string.

PNML.namespace — Function

Return XML namespace.

PNML.nets — Method

nets(model::PnmlModel) -> Vector{PnmlNet}

Return all nets of model.

PNML.nettype — Function

Return the PnmlType subtype representing the flavor (or pntd) of this kind of Petri Net Graph.

See also pnmltype

PNML.number_value — Method

number_value(s::AbstractString) -> Union{Nothing, Float64, Int64}

Parse string as a number. First try integer then float.

PNML.outs — Method

outs(p, transition_id::Symbol) -> LabelledArrays.LArray

Return arcs of p that have transition_id as the source.

PNML.pages — Function

Return vector of pages.

PNML.parse_and — Method

parse_and(node, pntd; kwargs...)

PNML.parse_arbitraryoperator — Method

parse_arbitraryoperator(node, pntd; kwargs...)

PNML.parse_arbitrarysort — Method

parse_arbitrarysort(node, pntd; kwargs...)

PNML.parse_arc — Method

parse_arc(node, pntd; kw...)

PNML.parse_bool — Method

parse_bool(node, pntd; kwargs...)

PNML.parse_booleanconstant — Method

parse_booleanconstant(node, pntd; kwargs...)

PNML.parse_condition — Method

parse_condition(node, pntd; kw...) -> Condition

Label of transition nodes.

Details

Condition is defined by the ISO Specification as a High-level Annotation, meaning it has <text> and <structure> elements. With all meaning in the element that the <structure> holds.

We extend Condition by allowing <structure> to have context instead of of a child element. This content is treated as either an Integer or Float64.

A Condition should evaluate to a boolean. We defer that evaluation to a higher level. See AbstractTerm.

PNML.parse_declaration — Method

parse_declaration(node, pntd; kw...) -> PNML.Declaration

Return Declaration label of 'net' and 'page' nodes.

PNML.parse_declarations — Method

parse_declarations(node, pntd; kw...) -> Vector{AbstractDeclaration}

Return an Vector{AbstractDeclaration} subtype,

PNML.parse_equality — Method

parse_equality(node, pntd; kwargs...)

PNML.parse_file — Method

parse_file(fname::AbstractString) -> PnmlModel

Build a PnmlModel from a file containing XML.

PNML.parse_graphics — Method

parse_graphics(node, pntd; kw...) -> Graphics

Arcs, Annotations and Nodes have different graphics semantics. Return a Graphics holding the union of possibilities.

PNML.parse_graphics_coordinate — Method

parse_graphics_coordinate(node, pntd; kw...) -> Coordinate

Return Coordinate. Specification seems to only use integers, we also allow real numbers.

PNML.parse_graphics_fill — Method

parse_graphics_fill(node, pntd; kw...) -> Fill

Return Fill

PNML.parse_graphics_font — Method

parse_graphics_font(node, pntd; kw...) -> Font

Return Font.

PNML.parse_graphics_line — Method

parse_graphics_line(node, pntd; kw...) -> Line

Return Line.

PNML.parse_hlinitialMarking — Method

parse_hlinitialMarking(node, pntd; kw...) -> HLMarking

PNML.parse_hlinscription — Method

parse_hlinscription(node, pntd; kw...) -> HLInscription

PNML.parse_imply — Method

parse_imply(node, pntd; kwargs...)

PNML.parse_inequality — Method

parse_inequality(node, pntd; kwargs...)

PNML.parse_initialMarking — Method

parse_initialMarking(node, pntd; kw...) -> PTMarking

PNML.parse_inscription — Method

parse_inscription(node, pntd; kw...) -> PTInscription

PNML.parse_label — Method

parse_label(node, pntd; kw...) -> Dict{Symbol, Any}

Should not often have a '<label>' tag, this will bark if one is found. Return minimal PnmlDict holding (tag,node), to defer parsing the xml.

PNML.parse_mulitsetsort — Method

parse_mulitsetsort(node, pntd; kwargs...)

PNML.parse_name — Method

parse_name(node, pntd; kw...) -> Name

Return Name holding text value and optional tool & GUI information.

PNML.parse_namedoperator — Method

parse_namedoperator(node, pntd; kw...)

PNML.parse_namedsort — Method

parse_namedsort(node, pntd; kw...)

PNML.parse_net — Function

parse_net(node)
parse_net(node, pntd; kw...)

Return a dictonary of the pnml net with keys matching their XML tag names.

PNML.parse_node — Function

Call the method matching node.name from tagmap if that mapping exists, otherwise parse as unclaimed_label wrapped in a PnmlLabel.

PNML.parse_not — Method

parse_not(node, pntd; kwargs...)

PNML.parse_or — Method

parse_or(node, pntd; kwargs...)

PNML.parse_page — Method

parse_page(node, pntd; kw...)

PNML requires at least one page.

PNML.parse_place — Method

parse_place(node, pntd; kw...)

PNML.parse_pnml — Method

parse_pnml(node; kw...)

Start parse from the pnml root node of a well formed XML document. Return a PnmlModel..

PNML.parse_pnml_common! — Method

parse_pnml_common!(d::Dict{Symbol, Any}, node, pntd; kw...) -> Union{Graphics, Vector{ToolInfo}}

Update d with any graphics, tools, and label child of node. Used by parse_pnml_node_common! & parse_pnml_label_common!.

Note that "labels" are the "everything else" option and this should be called after parsing any elements that has an expected tag. Any tag that is encountered in an unexpected location should be treated as an anonymous label for parsing.

PNML.parse_pnml_label_common! — Method

parse_pnml_label_common!(d::Dict{Symbol, Any}, node, pntd; kw...) -> Union{String, Graphics, Vector{ToolInfo}}

Update d with 'text' and 'structure' children of node, defering other tags to parse_pnml_common!.

PNML.parse_pnml_node_common! — Method

parse_pnml_node_common!(d::Dict{Symbol, Any}, node, pntd; kw...) -> Union{Graphics, Name, Vector{ToolInfo}}

Update d with name children, defering other tags to parse_pnml_common!.

PNML.parse_productsort — Method

parse_productsort(node, pntd; kwargs...)

PNML.parse_refPlace — Method

parse_refPlace(node, pntd; kw...)

PNML.parse_refTransition — Method

parse_refTransition(node, pntd; kw...)

PNML.parse_sort — Method

parse_sort(node, pntd; kw...)

Sorts are found within a <structure> element.

PNML.parse_str — Method

parse_str(str::AbstractString) -> PnmlModel

Build a PnmlModel from a string containing XML.

PNML.parse_structure — Method

parse_structure(node, pntd; kw...)

Return Structure wrapping a PnmlDict holding a <structure>. Should be inside of an label. A "claimed" label usually elids the <structure> level (does not call this method).

PNML.parse_term — Method

parse_term(node, pntd; kwargs...)

There will be no node <term>. Instead it is the interpertation of the child of some <structure> elements.

PNML.parse_text — Method

parse_text(node, pntd; kw...) -> String

Return the stripped string of nodecontent.

PNML.parse_tokengraphics — Method

parse_tokengraphics(node, pntd; kw...) -> TokenGraphics

High-level place-transition nets (HL-PTNet) have a toolspecific structure defined for token graphics. Contains <tokenposition> tags.

PNML.parse_tokenposition — Method

parse_tokenposition(node, pntd; kw...) -> Coordinate

Cartesian coordinate relative to containing element.

PNML.parse_toolspecific — Method

parse_toolspecific(node, pntd; kw...) -> ToolInfo

Return ToolInfo with tool & version attributes and content.

The content can be one or more well-formed xml elements. Each are wrapped in a PnmlLabel.

PNML.parse_transition — Method

parse_transition(node, pntd; kw...)

PNML.parse_tuple — Method

parse_tuple(node, pntd; kwargs...)

PNML.parse_type — Method

parse_type(node, pntd; kwargs...)

Defines the "sort" of tokens held by the place and semantics of the marking. NB: The "type" of a place is different from the "type" of a net or "pntd".

PNML.parse_unknowndecl — Method

parse_unknowndecl(node, pntd; kw...) -> PNML.UnknownDeclaration

PNML.parse_unparsed — Method

parse_unparsed(node, pntd; kwargs...)

PNML.parse_useroperator — Method

parse_useroperator(node, pntd; kwargs...) -> UserOperator

PNML.parse_usersort — Method

parse_usersort(node, pntd; kwargs...)

PNML.parse_variable — Method

parse_variable(node, pntd; kwargs...) -> Variable

PNML.parse_variabledecl — Method

parse_variabledecl(node, pntd; kw...)

PNML.pid — Function

pid(x) -> Symbol

Return pnml id symbol.

PNML.pid — Method

PnmlObjects are exected to have unique pnml ids.

PNML.place — Function

Return the place with id.

PNML.place_ids — Function

Return vector of place IDs.

PNML.places — Function

Return vector of places.

PNML.pnml_common_defaults — Method

pnml_common_defaults(node) -> Dict{Symbol, Any}

Return Dict of tags common to both pnml nodes and pnml labels. See also: pnml_label_defaults, pnml_node_defaults.

PNML.pnml_label_defaults — Method

pnml_label_defaults(node, xs...) -> Any

Merge xs into dictonary with default pnml HLannotation label tags. Used on pnml tags below a PnmlNode tag. Label level tags include: name, inscription, initialMarking. Notable differences from pnml_node_defaults: text, structure, no name tag. See also: pnml_common_defaults.

PNML.pnml_namespace — Method

pnml_namespace(node::EzXML.Node; missing_ns_fatal, default_ns) -> String

Return namespace of node. Whennodedoes not have a namespace return default value [pnml_ns`](@ref)."

PNML.pnml_node_defaults — Method

pnml_node_defaults(node, xs...) -> Any

Merge xs into dictonary with default pnml node tags. Used on: net, page ,place, transition, arc. Usually default value will be nothing or empty vector. See also: pnml_label_defaults, pnml_common_defaults.

PNML.rate — Function

Return rate value of transition. Mising rate labels are defaulted to 0.0.

PNML.rates — Function

Return a transition-id labelled vector of rate values for transitions of net s.

PNML.refplace — Function

Return reference place matching id.

PNML.refplace_ids — Function

PNML.refplaces — Function

Return vector of reference places.

PNML.reftransition — Function

Return reference transition matching id.

PNML.reftransition_ids — Function

PNML.reftransitions — Function

Return vector of reference transitions.

PNML.register_id! — Method

register_id!(reg::IDRegistry, s::AbstractString) -> Symbol

Register id symbol and return the symbol.

PNML.reset_registry! — Method

reset_registry!(reg::IDRegistry) -> Set{Symbol}

Empty the set of id symbols. Use case is unit tests. In normal use it should never be needed.

PNML.source — Method

source(arc) -> Symbol

Return symbol of source of arc.

PNML.src_arcs — Function

Return vector of arcs that have a source of transition id.

See also all_arcs, tgt_arcs.

PNML.structure — Method

Return structure field.

PNML.tag — Function

tag(x) -> Symbol

Return tag symbol.

PNML.target — Method

target(arc) -> Symbol

Return symbol of target of arc.

PNML.text — Method

Return text field

PNML.tgt_arcs — Function

Return vector of arcs that have a target of transition id.

See also all_arcs, src_arcs.

PNML.transition — Function

PNML.transition_function — Function

Transition function of a Petri Net. Each transition has an input vector and an output vector. Each labelled vector is indexed by the place on the other end of the arc. Values are inscriptions.

PNML.transition_ids — Function

PNML.transitions — Function

Return vector of transitions.

PNML.unclaimed_label — Function

Return tag => PnmlDict holding a pnml label and its children.

The main use-case is to wrap in a PnmlLabel, Structure, Term or other specialized label. These wrappers add type to the nested dictonary holding the contents of the label.

Differs from AnyElement in that any id attribute of node will be registered in IDRegistry as a unique identifier.

PNML.xmlnode — Function

xmlnode(x) -> XMLNode

Return attached xml node.

Macros

Docstrings for macros.

PNML.@xml_str — Macro

Parse xml string into EzXML node.