PNML Reference

PnmlTypeDefs Reference

PnmlIDRegistrys Reference

PNML Reference

PNML.PNML — Module

PNML

Petri Net Markup Language, is an XML-based format. PNML.jl reads a pnml model and emits an intermediate representation (IR).

The intermediate representation (IR) represents the XML tree via julia data structures: dictionaries, NamedTuples, LabelledArrays, strings, numbers, objects, vectors. The exact mixture changes as the project continues.

The tags of the XML are used as keys and names as much as possible.

What is accepted as values is ~~often~~ usually a superset of what a given pntd schema specifies. This can be thought of as duck-typing. Conforming to the pntd is not the role of the IR.

The pnml specification has layers. The core layer is useful and extendable. The standard defines extensions of the core for place-transition petri nets (integers) and high-level petri net graphs (many-sorted algebra). This package family adds non-standard continuous net (float64) support. Note that there is no RelaxNG schema file for these extensions

On top of the IR is (will be) implemented Petri Net adaptions and interpertations. This is the level that pntd conformance can be imposed. Adaption to julia packages for graphs, agents, and composing into the greater hive-mind.

Features that have not been started:

Write pnml file

Update pnml model

Create pnml model

Graphs.jl

Symbolics support for HLPNG (many-sorted algebra)

Features that are not complete:

HLPNG - many-sorted algebras are complex. Will build other infrastructure...

pntd specialize

toolspecific usage example

Features that work (perhaps in need of attention as changes are made):

continuous petri net (examples/lotka-volterra.jl)

pnml core: can read & print all Model Checking Contest (MCC) models

Imports

AbstractTrees

AutoHashEquals

Base

Core

DocStringExtensions

EzXML

FunctionWrappers

LabelledArrays

NamedTupleTools

PNML.PnmlIDRegistrys

PNML.PnmlTypeDefs

PrecompileTools

Preferences

PrettyPrinting

Reexport

Exports

AbstractContinuousNet

AbstractHLCore

AbstractPnmlCore

ContinuousNet

HLCoreNet

HLPNG

MalformedException

MissingIDException

OpenNet

PTNet

PT_HLPNG

PnmlCoreNet

PnmlException

PnmlIDRegistry

PnmlIDRegistrys

PnmlType

PnmlTypeDefs

StochasticNet

SymmetricNet

TimedNet

iscontinuous

isdiscrete

ishighlevel

isregistered

parse_file

parse_node

parse_pnml

parse_str

pnmltype

pntd_symbol

register_id!

registry

@xml_str

xmlroot

PNML.CONFIG — Constant

PnmlConfig

Configuration options

Options

indent_width::Int: Indention of nested lines. Defaults to 4.

warn_on_namespace::Bool: There are pnml files that break the rules &

do not have an xml namespace. Initial state of toggle defaults to true.

text_element_optional::Bool: There are pnml files that break the rules & do not have <text> elements.

Initial state of warning toggle defaults to true.

warn_on_fixup::Bool: When an missing value is replaced by a default value,

issue a warning. Initial state of "warnonfixup" toggle defaults to false.

warn_on_unclaimed::Bool: Issue warning when PNML label does not have a parser defined.

While allowed, there will be code required to do anything useful with the label. Initial state of "warn" toggle defaults to false.

verbose::Bool: Print information as runs. Initial state of "verbose" toggle defaults to false.

lock_registry::Bool: Default is true to use a lock, deefault Re.

PNML.Maybe — Type

Alias for union of type T or Nothing.

PNML.pnml_ns — Constant

Namespace expected for pnml XML.

PNML.tagmap — Constant

Map XML <tag> names to parser functions named "parse_tag". Allows associative lookup of parser using ideas from Tim Holy's MathML package.

PNML.AbstractDeclaration — Type

abstract type AbstractDeclaration <: PNML.AbstractLabel

Declarations are the core of high-level Petri Net. They define objects/names that are used for conditions, inscriptions, markings. They are attached to PNML nets and pages.

PNML.AbstractLabel — Type

abstract type AbstractLabel

Labels are attached to the Petri Net Graph objects. See AbstractPnmlObject.

PNML.AbstractOperator — Type

abstract type AbstractOperator <: PNML.AbstractTerm

Part of the high-level pnml many-sorted algebra.

PNML.AbstractPetriNet — Type

abstract type AbstractPetriNet{PNTD<:PnmlType}

Top-level of a single network in a pnml model that is some flavor of Petri Net. Note that pnml can represent nets that are not Petri Nets.

Here is where specialization and restriction are applied to achive Proper Petri Behavior.

See PnmlModel, PnmlType.

Extended

Additional constrants can be imposed. We want to run under the motto: "syntax is not semantics, quack".

Since a PNML.Document model can contain multiple networks it is possible that a higher-level will create multiple AbstractPetriNet instances, each a different type.

Multiple Pages can (are permitted to) be merged into one page by flatten_pages! without losing any Petri Net semantics.

PNML.AbstractPnmlNode — Type

abstract type AbstractPnmlNode{PNTD} <: PNML.AbstractPnmlObject{PNTD}

Petri Net Graph nodes are Place, Transition. They are the source or target of an Arc

PNML.AbstractPnmlObject — Type

abstract type AbstractPnmlObject{PNTD<:PnmlType}

Objects of a Petri Net Graph are pages, arcs, nodes.

PNML.AbstractPnmlTool — Type

abstract type AbstractPnmlTool

Tool specific objects can be attached to AbstractPnmlObjects and AbstractLabels subtypes.

PNML.AbstractSort — Type

abstract type AbstractSort

Part of the high-level pnml many-sorted algebra. See SortType.

NamedSort is an AbstractTerm that declares a definition using an AbstractSort. The pnml specification sometimes uses overlapping language.

From the 'primer': built-in sorts of Symmetric Nets are the following: booleans, integerrange, finite enumerations, cyclic enumerations, permutations and dots.

The eltype is expected to be a concrete subtype of Number such as Int, Bool or Float64.

PNML.AbstractTerm — Type

abstract type AbstractTerm

Terms are part of the multi-sorted algebra that is part of a High-Level Petri Net.

An abstract type in the pnml XML specification, concrete Terms are found within the <structure> element of a HLAnnotation label.

Notably, a Term is not a PnmlLabel.

See also Declaration, SortType, AbstractDeclaration.

PNML.AnyElement — Type

struct AnyElement

tag::Symbol

elements::Vector{PNML.AnyXmlNode}

xml::EzXML.Node

Hold well-formed XML in a Vector{AnyXmlNode}. See also ToolInfo and PnmlLabel.

PNML.AnyXmlNode — Type

Node in a tree formed from XML. tags are XML tags or attribute names. Leaf val are strings. NB: Assumes XML "content" nodes do not have child XML nodes.

PNML.ArbitraryOperator — Type

struct ArbitraryOperator{I<:PNML.AbstractSort} <: PNML.AbstractOperator

declaration::Symbol: Identity of operators's declaration.

input::PNML.AbstractSort

output::Vector{PNML.AbstractSort}

PNML.ArbitrarySort — Type

struct ArbitrarySort <: PNML.SortDeclaration

id::Symbol

name::String

Arbitrary sorts that can be used for constructing terms are reserved for/supported by HLPNG in the pnml specification.

PNML.Arc — Type

Edge of a Petri Net Markup Language graph that connects place and transition.

mutable struct Arc{PNTD, I} <: PNML.AbstractPnmlObject{PNTD}

pntd::Any

id::Symbol

source::Symbol

target::Symbol

inscription::Any

name::Union{Nothing, PNML.Name}

com::PNML.ObjectCommon

PNML.BoolSort — Type

Built-in sort whose eltype is Bool

PNML.Condition — Type

struct Condition <: PNML.Annotation

text::Union{Nothing, String}

value::Union{Bool, PNML.Term}

com::PNML.ObjectCommon

Label of a Transition that determines when the transition fires.

Examples

julia> c = Condition(false)
Condition(nothing, false, )

julia> c()
false

julia> c = Condition("xx", false)
Condition("xx", false, )

julia> c()
false

PNML.Coordinate — Type

struct Coordinate{T<:Union{Float64, Int64}}

x::Union{Float64, Int64}

y::Union{Float64, Int64}

Cartesian Coordinate are actually positive decimals. Ranges from 0 to 999.9.

PNML.Declaration — Type

struct Declaration <: PNML.Annotation

declarations::Vector{Any}

com::PNML.ObjectCommon

xml::Union{Nothing, EzXML.Node}

Label of a <net> or <page> that holds zero or more declarations. The declarations are used to define parts of the many-sorted algebra used by High-Level Petri Nets. Other PNTDs may introduce non-standard uses for declarations.

Notes: declarations is implemented as a collection of Any.

We can use infrastructure implemented for HL nets to provide nonstandard extensions.

PNML.DotSort — Type

Built-in sort whose eltype is Int

PNML.Fill — Type

Fill attributes as strings.

struct Fill

color::String

image::String

gradient_color::String

gradient_rotation::String

PNML.Font — Type

Font attributes as strings.

struct Font

family::String

style::String

weight::String

size::String

align::String

rotation::String

decoration::String

PNML.Graphics — Type

PNML Graphics can be attached to many parts of PNML models.

struct Graphics{T<:Union{Float64, Int64}}

dimension::PNML.Coordinate

fill::PNML.Fill

font::PNML.Font

line::PNML.Line

offset::PNML.Coordinate

positions::Array{PNML.Coordinate{T}, 1} where T<:Union{Float64, Int64}

PNML.HLAnnotation — Type

abstract type HLAnnotation <: PNML.AbstractLabel

Annotation label that uses <text> and <structure>.

PNML.HLInscription — Type

struct HLInscription{T<:PNML.Term} <: PNML.HLAnnotation

text::Union{Nothing, String}

term::PNML.Term

com::PNML.ObjectCommon

Labels an Arc. The <structure> element is a term in a many-sorted algebra. The term field

Examples

julia> i2 = HLInscription(Term(:value, 3))
HLInscription(nothing, Term(:value, 3), )

julia> i2()
3

julia> i3 = HLInscription("text", Term(:empty, 1))
HLInscription("text", Term(:empty, 1), )

julia> i3()
1

julia> i4 = HLInscription("text", Term(:value, 3))
HLInscription("text", Term(:value, 3), )

julia> i4()
3

PNML.HLInscription — Method

Evaluate a HLInscription. Returns a value of the same sort as TBD.

PNML.HLLabel — Type

struct HLLabel{PNTD} <: PNML.HLAnnotation

text::Union{Nothing, String}

structure::Union{Nothing, PNML.AnyElement}

com::PNML.ObjectCommon

xml::EzXML.Node

High-level pnml labels are expected to have <text> and <structure> elements. This concrete type is for "unclaimed" labels in a high-level petri net.

Some "claimed" HLAnnotation labels are Condition, Declaration, HLMarking, HLInscription.

PNML.HLMarking — Type

Label of a Place in a High-level Petri Net Graph. See AbstractHLCore, Term.

Multisets of a sort.

struct HLMarking{T<:PNML.AbstractTerm} <: PNML.HLAnnotation

text::Union{Nothing, String}

term::PNML.AbstractTerm

com::PNML.ObjectCommon

Examples

julia> m = HLMarking("the text", Term(:value, 3))
HLMarking("the text", Term(:value, 3), )

julia> m()
3

PNML.HLMarking — Method

Evaluate a HLMarking instance by returning its term.

PNML.HLPetriNet — Type

Wrap a single pnml net. Presumes that the net does not need to be flattened as all content is in first page.

struct HLPetriNet{PNTD} <: PNML.AbstractPetriNet{PNTD}

net::PNML.PnmlNet{PNTD} where PNTD

Details

PNML.HLPetriNet — Method

Construct from string of valid pnml XML, using the first network in model.

PNML.HarvestAny — Type

Wrap a function and two of its arguments.

PNML.Inscription — Type

struct Inscription{T<:Union{Float64, Int64}} <: PNML.Annotation

value::Union{Float64, Int64}

com::PNML.ObjectCommon

Labels an Arc.

Examples

julia> i = Inscription(3)
Inscription(3, )

julia> i()
3

PNML.Inscription — Method

Evaluate an Inscription's value.

PNML.IntegerSort — Type

"Built-in sort whose eltype is Int"

PNML.Line — Type

Line attributes as strings.

struct Line

color::String

shape::String

style::String

width::String

PNML.MalformedException — Type

struct MalformedException <: PnmlException

msg::String

PNML.Marking — Type

Number-valued label of Place. See PTNet, ContinuousNet.

Is a functor that returns the value.

struct Marking{N<:Union{Float64, Int64}} <: PNML.Annotation

value::Union{Float64, Int64}

com::PNML.ObjectCommon

Examples

julia> m = Marking(1)
Marking(1, )

julia> m()
1

julia> m = Marking(12.34)
Marking(12.34, )

julia> m()
12.34

PNML.Marking — Method

Evaluate a Marking instance by returning its value.

PNML.MissingIDException — Type

struct MissingIDException <: PnmlException

msg::String

Use exception to allow dispatch and additional data presentation to user.

PNML.MultisetSort — Type

struct MultisetSort <: PNML.AbstractSort

Wrap a AbstractSort. Use until specialized/cooked.

PNML.Name — Type

struct Name <: PNML.Annotation

text::String

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

Name is for display, possibly in a tool specific way.

PNML.NamedOperator — Type

struct NamedOperator{V, T} <: PNML.OperatorDeclaration

id::Symbol

name::String

parameter::Vector

def::Any

PNML.NamedSort — Type

struct NamedSort{S<:Union{PNML.AbstractSort, PNML.AnyElement}} <: PNML.SortDeclaration

id::Symbol

name::String

def::Union{PNML.AbstractSort, PNML.AnyElement}

PNML.NaturalSort — Type

"Built-in sort whose eltype is Int"

PNML.ObjectCommon — Type

struct ObjectCommon

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

labels::Vector{PNML.PnmlLabel}

Common infrastructure shared by PNML objects and labels. Some optional incidental bits are shared by most PNML objects are also collected here.

PNML.OperatorDeclaration — Type

abstract type OperatorDeclaration <: PNML.AbstractDeclaration

PNML.Page — Type

struct Page{PNTD<:PnmlType, P, T, A, RP, RT} <: PNML.AbstractPnmlObject{PNTD<:PnmlType}

pntd::PnmlType

id::Symbol

declaration::PNML.Declaration

name::Union{Nothing, PNML.Name}

com::PNML.ObjectCommon

pagedict::OrderedCollections.OrderedDict{Symbol, PNML.Page{PNTD, P, T, A, RP, RT}} where {PNTD<:PnmlType, P, T, A, RP, RT}

netdata::PNML.PnmlNetData

netsets::PNML.PnmlNetKeys

Contain all places, transitions & arcs. Pages are for visual presentation. There must be at least 1 Page for a valid pnml model.

PNTD binds the other type parameters together to express a specific PNG. See PnmlNet

PNML.Partition — Type

struct Partition{S, PE} <: PNML.SortDeclaration

id::Symbol

name::String

def::Any

element::Any

User-declared sort.

PNML.Place — Type

struct Place{PNTD, M, S<:PNML.SortType} <: PNML.AbstractPnmlNode{PNTD}

pntd::Any

id::Symbol

marking::Any

initialMarking::Any

sorttype::PNML.SortType

name::Union{Nothing, PNML.Name}

com::PNML.ObjectCommon

Place node of a Petri Net Markup Language graph.

PNML.PnmlConfig — Type

TODO

PNML.PnmlException — Type

abstract type PnmlException <: Exception

PNML.PnmlLabel — Type

struct PnmlLabel <: PNML.Annotation

tag::Symbol

elements::Vector{PNML.AnyXmlNode}

xml::EzXML.Node

Wrap a AnyXmlNode[] holding a pnml label. Use the XML tag as identifier.

Used for "unclaimed" labels that do not have, or we choose not to use, a dedicated parse method. Claimed labels will have a type/parser defined to make use of the structure defined by the pntd schema.

See also AnyElement. The difference is that AnyElement allows any well-formed XML, while PnmlLabel is restricted to PNML Labels (with extensions in PNML.jl).

PNML.PnmlModel — Type

struct PnmlModel

nets::Tuple{Vararg{PNML.PnmlNet}}

namespace::String

reg::PnmlIDRegistry

One or more Petri Nets and an ID Registry shared by all nets.

PNML.PnmlNet — Type

struct PnmlNet{PNTD<:PnmlType, P, T, A, RP, RT}

type::PnmlType

id::Symbol

pagedict::OrderedCollections.OrderedDict{Symbol, PNML.Page{PNTD, P, T, A, RP, RT}} where {PNTD<:PnmlType, P, T, A, RP, RT}

netdata::PNML.PnmlNetData

page_set::OrderedCollections.OrderedSet{Symbol}

declaration::PNML.Declaration

name::Union{Nothing, PNML.Name}

com::PNML.ObjectCommon

xml::EzXML.Node

One Petri Net of a PNML model.

PNML.PnmlNetData — Type

struct PnmlNetData{PNTD<:PnmlType, P, T, A, RP, RT}

pntd::PnmlType

place_dict::OrderedCollections.OrderedDict{Symbol}

transition_dict::OrderedCollections.OrderedDict{Symbol}

arc_dict::OrderedCollections.OrderedDict{Symbol}

refplace_dict::OrderedCollections.OrderedDict{Symbol}

reftransition_dict::OrderedCollections.OrderedDict{Symbol}

Collect each of the PnmlNodess & Arcs of a Petri Net Graph into one collection. Accessed via pnml ID key or iterate over values of an OrderedDict.

In the 'pnml' specification there is a Page structure that can be removed by flatten_pages!, removing some display-related information, leaving a functional Petri Net Graph as described in this structure. It is intended to be a per-PnmlNet database that is mutated as each page is parsed.

See PnmlNetKeys for page-level pnml ID of "owners" net data.

PNML.PnmlNetKeys — Type

struct PnmlNetKeys

page_set::OrderedCollections.OrderedSet{Symbol}

place_set::OrderedCollections.OrderedSet{Symbol}

transition_set::OrderedCollections.OrderedSet{Symbol}

arc_set::OrderedCollections.OrderedSet{Symbol}

reftransition_set::OrderedCollections.OrderedSet{Symbol}

refplace_set::OrderedCollections.OrderedSet{Symbol}

Per-page structure of OrderedSets of pnml IDs for each "owned" Page and AbstractPnmlObject.

PNML.PositiveSort — Type

"Built-in sort whose eltype is Int"

PNML.ProductSort — Type

struct ProductSort <: PNML.AbstractSort

Wrap a AnyElement. Use until specialized/cooked. Should contain an ordered collection of sorts.

PNML.RealSort — Type

"Built-in sort whose eltype is Float64" Real numbers are not part of the PNML Specification. We stick them into the type hierarchy for convenience rather than mathematical correctness.

PNML.RefPlace — Type

Reference Place node of a Petri Net Markup Language graph. For connections between pages.

struct RefPlace{PNTD} <: PNML.ReferenceNode{PNTD}

pntd::Any

id::Symbol

ref::Symbol

name::Union{Nothing, PNML.Name}

com::PNML.ObjectCommon

PNML.RefTransition — Type

Refrence Transition node of a Petri Net Markup Language graph. For connections between pages.

struct RefTransition{PNTD} <: PNML.ReferenceNode{PNTD}

pntd::Any

id::Symbol

ref::Symbol

name::Union{Nothing, PNML.Name}

com::PNML.ObjectCommon

PNML.ReferenceNode — Type

abstract type ReferenceNode{PNTD} <: PNML.AbstractPnmlNode{PNTD}

For common behavior shared by RefPlace, RefTransition used to connect Page together.

PNML.SimpleNet — Type

struct SimpleNet{PNTD} <: PNML.AbstractPetriNet{PNTD}

id::Symbol

net::PNML.PnmlNet{PNTD} where PNTD

TODO: Rename SimpleNet to TBD

SimpleNet is a concrete AbstractPetriNet wrapping a PnmlNet.

Uses a flattened net to avoid the page level of the pnml hierarchy.

A multi-page petri net can always be flattened by removing referenceTransitions & referencePlaces, and merging pages into the first page.

PNML.SortDeclaration — Type

abstract type SortDeclaration <: PNML.AbstractDeclaration

See NamedSort and [ArbitrarySort] as concrete subtypes.

PNML.SortType — Type

struct SortType{T<:PNML.AbstractSort}

text::Union{Nothing, String}

sort::PNML.AbstractSort

com::PNML.ObjectCommon

A places's <type> label's <structure> element holds a concrete subtype of AbstractSort. Defines the sort of a place, hence use of sorttype.

For high-level nets there will be a rich language of sorts using UserSort and NamedSort.

Notes:

NamedSort is a SortDeclaration. HLPNG adds ArbitrarySort.

UserSort holds the id symbol of a NamedSort.

Here 'type' means a 'term' from the many-sorted algebra.

We use sorts even for non-high-level nets for type-stability.

Expect eltype(::AbstractSort) to return a concrete subtype of Number.

PNML.Structure — Type

struct Structure

tag::Symbol

el::Vector{PNML.AnyXmlNode}

xml::EzXML.Node

High-level Annotation Labels place meaning in <structure>. Is is expected to contain an abstract syntax tree (ast) for the many-sorted algebra expressed in XML.

Note the structural similarity to PnmlLabel and AnyElement

Extra

There are various defined structure ast variants in pnml:

Sort Type of a Place [builtin, multi, product, user]

Place HLMarking [variable, operator]

Transition Condition [variable, operator]

Arc Inscription [variable, operator]

Declarations [sort, variable, operator]

These should all have dedicated parsers and objects as claimed lables. Here we provide a fallback for unclaimed labels.

PNML.Term — Type

struct Term <: PNML.AbstractTerm

tag::Symbol

elements::Union{Bool, Float64, Int64, Vector{PNML.AnyXmlNode}}

Note that Term is an abstract element in the pnml specification with no XML tag. Here we use it as a concrete wrapper around high-level many-sorted algebra terms AND EXTEND to also wrapping "single-sorted" values.

By adding Bool, Int, Float64 it is possible for PnmlCoreNet and ContinuousNet to use Terms, and for implenting default_bool_term, default_one_term, default_zero_term.

See also [iscont]

As part of the many-sorted algebra attached to nodes of a High-level Petri Net Graph, Terms are contained within the <structure> element of an annotation label, See [HLAnnotation`](@ref) concrete subtypes.

#TODO is it safe to assume that Bool, Int, Float64 are in the carrier set/basis set?

#! HL term is currently implemented as a wrapper of Vector{AnyXmlNode}. A tree of symbols with leafs that are strings.

Functor

(t::Term)([default_one_term(HLCoreNet())])

Term as functor requires a default value for missing values.

As a preliminary implementation evaluate a HL term by returning :value in elements or evaluating default_one_term(HLCoreNet()).

See parse_marking_term, parse_condition_term, parse_inscription_term, parse_type, parse_sorttype_term, AnyElement.

Warning: Much of the high-level is WORK-IN-PROGRES. The type parameter is a sort. We enumerate some of the built-in sorts allowed. Is expected that the term will evaluate to that type. Is that called a 'ground term'? 'basis set'? When the elements' value is a Vector{AnyXmlNode} external information is used to select the output type.

PNML.TokenGraphics — Type

struct TokenGraphics{T} <: PNML.AbstractPnmlTool

positions::Array{PNML.Coordinate{T}, 1} where T

TokenGraphics is <toolspecific> content. Combines the <tokengraphics> and <tokenposition> elements.

PNML.ToolInfo — Type

struct ToolInfo{T}

toolname::String

version::String

infos::Any

xml::EzXML.Node

ToolInfo holds a <toolspecific> tag.

It wraps a iteratable collection (currently vector) of well formed elements parsed into AnyElements for use by anything that understands toolname, version tool specifics.

PNML.Transition — Type

Transition node of a Petri Net Markup Language graph.

struct Transition{PNTD, C} <: PNML.AbstractPnmlNode{PNTD}

pntd::Any

id::Symbol

condition::Any

name::Union{Nothing, PNML.Name}

com::PNML.ObjectCommon

PNML.UnknownDeclaration — Type

struct UnknownDeclaration <: PNML.AbstractDeclaration

id::Symbol

name::String

nodename::String

content::Vector{Any}

PNML.UserOperator — Type

struct UserOperator <: PNML.AbstractOperator

declaration::Symbol: Identity of operators's declaration.

PNML.UserSort — Type

struct UserSort <: PNML.AbstractSort

Holds a reference id to a concrete subtype of SortDeclaration. NamedSort is used to construct a sort out of builtin types. Used in a Places sort type property.

PNML.Variable — Type

struct Variable <: PNML.AbstractTerm

variableDecl::Symbol

Example input: <variable refvariable="varx"/>

PNML.VariableDeclaration — Type

struct VariableDeclaration{S} <: PNML.AbstractDeclaration

id::Symbol

name::String

sort::Any

PNML.XMLNode — Type

Alias for EzXML.Node

PNML._anyelement_content! — Method

_anyelement_content!(
    vec,
    node::EzXML.Node,
    harvest!::PNML.HarvestAny
)

Apply harvest! to each child node of node, appending to vec.

PNML._evaluate — Function

_evaluate(x::Number) -> identity(x)
_evaluate(x::Base.Callable) -> (x)()

Return the value of "x", defaults to identity.

Examples

Since High-level PNML schemas are based on Natural numbers and booleans, it seems reasonable to assume Number, which includes Bool, for the non-callable type. A functor is expected as the callable type allowing expressions in the many-sorted algebra to be evaluated to aNumber`.

PNML._harvest_any! — Method

_harvest_any!(
    node::EzXML.Node,
    harvest!::PNML.HarvestAny
) -> Vector{PNML.AnyXmlNode}

Return AnyXmlNode vector holding a well-formed XML node.

If element node has any attributes &/or children, use attribute name or child's name as tag. (Empty is well-formed.)

Descend the well-formed XML using parser function harvest! on child nodes.

Note the assumption that "children" and "content" are mutually exclusive. Content is always a leaf element. However XML attributes can be anywhere in the hierarchy. And neither children nor content nor attribute may be present.

Leaf AnyXmlNode's contain an String or SubString.

PNML._match — Function

_match(ti::ToolInfo, name::AbstractString)
_match(ti::ToolInfo, name::String, version::String)
_match(ti::ToolInfo, namerex::Regex, versionrex::Regex)

Match toolname and version. Default is any version.

PNML._parse_decl_structure — Method

Assumes high-level semantics until someone specializes. See decl_structure.

PNML._parse_page! — Method

Call parse_page!, add page to dictionary and id set

PNML.add_toolinfo! — Method

add_toolinfo!(v::Vector{PNML.ToolInfo}, node, pntd, reg)

Add ToolInfo to vector, return nothing.

The UML from the pnml primer (and schemas) use <toolspecific> as the tag name for instances of the type ToolInfo.

PNML.all_arcs — Function

Return arcs that have a source or target of transition id.

See also src_arcs, tgt_arcs.

PNML.allchildren — Function

allchildren(
    tag::AbstractString,
    el::EzXML.Node
) -> Vector{EzXML.Node}
allchildren(
    tag::AbstractString,
    el::EzXML.Node,
    ns::AbstractString
) -> Vector{EzXML.Node}

Return vector of el's immediate children with tag.

PNML.anyelement — Function

Return AnyElement wraping a tag symbol and tuple holding a well-formed XML node. See ToolInfo for one intended use-case.

PNML.append_page! — Method

Append selected fields of r to fields of l. Some, like Names and xml, are omitted because they are scalar values, not collections.

pagedict & netdata (holding the arc and pnml nodes) are per-net data that is not modified here. netsets hold pnml IDs "owned"

PNML.arc — Function

Return arc with id if found, otherwise nothing.

PNML.arc_idset — Function

Return set arc ids.

PNML.arc_type — Function

Type of Arc.

PNML.arcs — Function

arcs(p::Page) -> iterator
arcs(n::PnmlNet) -> iterator
arcs(p::AbstractPetriNet) -> iterator

Return iterator over arc ids.

PNML.condition — Function

Return condition value of transition.

PNML.condition_type — Function

Type of Condition #!{conditionvaluetype(T)}`

PNML.condition_value_type — Function

Return value type.

PNML.conditions — Function

Return a labelled vector of condition values.

PNML.coordinate_type — Function

Return Coordinate{coordinate_value_type(T)}

PNML.coordinate_value_type — Function

Return type of value in a Coordinate.

PNML.currentMarkings — Method

currentMarkings(net) -> LVector{marking_value_type(n)}

LVector labelled with place id and holding marking's value.

PNML.decl_structure — Method

Return vector of AbstractDeclaration subtypes.

PNML.default_bool_term — Function

True as boolean or term with a value of true.

PNML.default_condition — Function

Return default condition based on PNTD. Has meaning of true or always. ```

PNML.default_inscription — Function

Return default inscription value based on PNTD. Has meaning of unity, as in one.

PNML.default_marking — Function

Return default marking value based on PNTD. Has meaning of empty, as in zero.

PNML.default_one_term — Function

One as integer, float, or empty term with a value of one.

PNML.default_sort — Function

Return instance of default sort based on PNTD.

PNML.default_sort_type — Function

PNML.default_sorttype — Function

Return instance of default place sort type based on PNTD.

PNML.default_zero_term — Function

Zero as integer, float, or empty term with a value of zero.

PNML.deref! — Function

deref!(net::PNML.PnmlNet) -> PNML.PnmlNet
deref!(net::PNML.PnmlNet, trim::Bool) -> PNML.PnmlNet

Remove reference nodes from arcs.

Operates on the PnmlNetData at the net level. Expects that the PnmlNetKeys of the firstpage will have to be cleaned as part of flatten_pages!,

Axioms

All ids in a network are unique in that they only have one instance in the XML.

A chain of reference Places or Transitions always ends at a Place or Transition.

All ids are valid.

No cycles.

PNML.deref_place — Function

deref_place(net, id[, trim::Bool] ) -> Symbol

Return id of referenced place. If trim is true (default) the reference is removed.

PNML.deref_transition — Function

deref_transition(net, id[, trim::Bool] ) -> Symbol

Return id of referenced transition. If trim is true (default) the reference is removed.

PNML.equals — Method

equals(a::PNML.AbstractSort, b::PNML.AbstractSort) -> Any

Are the sorts eltype the same?

PNML.find_net — Function

Return PnmlNet having id or nothing`.

PNML.find_nets — Function

Return nets matching pntd type given as string, symbol or pnmltype singleton.

PNML.first_net — Method

first_net(model) -> PNML.PnmlNet

Return first net contained by doc.

PNML.firstchild — Function

firstchild(
    tag::AbstractString,
    node::EzXML.Node
) -> Union{Nothing, EzXML.Node}
firstchild(
    tag::AbstractString,
    node::EzXML.Node,
    ns::AbstractString
) -> Union{Nothing, EzXML.Node}

Return up to 1 immediate child of el that is a tag.

PNML.firstpage — Method

Usually the only interesting page.

PNML.flatten_pages! — Function

flattenpages!(net::PnmlNet[; options]) flattenpages!(model::PnmlModel[; options])

Merge page content into the 1st page of the net or all nets of a model.

Options trim::Bool Remove refrence nodes (default false). See deref!. verbose::Bool Print breadcrumbs (default false).

PNML.get_label — Function

get_label(x, tag::Symbol) -> PnmlLabel

Return first label of x with a matching tagvalue.

PNML.get_toolinfo — Function

Return first toolinfo having a matching toolname and version.

PNML.getfirst — Method

Return first true f of v or nothing.

PNML.has_arc — Function

Return true if any arc has id.

PNML.has_label — Function

has_label(x, tag::Symbol) -> Bool

Does x have any label have a matching tagvalue.

PNML.has_labels — Function

has_labels(x) -> Bool

Does x have any labels.

PNML.has_name — Function

has_name(x) -> Bool

Return true if there is a name.

PNML.has_place — Function

Return true if there is any place with id?

PNML.has_refP — Function

PNML.has_refT — Function

PNML.has_structure — Method

Return true if label has a structure field.

PNML.has_text — Method

Return true if label has text field.

PNML.has_toolinfo — Function

Does any toolinfo attached to d have a matching toolname.

PNML.has_transition — Function

Is there a transition with id?

PNML.has_value — Method

PNML.has_xml — Function

has_xml(x) -> Bool

Return true if x has XML attached. Defaults to false.

PNML.in_out — Function

Return tuple of input, output labelled vectors with key of place ids and value of arc inscription's value for use as a transition function. #TODO When do these get called "pre" and "post"?

PNML.inc_indent — Method

Increment the :indent value by indent_width.

PNML.indent — Method

Return string of current indent size in io.

PNML.infos — Method

Content of a ToolInfo.

PNML.initialMarking — Function

Return a labelled vector with key of place id and value of its marking. Marking value is evaluated to be a number (Int or Float64). High-level nets evaluate a Term of the many-sorted algebra to an Int.

PNML.ins — Method

ins(p, transition_id::Symbol) -> LabelledArrays.LArray

Return labeled vector of arcs of p that have transition_id as the target.

PNML.inscription — Function

Return incription value of arc.

PNML.inscription_type — Function

Type of Inscription{inscriptionvaluetype(T)}.

PNML.inscription_value_type — Function

Return value type.

PNML.ispid — Method

ispid(x::Symbol) -> Base.Fix2{typeof(===), Symbol}

Return function to be used like: any(ispid(sym), iteratewithpid)

PNML.labels — Function

labels(x) -> Iterateable

Return iterator of labels attached to x.

PNML.marking — Function

Return marking value of a place.

PNML.marking_type — Function

Return type of Marking{markingvaluetype(T)}.

PNML.marking_value_type — Function

Return value type.

PNML.name — Function

name(x) -> String

Return name String. Default to empty string.

PNML.name — Method

Name of tool to for this tool specific information element.

PNML.nets — Method

nets(model::PNML.PnmlModel) -> Tuple{Vararg{PNML.PnmlNet}}

Return all nets of model.

PNML.nettype — Function

Return the PnmlType subtype representing the flavor (or pntd) of this kind of Petri Net Graph.

See also pnmltype

PNML.number_value — Method

number_value(
    _::Type{T<:Number},
    s::AbstractString
) -> Number

Parse string as a type T <: Number.

PNML.outs — Method

outs(p, transition_id::Symbol) -> LabelledArrays.LArray

Return labeled vector of arcs of p that have transition_id as the source.

PNML.page_type — Function

Type of Page.

PNML.pages — Function

Return vector of all pages.

PNML.pages — Method

Return iterator of Pages directly owned by net.

PNML.parse_and — Method

parse_and(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_arbitraryoperator — Method

parse_arbitraryoperator(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_arbitrarysort — Method

parse_arbitrarysort(node, pntd, reg) -> PNML.PnmlLabel

PNML.parse_arc — Method

parse_arc(node::XMLNode, pntd::PnmlType, idregistry) -> Arc{typeof(pntd), typeof(inscription)}

Construct an Arc with labels specialized for the PnmlType.

PNML.parse_bool — Method

parse_bool(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_booleanconstant — Method

parse_booleanconstant(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_condition — Method

parse_condition(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Condition

Label of transition nodes.

Details

Condition is defined by the ISO Specification as a High-level Annotation, meaning it has <text> and <structure> elements. With all meaning in the element that the <structure> holds evaluating to a boolean value. We extend this to anything that evaluates to a boolean value when treated as a functor.

A Condition should evaluate to a boolean. See AbstractTerm.

PNML.parse_condition_term — Method

parse_condition_term(
    cnode,
    pntd::PnmlType,
    idregistry
) -> PNML.Term

PNML.parse_declaration — Method

parse_declaration(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Declaration

Return Declaration label of 'net' and 'page' nodes. Assume behavior of a High-level Net label in that the meaning is in a <structure>.

Expected format: <declaration> <structure> <declarations> <namedsort/> <namedsort/> ...

PNML.parse_equality — Method

parse_equality(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_file — Method

parse_file(fname::AbstractString) -> PNML.PnmlModel

Build a PnmlModel from a file containing XML. See parse_str and parse_pnml.

PNML.parse_graphics — Method

parse_graphics(node, pntd, reg) -> PNML.Graphics

Arcs, Annotations and Nodes have different graphics semantics. Return a Graphics holding the union of possibilities.

PNML.parse_graphics_coordinate — Method

parse_graphics_coordinate(
    node,
    pntd,
    reg
) -> Union{PNML.Coordinate{Float64}, PNML.Coordinate{Int64}}

Return Coordinate. Specification seems to only use integers, we also allow real numbers.

PNML.parse_graphics_fill — Method

parse_graphics_fill(node, pntd, reg) -> PNML.Fill

Return Fill

PNML.parse_graphics_font — Method

parse_graphics_font(node, pntd, reg) -> PNML.Font

Return Font.

PNML.parse_graphics_line — Method

parse_graphics_line(node, pntd, reg) -> PNML.Line

Return Line.

PNML.parse_hlinitialMarking — Method

parse_hlinitialMarking(
    node::EzXML.Node,
    pntd::AbstractHLCore,
    idregistry::PnmlIDRegistry
) -> PNML.HLMarking{PNML.Term}

High-level initial marking labels are expected to have a Term in the <structure> child. We extend the pnml standard by allowing node content to be numeric: parsed to Int and Float64.

PNML.parse_hlinscription — Method

parse_hlinscription(
    node::EzXML.Node,
    pntd::AbstractHLCore,
    idregistry::PnmlIDRegistry
) -> PNML.HLInscription{PNML.Term}

hlinscriptions are expressions.

PNML.parse_imply — Method

parse_imply(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_inequality — Method

parse_inequality(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_initialMarking — Method

parse_initialMarking(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Marking

PNML.parse_inscription — Method

parse_inscription(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Inscription

PNML.parse_inscription_term — Method

parse_inscription_term(inode, pntd, idregistry) -> PNML.Term

PNML.parse_label — Method

parse_label(
    node::EzXML.Node,
    _::PnmlType,
    _::PnmlIDRegistry
) -> NamedTuple{(:tag, :xml), Tuple{Symbol, EzXML.Node}}

Should not often have a '<label>' tag, this will bark if one is found. Return NamedTuple (tag,node), to defer parsing the xml.

PNML.parse_marking_term — Method

parse_marking_term(marknode, pntd, idregistry) -> PNML.Term

PNML.parse_mulitsetsort — Method

parse_mulitsetsort(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_name — Method

parse_name(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Name

Return Name label holding text value and optional tool & GUI information.

PNML.parse_namedoperator — Method

parse_namedoperator(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.NamedOperator{PNML.VariableDeclaration, PNML.Term}

PNML.parse_namedsort — Method

parse_namedsort(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.NamedSort{PNML.AnyElement}

PNML.parse_net — Function

parse_net(
    node::EzXML.Node,
    idregistry::PnmlIDRegistry
) -> PNML.PnmlNet
parse_net(
    node::EzXML.Node,
    idregistry::PnmlIDRegistry,
    pntd_override::Union{Nothing, PnmlType}
) -> PNML.PnmlNet

Return a PnmlNet`.

PNML.parse_net_1 — Method

Parse net with a defined PnmlType. The PNTD is used to set the marking, inscription, condition and sort type parameters.

PNML.parse_node — Method

parse_node(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> Any

Call any method matching xml node's tagin [tagmap](@ref), otherwise parse as [unclaimed_label](@ref) wrapped in a [PnmlLabel`](@ref). All uses are expected to be pnml labels attached to pnml graph nodes, arcs, nets, pages, that are excluded from this parsing pathway.

PNML.parse_not — Method

parse_not(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_or — Method

parse_or(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_page! — Method

parse_page!(tup, node, pntd, idregistry) -> Page

Place Page in pagedict using id as the key.

PNML.parse_place — Method

parse_place(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Place

PNML.parse_pnml — Method

parse_pnml(xmlnode, idregistry) -> PnmlModel

Start parse from the root node of a well formed pnml XML document. Return a PnmlModel holding one or more PnmlNet.

PNML.parse_productsort — Method

parse_productsort(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_refPlace — Method

parse_refPlace(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.RefPlace

PNML.parse_refTransition — Method

parse_refTransition(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.RefTransition

PNML.parse_sort — Method

parse_sort(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.AnyElement

Sorts are found within a <structure> element.

PNML.parse_sorttype_term — Method

parse_sorttype_term(typenode, pntd, idregistry) -> Any

A concrete subtype of AbstractSort. Built from many different elements that contain a Sort: type, namedsort, variabledecl, multisetsort, productsort, numberconstant, partition...

Sort = BuiltInSort | MultisetSort | ProductSort | UserSort

PNML.parse_str — Method

parse_str(str::AbstractString) -> PNML.PnmlModel

Build a PnmlModel from a string containing XML. See parse_file and parse_pnml.

PNML.parse_structure — Method

parse_structure(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Structure

Return Structure holding a <structure>. Should be inside of an label. A "claimed" label usually elids the <structure> level (does not call this method).

PNML.parse_term — Method

parse_term(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

There will be no XML node 'term'. Instead it is the interpertation of the child of some 'structure' or def elements. The PNML specification describes Terms and Sorts as abstract types for the 'structure' element of some HLAnnotation.

PNML.parse_text — Method

parse_text(
    node::EzXML.Node,
    _::PnmlType,
    _::PnmlIDRegistry
) -> String

Return the stripped string of node's content.

PNML.parse_tokengraphics — Method

parse_tokengraphics(
    node::EzXML.Node,
    pntd::PnmlType,
    reg
) -> PNML.TokenGraphics

High-level place-transition nets (HL-PTNet) have a toolspecific structure defined for token graphics. Contains <tokenposition>s. Returns vector of Coordinates.

PNML.parse_tokenposition — Method

parse_tokenposition(
    node,
    pntd,
    reg
) -> Union{PNML.Coordinate{Float64}, PNML.Coordinate{Int64}}

Return Cartesian Coordinate relative to containing element.

PNML.parse_toolspecific — Method

parse_toolspecific(node, pntd, reg) -> PNML.ToolInfo

Return ToolInfo with tool & version attributes and content.

The content can be one or more well-formed xml elements. Each are wrapped in a PnmlLabel.

PNML.parse_transition — Method

parse_transition(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Transition{_A, PNML.Condition} where _A

PNML.parse_tuple — Method

parse_tuple(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_type — Function

Defines the "sort" of tokens held by the place and semantics of the marking. NB: The "type" of a place from many-sorted algebra is different from the Petri Net "type" of a net or "pntd". Neither is directly a julia type.

PNML.parse_unknowndecl — Method

parse_unknowndecl(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.UnknownDeclaration

PNML.parse_unparsed — Method

parse_unparsed(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.PnmlLabel

PNML.parse_useroperator — Method

parse_useroperator(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.UserOperator

PNML.parse_usersort — Method

parse_usersort(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
)

PNML.parse_variable — Method

parse_variable(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Variable

PNML.parse_variabledecl — Method

parse_variabledecl(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.VariableDeclaration{PNML.AnyElement}

PNML.pid — Function

pid(x) -> Symbol

Return pnml id symbol.

PNML.place — Function

Return the place with id.

PNML.place_idset — Function

Return vector of all place IDs.

PNML.place_type — Function

Type of Place.

PNML.places — Function

Return vector of all places.

PNML.pnml_namespace — Method

pnml_namespace(
    node::EzXML.Node;
    missing_ns_fatal,
    default_ns
) -> String

Return namespace. When node does not have a namespace return default value pnml_ns and warn or throw an error.

PNML.pnmlnet_type — Function

PNML.rate — Method

rate(transition) -> Float64

Return rate value of transition. Missing rate labels are defaulted to 0.0.

PNML.rate_value_type — Function

Return rate value type

PNML.rates — Function

Return a transition-id labelled vector of rate values for transitions of net.

PNML.refid — Method

Return the id of the referenced node.

PNML.refplace — Function

Return reference place matching id.

PNML.refplace_idset — Function

refplace_idset(x) -> Set{Symbol} #TODO iterator?

Return reference place pnml ids.

PNML.refplace_type — Function

Tyoe of RefPlace.

PNML.refplaces — Function

Return vector of all reference places.

PNML.reftransition — Function

Return reference transition matching id.

PNML.reftransition_idset — Function

reftransition_idset(x) -> Set{Symbol} #TODO iterator?

Return reference transition pnml ids.

PNML.reftransition_type — Function

Type of RefTransition.

PNML.reftransitions — Function

Return vector of all reference transitions.

PNML.save_config! — Function

save_config!(config::PnmlConfig=CONFIG)

Save a PNML.jl configuration config (by default, PNML) to your LocalPreferences.toml file using Preferences.jl.

The saved preferences will be automatically loaded next time you using PNML

Examples

julia> using PNML

julia> PNML.CONFIG.verbose = true;

julia> PNML.CONFIG.warn_on_unclaimed = true;     # Customize some defaults

julia> PNML.save_config!(PNML.CONFIG); # Will be automatically read next time you `using PNML`

PNML.source — Method

source(arc) -> Symbol

Return symbol of source of arc.

PNML.src_arcs — Function

Return arcs that have a source of transition id.

See also all_arcs, tgt_arcs.

PNML.structure — Method

Return structure field.

PNML.tag — Function

tag(x) -> Symbol

Return tag symbol.

PNML.target — Method

target(arc) -> Symbol

Return symbol of target of arc.

PNML.term_value_type — Function

Return value type.

PNML.text — Method

Return text field

PNML.text_content — Function

Find first :text and return content as string.

PNML.tgt_arcs — Function

Return arcs that have a target of transition id.

See also all_arcs, src_arcs.

PNML.transition — Function

PNML.transition_function — Function

Transition function of a Petri Net. Each transition has an input vector and an output vector. Each labelled vector is indexed by the place on the other end of the arc. Values are inscriptions of the arc.

PNML.transition_idset — Function

PNML.transition_type — Function

Type of TYransition.

PNML.transitions — Function

Return vector of all transitions.

PNML.unclaimed_label — Method

unclaimed_label(
    node::EzXML.Node,
    pntd::PnmlType,
    _::PnmlIDRegistry
) -> Pair{Symbol, Vector{PNML.AnyXmlNode}}

Return tag => AnyXmlNode holding well formed XML tree/forest.

The main use-case is to be wrapped in a PnmlLabel, AnyElement, Term or other specialized label.

PNML.version — Method

Version of tool for this tool specific information element.

PNML.xmlnode — Function

xmlnode(x) -> XMLNode

Return attached xml node.

PNML.xmlroot — Method

xmlroot(s::AbstractString) -> EzXML.Node

Parse string s into EzXML node.

PNML.@xml_str — Macro

Parse string into EzXML node.

See xmlroot.

PnmlTypeDefs Reference

PNML.PnmlTypeDefs — Module

Petri Net Type Definition (pntd) URI mapped to PnmlType subtype singleton.

PNML.PnmlTypeDefs.default_pntd_map — Constant

Map from Petri Net Type Definition (pntd) URI to Symbol. Allows multiple strings to map to the same pntd.

There is a companion map pnmltype_map that takes the symbol to a type object.

The URI is a string and may be the full URL of a pntd schema, just the schema file name, or a placeholder for a future schema.

For readability, the 'pntd symbol' should match the name used in the URI with inconvinient characters removed or replaced. For example, '-' is replaced by '_'.

PNML.PnmlTypeDefs.pnmltype_map — Constant

The key Symbols are the supported kinds of Petri Nets.

PNML.PnmlTypeDefs.AbstractContinuousNet — Type

abstract type AbstractContinuousNet <: PnmlType

Uses floating point numbers for markings, inscriptions. Most of the functionality is shared with AbstractPnmlCore. This seperates the

PNML.PnmlTypeDefs.AbstractHLCore — Type

abstract type AbstractHLCore <: PnmlType

Base of High Level Petri Net pntds which add large extensions to PNML core. hlinitialMarking, hlinscription, and defined label structures.

See HLCoreNet, SymmetricNet, PT_HLPNG and others.

PNML.PnmlTypeDefs.AbstractPnmlCore — Type

abstract type AbstractPnmlCore <: PnmlType

Base of token/integer-based Petri Net pntds.

See PnmlCoreNet, PTNet and others.

PNML.PnmlTypeDefs.ContinuousNet — Type

struct ContinuousNet <: AbstractContinuousNet

TODO: Continuous Petri Net

PNML.PnmlTypeDefs.HLCoreNet — Type

struct HLCoreNet <: AbstractHLCore

HLCoreNet can be used for generic high-level nets. We try to implement and test all function at `PnmlCoreNet level, but expect to find use for a concrete type at this level for testing high-level extensions.

PNML.PnmlTypeDefs.HLPNG — Type

struct HLPNG <: AbstractHLCore

High-Level Petri Net Graphs (HLPNGs) are the most intricate High-Level Petri Net schema. It extends SymmetricNet, including with

declarations for sorts and functions (ArbitraryDeclarations)

sorts for Integer, String, and List

PNML.PnmlTypeDefs.OpenNet — Type

struct OpenNet <: AbstractContinuousNet

TODO: Open Petri Net

PNML.PnmlTypeDefs.PTNet — Type

struct PTNet <: AbstractPnmlCore

Place-Transition Petri Nets add small extensions to core PNML. Integer-valued initialMarking and inscription.

The grammer file is ptnet.pnml so we name it PTNet. Note that 'PT' is often the prefix for XML tags specialized for this net type.

PNML.PnmlTypeDefs.PT_HLPNG — Type

struct PT_HLPNG <: AbstractHLCore

Place-Transition Net in HLCoreNet notation.

PNML.PnmlTypeDefs.PnmlCoreNet — Type

struct PnmlCoreNet <: AbstractPnmlCore

The most minimal concrete Petri Net.

Used to implement and test the core PNML support. Covers the complete graph infrastructure including labels attached to nodes and arcs.

PNML.PnmlTypeDefs.PnmlType — Type

abstract type PnmlType

Abstract root of a dispatch type based on Petri Net Type Definition (pntd).

Each Petri Net Markup Language (PNML) network element will have a single pntd URI as a required 'type' XML attribute. That URI should refer to a RelaxNG schema defining the syntax and semantics of the XML model.

Selected abbreviations, URIs that do not resolve to a valid schema file, are also allowed.

Refer to pntd_symbol and pnmltype for how to get from the URI to a singleton.

PNML.PnmlTypeDefs.StochasticNet — Type

struct StochasticNet <: AbstractHLCore

TODO: Stochastic Petri Net

PNML.PnmlTypeDefs.SymmetricNet — Type

struct SymmetricNet <: AbstractHLCore

Symmetric Petri Net is the best-worked use case in the primer and ISO specification part 2.

PNML.PnmlTypeDefs.TimedNet — Type

struct TimedNet <: AbstractHLCore

TODO: Timed Petri Net

PNML.PnmlTypeDefs.add_nettype! — Method

add_nettype!(
    elements::AbstractDict,
    s::Symbol,
    pntd::PnmlType
) -> PnmlType

Add or replace mapping from symbol s to nettype dispatch singleton t.

PNML.PnmlTypeDefs.iscontinuous — Function

Values are floating point.

PNML.PnmlTypeDefs.isdiscrete — Function

Values are integers.

PNML.PnmlTypeDefs.ishighlevel — Function

Values are many-sorted.

PNML.PnmlTypeDefs.pnmltype — Function

pnmltype(pntd::T, reg)
pnmltype(uri::AbstractString, reg)
pnmltype(s::Symbol; pnmltype_map=pnmltype_map, reg)

Map either a text string or a symbol to a dispatch type singlton.

While that string may be a URI for a pntd, we treat it as a simple string without parsing. The PnmlTypeDefs.pnmltype_map and PnmlTypeDefs.default_pntd_map are both assumed to be correct here.

Unknown or empty uri will map to symbol :pnmlcore. Unknown symbol throws a DomainError exception.

Examples

julia> pnmltype(PnmlCoreNet())
PnmlCoreNet()

julia> pnmltype("nonstandard")
PnmlCoreNet()

julia> pnmltype(:symmetric)
SymmetricNet()

PNML.PnmlTypeDefs.pntd_symbol — Method

pntd_symbol(s::AbstractString) -> Symbol

Map string s to a pntd symbol using default_pntd_map. Any unknown s is mapped to :pnmlcore. Returned symbol is suitable for pnmltype to use to index into pnmltype_map.

Examples

julia> PnmlTypeDefs.pntd_symbol("foo")
:pnmlcore

PnmlIDRegistrys Reference

PNML.PnmlIDRegistrys — Module

Petri Net Markup Language identifier registry.

PNML.PnmlIDRegistrys.PnmlIDRegistry — Type

Holds a set of pnml id symbols and a lock to allow safe reentrancy.

struct PnmlIDRegistry{L}

ids::Set{Symbol}

lk::Any

Base.isempty — Method

isempty(
    idregistry::PnmlIDRegistry{L<:Base.AbstractLock}
) -> Bool

Is the set of id symbols empty?

PNML.PnmlIDRegistrys.isregistered — Method

isregistered(reg::PnmlIDRegistry, s::AbstractString) -> Bool

Return true if s is registered in reg.

PNML.PnmlIDRegistrys.register_id! — Method

register_id!(
    idregistry::PnmlIDRegistry,
    s::AbstractString
) -> Symbol

Register id symbol and return the symbol.

PNML.PnmlIDRegistrys.registry — Function

registry([lock]) -> PnmlIDRegistry

Construct a PNML ID registry using the supplied lock or a ReentrantLock`.

PNML.PnmlIDRegistrys.reset! — Method

reset!(
    idregistry::PnmlIDRegistry{L<:Base.AbstractLock}
) -> Any

Empty the set of id symbols. Use case is unit tests. In normal use it should never be needed.