Layers of Abstraction

Graphics are elided from this discussion

Everywhere there are ToolInfos in this discussion one may assume that there is also an optional Graphics possible.

While we parse such XML into "containers of strings" and Coordinates. No further use is implemented or planned. And no discussion of use is present.

Intermediate Representation

The intermediate representation (IR) is between the XML model and a "usable" network. Many different flavors of Petri Nets are expected to be implemented using the IR.

The IR is constructed by traversing the XML and using tag names as dictonary keys.

In the first part of parsing, a named tuple is filled with appropriate initial values for each xml tag. Then optional child keys have values bound as they are parsed.

The second part of parsing instantiates objects using the named tuple as input.

The structure of the IR follows the tree structure of a well-formed XML document and the PNML specification.

XML attribute names and child element tag names are used for keys of the same dictonary. The pnml specification/schemas do not use colliding names. However, the <toolspecific> tag's content is not required to be valid pnml, just well-formed XML. We assume nobody would use colliding names intentionally.

The crude structure required by the pnmlcore schema: PnmlModel

Net

Page

- Places, Marking
- Transitions, Condition
- Arcs, Inscription
- Toolinfos
- Labels
- Subpages

Name

- Toolinfos
- Labels

The IR is implemented under the assumption the the input pnml file is valid. All tags are assumed to be meaningful to the resulting network. The pnmlcore schema requires undefined tags will be considered pnml labels. The IR is capable of handling arbitrary labels. Many label tags from higherlevel pnml schemas are recognized by the IR parsers.

While the Petri Net Type Definition (pntd) is present in every valid net, it was not necessary to consult the type during creation of the IR. It is expected that conforming to pntd will be done at a higher level.

Some parts of pnml are complicated. Not yet completed bits may be implemented as wrappers holdind unparsed XML. In fact, parts of pnml are specified as holding any well-formed XML.

The crude structure required by the pnmlcore schema: PnmlModel

Net

Pages

Places, Marking, unclaimed labels [SortType] [Capacity]

Transitions, Condition, unclaimed labels [rate]

Arcs, Inscription, unclaimed labels [ArcType]

Toolinfos [TokenGraphics]

Labels unclaimed, [Declaration]

Subpages

Name (everybody has name)

Toolinfos

Labels

It is expected that conforming to pntd will be done at a higher level.

Level	Sorts

Core	Bool, Int

PT	Bool, Int

Continuous	Bool, Float64

Hybrid	Bool, Int, Float64

high-level	Bool, Int, Term

Symmetric	Bool, Int, Term

PTHLPNG	Bool, Int, Term

Core Layer

What is permitted by the specification in a XML file will be a subset of the implementation.

Concepts from High-Level Petri Nets will be used in the Core layer.

PNTD	Sort value

PnmlType	Int

AbstractContinuousNet	Float64

AbstractHLCore	DotConstant())

Use Union{Bool, Int, Float64, DictType} as the set of types that a Term's can contain. Consider Bool, Int, Float64 as builtin-sorts, and DictType as "user defined" sorts.

Type Lookup Layers

Petri Net Graph Object Types are parameterized by Label Types. What labels are "allowed" (syntax vs. semantics vs. schema vs. specification) is parameterized on the PNTD (Petri Net Type Definition).

See PnmlNets & AbstractPnmlObjects

pnmltype_map

PNML.PnmlTypeDefs.pnmltype_map — Constant

The key Symbols are the supported kinds of Petri Nets. Maps to singletons.

PNML.PnmlTypeDefs.pnmltype_map

IdDict{Symbol, PnmlType} with 7 entries:
  :pnmlcore   => PnmlCoreNet()
  :symmetric  => SymmetricNet()
  :ptnet      => PTNet()
  :continuous => ContinuousNet()
  :hlnet      => HLPNG()
  :hlcore     => HLCoreNet()
  :pt_hlpng   => PT_HLPNG()

pnmlnet_type

PNML.pnmlnet_type — Function

PnmlCoreNet()   -> PNML.PnmlNet{PnmlCoreNet, PNML.Place{PnmlCoreNet, PNML.Marking{Int64}}, PNML.Transition{PnmlCoreNet, PNML.Condition}, PNML.Arc{PNML.Inscription{Int64}}, PNML.RefPlace, PNML.RefTransition}
SymmetricNet()  -> PNML.PnmlNet{SymmetricNet, PNML.Place{SymmetricNet, PNML.HLMarking}, PNML.Transition{SymmetricNet, PNML.Condition}, PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}, PNML.RefPlace, PNML.RefTransition}
PTNet()         -> PNML.PnmlNet{PTNet, PNML.Place{PTNet, PNML.Marking{Int64}}, PNML.Transition{PTNet, PNML.Condition}, PNML.Arc{PNML.Inscription{Int64}}, PNML.RefPlace, PNML.RefTransition}
ContinuousNet() -> PNML.PnmlNet{ContinuousNet, PNML.Place{ContinuousNet, PNML.Marking{Float64}}, PNML.Transition{ContinuousNet, PNML.Condition}, PNML.Arc{PNML.Inscription{Float64}}, PNML.RefPlace, PNML.RefTransition}
HLPNG()         -> PNML.PnmlNet{HLPNG, PNML.Place{HLPNG, PNML.HLMarking}, PNML.Transition{HLPNG, PNML.Condition}, PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}, PNML.RefPlace, PNML.RefTransition}
HLCoreNet()     -> PNML.PnmlNet{HLCoreNet, PNML.Place{HLCoreNet, PNML.HLMarking}, PNML.Transition{HLCoreNet, PNML.Condition}, PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}, PNML.RefPlace, PNML.RefTransition}
PT_HLPNG()      -> PNML.PnmlNet{PT_HLPNG, PNML.Place{PT_HLPNG, PNML.HLMarking}, PNML.Transition{PT_HLPNG, PNML.Condition}, PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}, PNML.RefPlace, PNML.RefTransition}

page_type

PnmlCoreNet()   -> PNML.Page{PnmlCoreNet, PNML.Place{PnmlCoreNet, PNML.Marking{Int64}}, PNML.Transition{PnmlCoreNet, PNML.Condition}, PNML.Arc{PNML.Inscription{Int64}}, PNML.RefPlace, PNML.RefTransition}
SymmetricNet()  -> PNML.Page{SymmetricNet, PNML.Place{SymmetricNet, PNML.HLMarking}, PNML.Transition{SymmetricNet, PNML.Condition}, PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}, PNML.RefPlace, PNML.RefTransition}
PTNet()         -> PNML.Page{PTNet, PNML.Place{PTNet, PNML.Marking{Int64}}, PNML.Transition{PTNet, PNML.Condition}, PNML.Arc{PNML.Inscription{Int64}}, PNML.RefPlace, PNML.RefTransition}
ContinuousNet() -> PNML.Page{ContinuousNet, PNML.Place{ContinuousNet, PNML.Marking{Float64}}, PNML.Transition{ContinuousNet, PNML.Condition}, PNML.Arc{PNML.Inscription{Float64}}, PNML.RefPlace, PNML.RefTransition}
HLPNG()         -> PNML.Page{HLPNG, PNML.Place{HLPNG, PNML.HLMarking}, PNML.Transition{HLPNG, PNML.Condition}, PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}, PNML.RefPlace, PNML.RefTransition}
HLCoreNet()     -> PNML.Page{HLCoreNet, PNML.Place{HLCoreNet, PNML.HLMarking}, PNML.Transition{HLCoreNet, PNML.Condition}, PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}, PNML.RefPlace, PNML.RefTransition}
PT_HLPNG()      -> PNML.Page{PT_HLPNG, PNML.Place{PT_HLPNG, PNML.HLMarking}, PNML.Transition{PT_HLPNG, PNML.Condition}, PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}, PNML.RefPlace, PNML.RefTransition}

place_type

PnmlCoreNet()   -> PNML.Place{PnmlCoreNet, PNML.Marking{Int64}}
SymmetricNet()  -> PNML.Place{SymmetricNet, PNML.HLMarking}
PTNet()         -> PNML.Place{PTNet, PNML.Marking{Int64}}
ContinuousNet() -> PNML.Place{ContinuousNet, PNML.Marking{Float64}}
HLPNG()         -> PNML.Place{HLPNG, PNML.HLMarking}
HLCoreNet()     -> PNML.Place{HLCoreNet, PNML.HLMarking}
PT_HLPNG()      -> PNML.Place{PT_HLPNG, PNML.HLMarking}

transition_type

PnmlCoreNet()   -> PNML.Transition{PnmlCoreNet, PNML.Condition}
SymmetricNet()  -> PNML.Transition{SymmetricNet, PNML.Condition}
PTNet()         -> PNML.Transition{PTNet, PNML.Condition}
ContinuousNet() -> PNML.Transition{ContinuousNet, PNML.Condition}
HLPNG()         -> PNML.Transition{HLPNG, PNML.Condition}
HLCoreNet()     -> PNML.Transition{HLCoreNet, PNML.Condition}
PT_HLPNG()      -> PNML.Transition{PT_HLPNG, PNML.Condition}

arc_type

PnmlCoreNet()   -> PNML.Arc{PNML.Inscription{Int64}}
SymmetricNet()  -> PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}
PTNet()         -> PNML.Arc{PNML.Inscription{Int64}}
ContinuousNet() -> PNML.Arc{PNML.Inscription{Float64}}
HLPNG()         -> PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}
HLCoreNet()     -> PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}
PT_HLPNG()      -> PNML.Arc{PNML.HLInscription{PNML.PnmlMultiset}}

Label Types

AbstractLabels are parameterized by Value Types.

marking_type

PnmlCoreNet()   -> PNML.Marking{Int64}
SymmetricNet()  -> PNML.HLMarking
PTNet()         -> PNML.Marking{Int64}
ContinuousNet() -> PNML.Marking{Float64}
HLPNG()         -> PNML.HLMarking
HLCoreNet()     -> PNML.HLMarking
PT_HLPNG()      -> PNML.HLMarking

condition_type

PnmlCoreNet()   -> PNML.Condition
SymmetricNet()  -> PNML.Condition
PTNet()         -> PNML.Condition
ContinuousNet() -> PNML.Condition
HLPNG()         -> PNML.Condition
HLCoreNet()     -> PNML.Condition
PT_HLPNG()      -> PNML.Condition

inscription_type

PnmlCoreNet()   -> PNML.Inscription{Int64}
SymmetricNet()  -> PNML.HLInscription{PNML.PnmlMultiset}
PTNet()         -> PNML.Inscription{Int64}
ContinuousNet() -> PNML.Inscription{Float64}
HLPNG()         -> PNML.HLInscription{PNML.PnmlMultiset}
HLCoreNet()     -> PNML.HLInscription{PNML.PnmlMultiset}
PT_HLPNG()      -> PNML.HLInscription{PNML.PnmlMultiset}

refplace_type

PnmlCoreNet()   -> PNML.RefPlace
SymmetricNet()  -> PNML.RefPlace
PTNet()         -> PNML.RefPlace
ContinuousNet() -> PNML.RefPlace
HLPNG()         -> PNML.RefPlace
HLCoreNet()     -> PNML.RefPlace
PT_HLPNG()      -> PNML.RefPlace

reftransition_type

PnmlCoreNet()   -> PNML.RefTransition
SymmetricNet()  -> PNML.RefTransition
PTNet()         -> PNML.RefTransition
ContinuousNet() -> PNML.RefTransition
HLPNG()         -> PNML.RefTransition
HLCoreNet()     -> PNML.RefTransition
PT_HLPNG()      -> PNML.RefTransition

Value Types

TBD