Layers of Abstraction

Graphics are elided from this discussion

Everywhere there are ToolInfos in this discussion one may assume that there is also an optional Graphics possible.

While we parse such XML into "containers of strings" and Coordinates. No further use is implemented or planned. And no discussion of use is present.

Intermediate Representation

The intermediate representation (IR) is between the XML model and a "usable" network. Many different flavors of Petri Nets are expected to be implemented using the IR.

XML attribute names and child element tag names are used for keys of the same dictonary. The pnml specification/schemas do not use colliding names. However, the <toolspecific> tag's content is not required to be valid pnml, just well-formed XML. We assume nobody would use colliding names intentionally.

The crude structure required by the pnmlcore schema: PnmlModel

Net

Page

- Places, Marking
- Transitions, Condition
- Arcs, Inscription
- Toolinfos
- Labels
- Subpages

Name

- Toolinfos
- Labels

The IR is implemented under the assumption the the input pnml file is valid. All tags are assumed to be meaningful to the resulting network. The pnmlcore schema requires undefined tags on objects will be considered pnml labels. The IR is capable of handling arbitrary labels. Many label tags from higherlevel pnml schemas are recognized by the IR core parser.

Some parts of pnml are complicated. Not yet completed bits may be implemented as wrappers holdind unparsed XML. In fact, parts of pnml are specified as holding any well-formed XML.

The crude structure required by the pnmlcore schema: PnmlModel

Net

Pages

Places, Marking, unclaimed labels [SortType] [Capacity]

Transitions, Condition, unclaimed labels [TransitionRate]

Arcs, Inscription, unclaimed labels [ArcType]

Toolinfos [Labels.TokenGraphics]

Labels unclaimed, [Declaration]

Subpages

Name (everybody has name)

Toolinfos

Labels

It is expected that conforming to pntd will be done at a higher level.

Core Layer

What is permitted by the specification will be a subset of the implementation.

Concepts from High-Level Petri Nets will be used in the Core layer.

Data Storage

Declaration Dictionaries

The net-global storage resides here.

DeclDict holds unordered collections indexed by REFID symbol.

variabledecls
namedsorts
arbitrarysorts
partitionsorts
namedoperators
arbitraryoperators
partitionops
feconstants
usersorts
useroperators

The XML file format allows declarations to be declared in <net> and <page> elements.

All Declaration labels for a net share the same DeclDict as a ScopedValue.

XML XPath is used to gather this information before parsing the rest of the elements. Allows using DeclDict while parsing.

Net Data Dictionaries

This is where the graph node storage resides.

PnmlNetData contains ordered collections of the graph node objects, indexed by REFID symbols.

place_dict
transition_dict
arc_dict
refplace_dict
reftransition_dict

The XML file format distributes a <net> over one or more <page>s. As the pages are parsed, the nodes are appended to a PnmlNetData dictionary and a PnmlNetKeys set.

The PnmlNetData dictionaries maintain insertion order.

Each graph node may have labels attached. What labels depends on the PnmlTypeDefs

ID Sets

PnmlNetKeys contains ordered sets of REFID symbols.

page_set
place_set
transition_set
arc_set
reftransition_set
refplace_set

Each PnmlNetKeys set maintains insertion order.

Uses REFIDs to keep track of which page owns which graph nodes or sub-page. We always use the flatten_pages! version. Testing of non-flattened nets is very minimal.

Warning

After flatten_pages! the PnmlNetKeys of the only remaining page are assumed to be the same as the keys of corresponding PnmlNetData dictionary.

Type Lookup Layers

Petri Net Graph Object Types are parameterized by Label Types. What labels are "allowed" (syntax vs. semantics vs. schema vs. specification) is parameterized on the PNTD (Petri Net Type Definition).

See PnmlNets & AbstractPnmlObjects

pnmltype_map

PNML.PnmlTypeDefs.pnmltype_map — Constant

The key Symbols are the supported kinds of Petri Nets. Maps to singletons.

PNML.PnmlTypeDefs.pnmltype_map

IdDict{Symbol, PNML.PnmlTypeDefs.PnmlType} with 7 entries:
  :pnmlcore   => PnmlCoreNet()
  :symmetric  => SymmetricNet()
  :ptnet      => PTNet()
  :continuous => ContinuousNet()
  :hlnet      => HLPNG()
  :hlcore     => HLCoreNet()
  :pt_hlpng   => PT_HLPNG()

pnmlnet_type

PNML.pnmlnet_type — Function

PNML.PnmlTypeDefs.PnmlCoreNet() -> PnmlNet{PNML.PnmlTypeDefs.PnmlCoreNet, Place{PNML.PnmlTypeDefs.PnmlCoreNet, PNML.Labels.Marking}, Transition{PNML.PnmlTypeDefs.PnmlCoreNet, PNML.Labels.Condition}, Arc{PNML.Labels.Inscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.SymmetricNet() -> PnmlNet{PNML.PnmlTypeDefs.SymmetricNet, Place{PNML.PnmlTypeDefs.SymmetricNet, PNML.Labels.HLMarking}, Transition{PNML.PnmlTypeDefs.SymmetricNet, PNML.Labels.Condition}, Arc{PNML.Labels.HLInscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.PTNet() -> PnmlNet{PNML.PnmlTypeDefs.PTNet, Place{PNML.PnmlTypeDefs.PTNet, PNML.Labels.Marking}, Transition{PNML.PnmlTypeDefs.PTNet, PNML.Labels.Condition}, Arc{PNML.Labels.Inscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.ContinuousNet() -> PnmlNet{PNML.PnmlTypeDefs.ContinuousNet, Place{PNML.PnmlTypeDefs.ContinuousNet, PNML.Labels.Marking}, Transition{PNML.PnmlTypeDefs.ContinuousNet, PNML.Labels.Condition}, Arc{PNML.Labels.Inscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.HLPNG() -> PnmlNet{PNML.PnmlTypeDefs.HLPNG, Place{PNML.PnmlTypeDefs.HLPNG, PNML.Labels.HLMarking}, Transition{PNML.PnmlTypeDefs.HLPNG, PNML.Labels.Condition}, Arc{PNML.Labels.HLInscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.HLCoreNet() -> PnmlNet{PNML.PnmlTypeDefs.HLCoreNet, Place{PNML.PnmlTypeDefs.HLCoreNet, PNML.Labels.HLMarking}, Transition{PNML.PnmlTypeDefs.HLCoreNet, PNML.Labels.Condition}, Arc{PNML.Labels.HLInscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.PT_HLPNG() -> PnmlNet{PNML.PnmlTypeDefs.PT_HLPNG, Place{PNML.PnmlTypeDefs.PT_HLPNG, PNML.Labels.HLMarking}, Transition{PNML.PnmlTypeDefs.PT_HLPNG, PNML.Labels.Condition}, Arc{PNML.Labels.HLInscription}, RefPlace, RefTransition}

page_type

PNML.PnmlTypeDefs.PnmlCoreNet() -> Page{PNML.PnmlTypeDefs.PnmlCoreNet, Place{PNML.PnmlTypeDefs.PnmlCoreNet, PNML.Labels.Marking}, Transition{PNML.PnmlTypeDefs.PnmlCoreNet, PNML.Labels.Condition}, Arc{PNML.Labels.Inscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.SymmetricNet() -> Page{PNML.PnmlTypeDefs.SymmetricNet, Place{PNML.PnmlTypeDefs.SymmetricNet, PNML.Labels.HLMarking}, Transition{PNML.PnmlTypeDefs.SymmetricNet, PNML.Labels.Condition}, Arc{PNML.Labels.HLInscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.PTNet() -> Page{PNML.PnmlTypeDefs.PTNet, Place{PNML.PnmlTypeDefs.PTNet, PNML.Labels.Marking}, Transition{PNML.PnmlTypeDefs.PTNet, PNML.Labels.Condition}, Arc{PNML.Labels.Inscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.ContinuousNet() -> Page{PNML.PnmlTypeDefs.ContinuousNet, Place{PNML.PnmlTypeDefs.ContinuousNet, PNML.Labels.Marking}, Transition{PNML.PnmlTypeDefs.ContinuousNet, PNML.Labels.Condition}, Arc{PNML.Labels.Inscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.HLPNG() -> Page{PNML.PnmlTypeDefs.HLPNG, Place{PNML.PnmlTypeDefs.HLPNG, PNML.Labels.HLMarking}, Transition{PNML.PnmlTypeDefs.HLPNG, PNML.Labels.Condition}, Arc{PNML.Labels.HLInscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.HLCoreNet() -> Page{PNML.PnmlTypeDefs.HLCoreNet, Place{PNML.PnmlTypeDefs.HLCoreNet, PNML.Labels.HLMarking}, Transition{PNML.PnmlTypeDefs.HLCoreNet, PNML.Labels.Condition}, Arc{PNML.Labels.HLInscription}, RefPlace, RefTransition}
PNML.PnmlTypeDefs.PT_HLPNG() -> Page{PNML.PnmlTypeDefs.PT_HLPNG, Place{PNML.PnmlTypeDefs.PT_HLPNG, PNML.Labels.HLMarking}, Transition{PNML.PnmlTypeDefs.PT_HLPNG, PNML.Labels.Condition}, Arc{PNML.Labels.HLInscription}, RefPlace, RefTransition}

place_type

PNML.PnmlTypeDefs.PnmlCoreNet() -> Place{PNML.PnmlTypeDefs.PnmlCoreNet, PNML.Labels.Marking}
PNML.PnmlTypeDefs.SymmetricNet() -> Place{PNML.PnmlTypeDefs.SymmetricNet, PNML.Labels.HLMarking}
PNML.PnmlTypeDefs.PTNet() -> Place{PNML.PnmlTypeDefs.PTNet, PNML.Labels.Marking}
PNML.PnmlTypeDefs.ContinuousNet() -> Place{PNML.PnmlTypeDefs.ContinuousNet, PNML.Labels.Marking}
PNML.PnmlTypeDefs.HLPNG() -> Place{PNML.PnmlTypeDefs.HLPNG, PNML.Labels.HLMarking}
PNML.PnmlTypeDefs.HLCoreNet() -> Place{PNML.PnmlTypeDefs.HLCoreNet, PNML.Labels.HLMarking}
PNML.PnmlTypeDefs.PT_HLPNG() -> Place{PNML.PnmlTypeDefs.PT_HLPNG, PNML.Labels.HLMarking}

transition_type

PNML.PnmlTypeDefs.PnmlCoreNet() -> Transition{PNML.PnmlTypeDefs.PnmlCoreNet, PNML.Labels.Condition}
PNML.PnmlTypeDefs.SymmetricNet() -> Transition{PNML.PnmlTypeDefs.SymmetricNet, PNML.Labels.Condition}
PNML.PnmlTypeDefs.PTNet() -> Transition{PNML.PnmlTypeDefs.PTNet, PNML.Labels.Condition}
PNML.PnmlTypeDefs.ContinuousNet() -> Transition{PNML.PnmlTypeDefs.ContinuousNet, PNML.Labels.Condition}
PNML.PnmlTypeDefs.HLPNG() -> Transition{PNML.PnmlTypeDefs.HLPNG, PNML.Labels.Condition}
PNML.PnmlTypeDefs.HLCoreNet() -> Transition{PNML.PnmlTypeDefs.HLCoreNet, PNML.Labels.Condition}
PNML.PnmlTypeDefs.PT_HLPNG() -> Transition{PNML.PnmlTypeDefs.PT_HLPNG, PNML.Labels.Condition}

arc_type

PNML.PnmlTypeDefs.PnmlCoreNet() -> Arc{PNML.Labels.Inscription}
PNML.PnmlTypeDefs.SymmetricNet() -> Arc{PNML.Labels.HLInscription}
PNML.PnmlTypeDefs.PTNet() -> Arc{PNML.Labels.Inscription}
PNML.PnmlTypeDefs.ContinuousNet() -> Arc{PNML.Labels.Inscription}
PNML.PnmlTypeDefs.HLPNG() -> Arc{PNML.Labels.HLInscription}
PNML.PnmlTypeDefs.HLCoreNet() -> Arc{PNML.Labels.HLInscription}
PNML.PnmlTypeDefs.PT_HLPNG() -> Arc{PNML.Labels.HLInscription}

Label Types

AbstractLabels are parameterized by Value Types.

marking_type

PNML.PnmlTypeDefs.PnmlCoreNet() -> PNML.Labels.Marking
PNML.PnmlTypeDefs.SymmetricNet() -> PNML.Labels.HLMarking
PNML.PnmlTypeDefs.PTNet() -> PNML.Labels.Marking
PNML.PnmlTypeDefs.ContinuousNet() -> PNML.Labels.Marking
PNML.PnmlTypeDefs.HLPNG() -> PNML.Labels.HLMarking
PNML.PnmlTypeDefs.HLCoreNet() -> PNML.Labels.HLMarking
PNML.PnmlTypeDefs.PT_HLPNG() -> PNML.Labels.HLMarking

condition_type

PNML.PnmlTypeDefs.PnmlCoreNet() -> PNML.Labels.Condition
PNML.PnmlTypeDefs.SymmetricNet() -> PNML.Labels.Condition
PNML.PnmlTypeDefs.PTNet() -> PNML.Labels.Condition
PNML.PnmlTypeDefs.ContinuousNet() -> PNML.Labels.Condition
PNML.PnmlTypeDefs.HLPNG() -> PNML.Labels.Condition
PNML.PnmlTypeDefs.HLCoreNet() -> PNML.Labels.Condition
PNML.PnmlTypeDefs.PT_HLPNG() -> PNML.Labels.Condition

inscription_type

PNML.PnmlTypeDefs.PnmlCoreNet() -> PNML.Labels.Inscription
PNML.PnmlTypeDefs.SymmetricNet() -> PNML.Labels.HLInscription
PNML.PnmlTypeDefs.PTNet() -> PNML.Labels.Inscription
PNML.PnmlTypeDefs.ContinuousNet() -> PNML.Labels.Inscription
PNML.PnmlTypeDefs.HLPNG() -> PNML.Labels.HLInscription
PNML.PnmlTypeDefs.HLCoreNet() -> PNML.Labels.HLInscription
PNML.PnmlTypeDefs.PT_HLPNG() -> PNML.Labels.HLInscription

refplace_type

PNML.PnmlTypeDefs.PnmlCoreNet() -> RefPlace
PNML.PnmlTypeDefs.SymmetricNet() -> RefPlace
PNML.PnmlTypeDefs.PTNet() -> RefPlace
PNML.PnmlTypeDefs.ContinuousNet() -> RefPlace
PNML.PnmlTypeDefs.HLPNG() -> RefPlace
PNML.PnmlTypeDefs.HLCoreNet() -> RefPlace
PNML.PnmlTypeDefs.PT_HLPNG() -> RefPlace

reftransition_type

PNML.PnmlTypeDefs.PnmlCoreNet() -> RefTransition
PNML.PnmlTypeDefs.SymmetricNet() -> RefTransition
PNML.PnmlTypeDefs.PTNet() -> RefTransition
PNML.PnmlTypeDefs.ContinuousNet() -> RefTransition
PNML.PnmlTypeDefs.HLPNG() -> RefTransition
PNML.PnmlTypeDefs.HLCoreNet() -> RefTransition
PNML.PnmlTypeDefs.PT_HLPNG() -> RefTransition

Value Types

TBD