PNML Reference

PnmlTypeDefs Reference

PnmlIDRegistrys Reference

PNML Reference

PNML.PNML — Module

PNML

Petri Net Markup Language, is an XML-based format. PNML.jl reads a pnml model and emits an intermediate representation (IR).

The intermediate representation (IR) represents the XML tree via julia data structures: dictionaries, NamedTuples, LabelledArrays, strings, numbers, objects, vectors. The exact mixture changes as the project continues.

The tags of the XML are used as keys and names as much as possible.

What is accepted as values is ~~often~~ usually a superset of what a given pntd schema specifies. This can be thought of as duck-typing. Conforming to the pntd is not the role of the IR.

The pnml specification has layers. The core layer is useful and extendable. The standard defines extensions of the core for place-transition petri nets (integers) and high-level petri net graphs (many-sorted algebra). This package family adds non-standard continuous net (float64) support. Note that there is no RelaxNG schema file for these extensions

On top of the IR is (will be) implemented Petri Net adaptions and interpertations. This is the level that pntd conformance can be imposed. Adaption to julia packages for graphs, agents, and composing into the greater hive-mind.

Features that have not been started:

Write pnml file

Update pnml model

Create pnml model

Graphs.jl

Symbolics support for HLPNG (many-sorted algebra)

Features that are not complete:

HLPNG - many-sorted algebras are complex. Will build other infrastructure...

pntd specialize

toolspecific usage example

Features that work (perhaps in need of attention as changes are made):

continuous petri net (examples/lotka-volterra.jl)

pnml core: can read & print all Model Checking Contest (MCC) models

Imports

AutoHashEquals

Base

Core

DecFP

DocStringExtensions

FunctionWrappers

LabelledArrays

NamedTupleTools

PNML.PnmlIDRegistrys

PNML.PnmlTypeDefs

PrecompileTools

Preferences

PrettyPrinting

Reexport

Exports

AbstractContinuousNet

AbstractHLCore

AbstractPnmlCore

ContinuousNet

HLCoreNet

HLPNG

MalformedException

MissingIDException

PTNet

PT_HLPNG

PnmlCoreNet

PnmlException

PnmlIDRegistry

PnmlIDRegistrys

PnmlType

PnmlTypeDefs

SymmetricNet

all_nettypes

iscontinuous

isdiscrete

ishighlevel

isregistered

parse_file

parse_pnml

parse_str

pnmltype

pntd_symbol

register_id!

registry

@xml_str

xmlroot

PNML.CONFIG — Constant

See PnmlConfig for default values.

PNML.Maybe — Type

Alias for Union{T, Nothing}.

PNML.pnml_ns — Constant

Namespace expected for pnml XML.

PNML.sort_ids — Constant

Tags used in sort XML elements.

PNML.AbstractDeclaration — Type

abstract type AbstractDeclaration

Declarations define objects/names that are used for high-level terms in conditions, inscriptions, markings. The definitions are attached to PNML nets and/or pages using a PNML Label defined in a <declarations> tag.

PNML.AbstractLabel — Type

abstract type AbstractLabel

Labels are attached to the Petri Net Graph objects. See AbstractPnmlObject.

PNML.AbstractOperator — Type

abstract type AbstractOperator <: PNML.AbstractTerm

Part of the high-level pnml many-sorted algebra.

"...can be a built-in constant or a built-in operator, a multiset operator which among others can construct a multiset from an enumeration of its elements, or a tuple operator. Each operator has a sequence of sorts as its input sorts, and exactly one output sort, which defines its signature."

See NamedOperator and ArbitraryOperator.

PNML.AbstractPetriNet — Type

abstract type AbstractPetriNet{PNTD<:PnmlType}

Top-level of a single network in a pnml model that is some flavor of Petri Net. Note that pnml can represent nets that are not Petri Nets.

Here is where specialization and restriction are applied to achive Proper Petri Behavior.

See PnmlModel, PnmlType.

Extended

Additional constrants can be imposed. We want to run under the motto: "syntax is not semantics, quack".

Since a PNML.Document model can contain multiple networks it is possible that a higher-level will create multiple AbstractPetriNet instances, each a different type.

Multiple Pages can (are permitted to) be merged into one page by flatten_pages! without losing any Petri Net semantics.

PNML.AbstractPnmlNode — Type

abstract type AbstractPnmlNode{PNTD} <: PNML.AbstractPnmlObject{PNTD}

Petri Net Graph nodes are Place, Transition. They are the source or target of an Arc

PNML.AbstractPnmlObject — Type

abstract type AbstractPnmlObject{PNTD<:PnmlType}

Objects of a Petri Net Graph are pages, arcs, nodes.

PNML.AbstractPnmlTool — Type

abstract type AbstractPnmlTool

Tool specific objects can be attached to AbstractPnmlObjects and AbstractLabels subtypes.

PNML.AbstractSort — Type

abstract type AbstractSort

Part of the high-level pnml many-sorted algebra. See SortType.

NamedSort is an AbstractTerm that declares a definition using an AbstractSort. The pnml specification sometimes uses overlapping language.

From the 'primer': built-in sorts of Symmetric Nets are the following: booleans, integerrange, finite enumerations, cyclic enumerations, permutations and dots.

The eltype is expected to be a concrete subtype of Number such as Int, Bool or Float64.

Extras

Notes:

NamedSort is a SortDeclaration. HLPNG adds ArbitrarySort.

UserSort holds the id symbol of a NamedSort.

Here 'type' means a 'term' from the many-sorted algebra.

We use sorts even for non-high-level nets.

Expect eltype(::AbstractSort) to return a concrete subtype of Number.

PNML.AbstractTerm — Type

abstract type AbstractTerm

Terms are part of the multi-sorted algebra that is part of a High-Level Petri Net.

An abstract type in the pnml XML specification, concrete Terms are found within the <structure> element of a label.

Notably, a Term is not a PnmlLabel (or a PNML Label).

See also Declaration, SortType, AbstractDeclaration.

PNML.Annotation — Type

abstract type Annotation <: PNML.AbstractLabel

Label that may be displayed. Differs from an Attribute Label by possibly having a Graphics field.

PNML.AnyElement — Type

struct AnyElement

tag::Symbol

elements::Vector{PNML.AnyXmlNode}

Hold well-formed XML in a Vector{AnyXmlNode}. See also ToolInfo and PnmlLabel.

PNML.AnyXmlNode — Type

Node in a tree formed from XML. tags are XML tags or attribute names. Leaf val are strings. NB: Assumes XML "content" nodes do not have child XML nodes.

PNML.ArbitraryOperator — Type

struct ArbitraryOperator{I<:PNML.AbstractSort} <: PNML.AbstractOperator

declaration::Symbol

input::Vector{PNML.AbstractSort}

output::PNML.AbstractSort

"...arbitrary sorts and operators do not come with a definition of the sort or operation; they just introduce a new symbol without giving a definition for it. "

See [`UserOperator`](@ref)

PNML.ArbitrarySort — Type

struct ArbitrarySort <: PNML.SortDeclaration

id::Symbol

name::Union{String, SubString}

Arbitrary sorts that can be used for constructing terms are reserved for/supported by HLPNG in the pnml specification.

PNML.Arc — Type

Edge of a Petri Net Markup Language graph that connects place and transition.

mutable struct Arc{PNTD, I} <: PNML.AbstractPnmlObject{PNTD}

pntd::Any

id::Symbol

source::Symbol

target::Symbol

inscription::Any

name::Union{Nothing, PNML.Name}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

labels::Vector{PNML.PnmlLabel}

PNML.BoolSort — Type

Built-in sort whose eltype is Bool

Operators: and, or, not, imply Functions: equality, inequality

PNML.Condition — Type

struct Condition <: PNML.Annotation

text::Union{Nothing, String}

value::PNML.Term

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

Label of a Transition that determines when the transition fires.

Examples

julia> c = Condition(false)
Condition(nothing, Term(:bool, false), nothing, [])

julia> c()
false

julia> c = Condition("xx", false)
Condition("xx", Term(:bool, false), nothing, [])

julia> c()
false

PNML.Coordinate — Type

struct Coordinate{T<:DecFP.DecimalFloatingPoint}

x::DecFP.DecimalFloatingPoint

y::DecFP.DecimalFloatingPoint

Cartesian Coordinate are actually positive decimals. Ranges from 0 to 999.9.

PNML.Coordinate — Method

Construct a Coordinate from mixed Int, Float64.

PNML.CyclicEnumerationSort — Type

struct CyclicEnumerationSort <: PNML.AbstractSort

PNML.Declaration — Type

struct Declaration <: PNML.Annotation

declarations::Vector{Any}

text::Union{Nothing, String}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

Label of a <net> or <page> that holds zero or more declarations. The declarations are used to define parts of the many-sorted algebra used by High-Level Petri Nets. Other PNTDs may introduce non-standard uses for declarations.

Notes: declarations is implemented as a collection of Any.

We can use infrastructure implemented for HL nets to provide nonstandard extensions.

PNML.DotSort — Type

Built-in sort whose eltype is Int

PNML.Fill — Type

Fill attributes as strings.

struct Fill

color::String

image::String

gradient_color::String

gradient_rotation::String

PNML.FiniteEnumerationSort — Type

struct FiniteEnumerationSort <: PNML.AbstractSort

PNML.FiniteIntRangeSort — Type

struct FiniteIntRangeSort{T} <: PNML.AbstractSort

PNML.Font — Type

Font attributes as strings.

struct Font

family::String

style::String

weight::String

size::String

align::String

rotation::String

decoration::String

PNML.Graphics — Type

PNML Graphics can be attached to many parts of PNML models.

struct Graphics{T<:DecFP.Dec32}

dimension::PNML.Coordinate{T} where T<:DecFP.Dec32

fill::PNML.Fill

font::PNML.Font

line::PNML.Line

offset::PNML.Coordinate{T} where T<:DecFP.Dec32

positions::Array{PNML.Coordinate{T}, 1} where T<:DecFP.Dec32

PNML.HLAnnotation — Type

abstract type HLAnnotation <: PNML.AbstractLabel

Annotation label that uses <text> and <structure>.

PNML.HLInscription — Type

struct HLInscription{T<:PNML.Term} <: PNML.HLAnnotation

text::Union{Nothing, String}

term::PNML.Term

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

Labels an Arc. The <structure> element is a term in a many-sorted algebra. The term field TBD. See also Inscription

Examples

julia> i2 = HLInscription(Term(:value, 3))
HLInscription(nothing, Term(:value, 3), nothing, [])

julia> i2()
3

julia> i3 = HLInscription("text", Term(:empty, 1))
HLInscription("text", Term(:empty, 1), nothing, [])

julia> i3()
1

julia> i4 = HLInscription("text", Term(:value, 3))
HLInscription("text", Term(:value, 3), nothing, [])

julia> i4()
3

PNML.HLInscription — Method

Evaluate a HLInscription. Returns a value of the same sort as TBD.

PNML.HLLabel — Type

struct HLLabel{PNTD} <: PNML.Annotation

text::Union{Nothing, String}

structure::Union{Nothing, PNML.AnyElement}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

High-level pnml labels are expected to have <text> and <structure> elements. This concrete type is for "unclaimed" labels in a high-level petri net.

Some "claimed" HLAnnotation labels are Condition, Declaration, HLMarking, HLInscription.

PNML.HLMarking — Type

struct HLMarking{T<:PNML.AbstractTerm} <: PNML.HLAnnotation

text::Union{Nothing, String}

term::PNML.AbstractTerm

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

Multiset of a sort labelling of a Place in a High-level Petri Net Graph. See AbstractHLCore, Term, Marking.

Is a functor that returns the evaluated value.

Examples

julia> m = HLMarking("the text", Term(:value, 3))
HLMarking("the text", Term(:value, 3), nothing, [])

julia> m()
3

PNML.HLMarking — Method

Evaluate a HLMarking instance by returning its term.

PNML.HLPetriNet — Type

Wrap a single pnml net. Presumes that the net does not need to be flattened as all content is in first page.

struct HLPetriNet{PNTD} <: PNML.AbstractPetriNet{PNTD}

net::PNML.PnmlNet{PNTD} where PNTD

Details

PNML.HLPetriNet — Method

Construct from string of valid pnml XML, using the first network in model.

PNML.HarvestAny — Type

Wrap a function and two of its arguments.

PNML.Inscription — Type

struct Inscription{T<:Union{Float64, Int64}} <: PNML.Annotation

value::Union{Float64, Int64}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

Labels an Arc. See also HLInscription.

Examples

julia> i = Inscription(3)
Inscription(3, nothing, [])

julia> i()
3

PNML.Inscription — Method

Evaluate an Inscription's value.

PNML.IntegerSort — Type

Built-in sort whose eltype is Int

PNML.Line — Type

Line attributes as strings.

struct Line

color::String

shape::String

style::String

width::String

PNML.ListSort — Type

struct ListSort <: PNML.AbstractSort

PNML.MalformedException — Type

struct MalformedException <: PnmlException

msg::String

PNML.Marking — Type

struct Marking{N<:Union{Float64, Int64}} <: PNML.Annotation

value::Union{Float64, Int64}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

Number-valued label of Place. See PTNet, ContinuousNet, HLMarking.

Is a functor that returns the value.

Examples

julia> m = Marking(1)
Marking(1, nothing, [])

julia> m()
1

julia> m = Marking(12.34)
Marking(12.34, nothing, [])

julia> m()
12.34

PNML.Marking — Method

Evaluate Marking instance by returning its evaluated value.

PNML.MissingIDException — Type

struct MissingIDException <: PnmlException

msg::String

Use exception to allow dispatch and additional data presentation to user.

PNML.MultisetSort — Type

struct MultisetSort <: PNML.AbstractSort

Wrap a UserSort. Warning: do not cause recursive multiset Sorts.

PNML.Name — Type

struct Name <: PNML.Annotation

text::String

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

Name is for display, possibly in a tool specific way.

PNML.NamedOperator — Type

struct NamedOperator{V, T} <: PNML.OperatorDeclaration

id::Symbol

name::Union{String, SubString}

parameter::Vector

def::Any

See UserOperator

PNML.NamedSort — Type

struct NamedSort{S<:Union{PNML.AbstractSort, PNML.AnyElement}} <: PNML.SortDeclaration

id::Symbol

name::Union{String, SubString}

def::Union{PNML.AbstractSort, PNML.AnyElement}

PNML.NaturalSort — Type

Built-in sort whose eltype is Int

PNML.OperatorDeclaration — Type

abstract type OperatorDeclaration <: PNML.AbstractDeclaration

PNML.Page — Type

struct Page{PNTD<:PnmlType, P, T, A, RP, RT} <: PNML.AbstractPnmlObject{PNTD<:PnmlType}

pntd::PnmlType

id::Symbol

declaration::PNML.Declaration

name::Union{Nothing, PNML.Name}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

labels::Vector{PNML.PnmlLabel}

pagedict::OrderedCollections.OrderedDict{Symbol, PNML.Page{PNTD, P, T, A, RP, RT}} where {PNTD<:PnmlType, P, T, A, RP, RT}

netdata::PNML.PnmlNetData

netsets::PNML.PnmlNetKeys

Contain all places, transitions & arcs. Pages are for visual presentation. There must be at least 1 Page for a valid pnml model.

PNTD binds the other type parameters together to express a specific PNG. See PnmlNet

PNML.PartitionElement — Type

struct PartitionElement

id::Symbol

name::Union{String, SubString}

terms::Vector{PNML.UserOperator}

Partition Element is part of a Partition Sort.

PNML.PartitionSort — Type

struct PartitionSort{S, PE} <: PNML.SortDeclaration

id::Symbol

name::Union{String, SubString}

def::Any

element::Any

Partition sort.

PNML.Place — Type

struct Place{PNTD, M, S<:PNML.SortType} <: PNML.AbstractPnmlNode{PNTD}

pntd::Any

id::Symbol

initialMarking::Any

sorttype::PNML.SortType

name::Union{Nothing, PNML.Name}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

labels::Vector{PNML.PnmlLabel}

Place node of a Petri Net Markup Language graph.

PNML.PnmlConfig — Type

Configuration with default values that can be overidden by a LocalPreferences.toml.

Options

indent_width::Int: Indention of nested lines.

warn_on_namespace::Bool: There are pnml files that break the rules & do not have an xml namespace.

text_element_optional::Bool: There are pnml files that break the rules & do not have <text> elements.

warn_on_fixup::Bool: When an missing value is replaced by a default value, issue a warning.

warn_on_unclaimed::Bool: Issue warning when PNML label does not have a parser defined. While allowed, there will be code required to do anything useful with the label.

verbose::Bool: Print information as runs.

lock_registry::Bool: Lock registry with a ReentrantLock.

PNML.PnmlException — Type

abstract type PnmlException <: Exception

PNML.PnmlLabel — Type

struct PnmlLabel <: PNML.Annotation

tag::Symbol

elements::Vector{PNML.AnyXmlNode}

Wrap a AnyXmlNode[] holding a PNML Label. Use the XML tag as identifier.

Used for "unclaimed" labels that do not have, or we choose not to use, a dedicated parse method. Claimed labels will have a type/parser defined to make use of the structure defined by the pntd schema.

See also AnyElement. The difference is that AnyElement allows any well-formed XML, while PnmlLabel is restricted to PNML Labels (with extensions in PNML.jl).

PNML.PnmlModel — Type

struct PnmlModel

nets::Tuple{Vararg{PNML.PnmlNet}}

namespace::String

reg::PnmlIDRegistry

One or more Petri Nets and an ID Registry shared by all nets.

PNML.PnmlNet — Type

struct PnmlNet{PNTD<:PnmlType, P, T, A, RP, RT}

type::PnmlType

id::Symbol

pagedict::OrderedCollections.OrderedDict{Symbol, PNML.Page{PNTD, P, T, A, RP, RT}} where {PNTD<:PnmlType, P, T, A, RP, RT}

netdata::PNML.PnmlNetData

page_set::OrderedCollections.OrderedSet{Symbol}

declaration::PNML.Declaration

name::Union{Nothing, PNML.Name}

tools::Vector{PNML.ToolInfo}

labels::Vector{PNML.PnmlLabel}

One Petri Net of a PNML model.

PNML.PnmlNetData — Type

struct PnmlNetData{PNTD<:PnmlType, P, T, A, RP, RT}

pntd::PnmlType

place_dict::OrderedCollections.OrderedDict{Symbol}

transition_dict::OrderedCollections.OrderedDict{Symbol}

arc_dict::OrderedCollections.OrderedDict{Symbol}

refplace_dict::OrderedCollections.OrderedDict{Symbol}

reftransition_dict::OrderedCollections.OrderedDict{Symbol}

Collect each of the PnmlNodess & Arcs of a Petri Net Graph into one collection. Accessed via pnml ID key or iterate over values of an OrderedDict.

In the 'pnml' specification there is a Page structure that can be removed by flatten_pages!, removing some display-related information, leaving a functional Petri Net Graph as described in this structure. It is intended to be a per-PnmlNet database that is mutated as each page is parsed.

See PnmlNetKeys for page-level pnml ID of "owners" net data.

PNML.PnmlNetKeys — Type

struct PnmlNetKeys

page_set::OrderedCollections.OrderedSet{Symbol}

place_set::OrderedCollections.OrderedSet{Symbol}

transition_set::OrderedCollections.OrderedSet{Symbol}

arc_set::OrderedCollections.OrderedSet{Symbol}

reftransition_set::OrderedCollections.OrderedSet{Symbol}

refplace_set::OrderedCollections.OrderedSet{Symbol}

Per-page structure of OrderedSets of pnml IDs for each "owned" Page and AbstractPnmlObject.

PNML.PositiveSort — Type

Built-in sort whose eltype is Int

PNML.ProductSort — Type

struct ProductSort <: PNML.AbstractSort

An ordered collection of sorts.

PNML.RealSort — Type

"Built-in sort whose eltype is Float64"

Real numbers are not part of the PNML Specification. We stick them into the type hierarchy for convenience rather than mathematical correctness.

PNML.RefPlace — Type

Reference Place node of a Petri Net Markup Language graph. For connections between pages.

struct RefPlace{PNTD} <: PNML.ReferenceNode{PNTD}

pntd::Any

id::Symbol

ref::Symbol

name::Union{Nothing, PNML.Name}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

labels::Vector{PNML.PnmlLabel}

PNML.RefTransition — Type

Refrence Transition node of a Petri Net Markup Language graph. For connections between pages.

struct RefTransition{PNTD} <: PNML.ReferenceNode{PNTD}

pntd::Any

id::Symbol

ref::Symbol

name::Union{Nothing, PNML.Name}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

labels::Vector{PNML.PnmlLabel}

PNML.ReferenceNode — Type

abstract type ReferenceNode{PNTD} <: PNML.AbstractPnmlNode{PNTD}

For common behavior shared by RefPlace, RefTransition used to connect Page together.

PNML.SimpleNet — Type

struct SimpleNet{PNTD} <: PNML.AbstractPetriNet{PNTD}

id::Symbol

net::PNML.PnmlNet{PNTD} where PNTD

TODO: Rename SimpleNet to TBD

SimpleNet is a concrete AbstractPetriNet wrapping a PnmlNet.

Uses a flattened net to avoid the page level of the pnml hierarchy.

Note: A multi-page petri net can always be flattened by removing referenceTransitions & referencePlaces, and merging pages into the first page.

PNML.SortDeclaration — Type

abstract type SortDeclaration <: PNML.AbstractDeclaration

See NamedSort and [ArbitrarySort] as concrete subtypes.

PNML.SortType — Type

struct SortType <: PNML.Annotation

text::Union{Nothing, String}

sort::Base.RefValue{PNML.AbstractSort}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

A places's <type> label's <structure> element holds a concrete subtype of AbstractSort. Defines the sort of a place, hence use of sorttype.

For high-level nets there will be a rich language of sorts using UserSort and NamedSort. For other PnmlNets they may still be used internally

PNML.StringSort — Type

struct StringSort <: PNML.AbstractSort

PNML.Structure — Type

struct Structure

tag::Symbol

el::Vector{PNML.AnyXmlNode}

High-level Annotation Labels place meaning in <structure> that is consumed by "claimed" labels. Is is expected to contain an abstract syntax tree (ast) for the many-sorted algebra expressed in XML. We implement this to allow use of <structure> tags by other PnmlTypes.

Extra

There are various defined structure ast variants in pnml:

Sort Type of a Place [builtin, multi, product, user]

Place HLMarking [variable, operator]

Transition Condition [variable, operator]

Arc Inscription [variable, operator]

Declarations [sort, variable, operator]

These should all have dedicated parsers and objects as claimed labels. Here we provide a fallback for unclaimed tags.

PNML.Term — Type

struct Term <: PNML.AbstractTerm

tag::Symbol

elements::Union{Bool, Float64, Int64, Vector{PNML.AnyXmlNode}}

Note that Term is an abstract element in the pnml specification with no XML tag, we call that AbstractTerm. Here we useTerm` as a concrete wrapper around unparsed high-level many-sorted algebra terms AND EXTEND to also wrapping "single-sorted" values.

By adding Bool, Int, Float64 it is possible for PnmlCoreNet and ContinuousNet to use Terms, and for implenting default_bool_term, default_one_term, default_zero_term.

See also iscontinuous

As part of the many-sorted algebra attached to nodes of a High-level Petri Net Graph, Terms are contained within the <structure> element of an annotation label, See [HLAnnotation`](@ref) concrete subtypes.

#TODO is it safe to assume that Bool, Int, Float64 are in the carrier set/basis set?

#! HL term is currently implemented as a wrapper of Vector{AnyXmlNode}. A tree of symbols with leafs that are strings.

Functor

(t::Term)([default_one_term(HLCoreNet())])

Term as functor requires a default value for missing values.

As a preliminary implementation evaluate a HL term by returning :value in elements or evaluating default_one_term(HLCoreNet()).

See parse_marking_term, parse_condition_term, parse_inscription_term, parse_type, parse_sorttype_term, AnyElement.

Warning: Much of the high-level is WORK-IN-PROGRESS. The type parameter is a sort. We enumerate some of the built-in sorts allowed. Is expected that the term will evaluate to that type. Is that called a 'ground term'? 'basis set'? When the elements' value is a Vector{AnyXmlNode} external information is used to select the output type.

PNML.TokenGraphics — Type

struct TokenGraphics{T<:DecFP.DecimalFloatingPoint} <: PNML.AbstractPnmlTool

positions::Array{PNML.Coordinate{T}, 1} where T<:DecFP.DecimalFloatingPoint

TokenGraphics is <toolspecific> content. Combines the <tokengraphics> and <tokenposition> elements.

PNML.ToolInfo — Type

struct ToolInfo{T}

toolname::String

version::String

infos::Any

ToolInfo holds a <toolspecific> tag.

It wraps a iteratable collection (currently vector) of well formed elements parsed into AnyElements for use by anything that understands toolname, version tool specifics.

PNML.Transition — Type

Transition node of a Petri Net Markup Language graph.

struct Transition{PNTD, C} <: PNML.AbstractPnmlNode{PNTD}

pntd::Any

id::Symbol

condition::Any

name::Union{Nothing, PNML.Name}

graphics::Union{Nothing, PNML.Graphics}

tools::Vector{PNML.ToolInfo}

labels::Vector{PNML.PnmlLabel}

PNML.TransitionRate — Type

PNML.UnknownDeclaration — Type

struct UnknownDeclaration <: PNML.AbstractDeclaration

id::Symbol

name::Union{String, SubString}

nodename::Union{String, SubString}

content::Vector{Any}

PNML.UserOperator — Type

struct UserOperator <: PNML.AbstractOperator

declaration::Symbol

User defined operators only define an abbreviation. See NamedOperator

PNML.UserSort — Type

struct UserSort <: PNML.AbstractSort

Holds a reference id to a concrete subtype of SortDeclaration.

NamedSort is used to construct a sort out of builtin types. Used in a Places sort type property.

PNML.Variable — Type

struct Variable <: PNML.AbstractTerm

variableDecl::Symbol

Variable refers to a varaible declaration. Example input: <variable refvariable="varx"/>.

#TODO examples of use, modifying and accessing

PNML.VariableDeclaration — Type

struct VariableDeclaration{S} <: PNML.AbstractDeclaration

id::Symbol

name::Union{String, SubString}

sort::Any

PNML.XMLNode — Type

Alias for EzXML.Node

PNML._evaluate — Function

_evaluate(x::Number) -> identity(x)
_evaluate(x::Base.Callable) -> (x)()

Return the value of "x", defaults to identity.

Examples

Since High-level PNML schemas are based on Natural numbers and booleans, it seems reasonable to assume Number, which includes Bool, for the non-callable type. A functor is expected as the callable type allowing expressions in the many-sorted algebra to be evaluated to a Number.

PNML._harvest_any — Method

_harvest_any(
    node::EzXML.Node,
    harvest!::PNML.HarvestAny
) -> Vector{PNML.AnyXmlNode}

Return AnyXmlNode vector holding a well-formed XML node.

If element node has any attributes &/or children, use attribute name or child's name as tag. (Empty is well-formed.)

Descend the well-formed XML using parser function harvest! on child nodes.

Note the assumption that "children" and "content" are mutually exclusive. Content is always a leaf element. However XML attributes can be anywhere in the hierarchy. And neither children nor content nor attribute may be present.

Leaf AnyXmlNode's contain an String or SubString.

PNML._match — Function

_match(ti::ToolInfo, name::AbstractString)
_match(ti::ToolInfo, name::String, version::String)
_match(ti::ToolInfo, namerex::Regex, versionrex::Regex)

Match toolname and version. Default is any version.

PNML._parse_decl_structure — Method

Assumes high-level semantics until someone specializes. See decl_structure.

PNML._parse_page! — Method

Call parse_page!, add page to dictionary and id set

PNML.add_toolinfo! — Method

add_toolinfo!(infos, node, pntd, reg)

Add ToolInfo to infos, return nothing.

The UML from the pnml primer (and schemas) use <toolspecific> as the tag name for instances of the type ToolInfo.

PNML.all_arcs — Function

Return arcs that have a source or target of transition id.

See also src_arcs, tgt_arcs.

PNML.allchildren — Function

allchildren(
    tag::AbstractString,
    el::EzXML.Node
) -> Vector{EzXML.Node}
allchildren(
    tag::AbstractString,
    el::EzXML.Node,
    ns::AbstractString
) -> Vector{EzXML.Node}

Return vector of el's immediate children with tag.

PNML.allpages — Method

PNML.anyelement — Method

anyelement(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.AnyElement

Return AnyElement holding a well-formed XML node. See ToolInfo for one intended use-case.

PNML.append_page! — Method

Append selected fields of r to fields of l. Some, like Names and xml, are omitted because they are scalar values, not collections.

pagedict & netdata (holding the arc and pnml nodes) are per-net data that is not modified here. netsets hold pnml IDs "owned"

PNML.arc — Function

Return arc with id if found, otherwise nothing.

PNML.arc — Method

Return Arc from 's' to 't' or nothing`. Assumes there is at most one.

PNML.arc_idset — Function

Return iterator over arc ids.

PNML.arc_type — Function

Type of Arc.

PNML.arcs — Function

arcs(n::PnmlNet) -> iterator
arcs(p::AbstractPetriNet) -> iterator

Return iterator over arc ids.

PNML.condition — Function

Return condition's value of transition.

PNML.condition — Method

Return value of condition.

PNML.condition_type — Function

Type of Condition.

PNML.condition_value_type — Function

Return value type.

PNML.conditions — Function

Return a labelled vector of condition values.

PNML.conditions — Method

conditions(petrinet) -> LVector{condition_value_type(pntd)}

LVector labelled with transition id and holding its condition's value.

PNML.coordinate_type — Function

Return Coordinate{coordinate_value_type(T)}

PNML.coordinate_value_type — Function

Return type of value in a Coordinate.

PNML.decl_structure — Method

Return vector of AbstractDeclaration subtypes.

PNML.default_bool_term — Function

True as boolean or term with a value of true.

PNML.default_condition — Function

Return default condition based on PNTD. Has meaning of true or always. ```

PNML.default_inscription — Function

Return default inscription value based on PNTD. Has meaning of unity, as in one.

PNML.default_marking — Function

Return default marking value based on PNTD. Has meaning of empty, as in zero.

PNML.default_one_term — Function

One as integer, float, or empty term with a value of one.

PNML.default_sort — Function

Return instance of default sort based on PNTD.

PNML.default_sorttype — Function

Return instance of default place sort type based on PNTD.

PNML.default_zero_term — Function

Zero as integer, float, or empty term with a value of zero.

PNML.deref! — Function

deref!(net::PNML.PnmlNet) -> PNML.PnmlNet
deref!(net::PNML.PnmlNet, trim::Bool) -> PNML.PnmlNet

Remove reference nodes from arcs.

Operates on the PnmlNetData at the net level. Expects that the PnmlNetKeys of the firstpage will have to be cleaned as part of flatten_pages!,

Axioms

All ids in a network are unique in that they only have one instance in the XML.

A chain of reference Places or Transitions always ends at a Place or Transition.

All ids are valid.

No cycles.

PNML.deref_place — Function

deref_place(net, id[, trim::Bool] ) -> Symbol

Return id of referenced place. If trim is true (default) the reference is removed.

PNML.deref_transition — Function

deref_transition(net, id[, trim::Bool] ) -> Symbol

Return id of referenced transition. If trim is true (default) the reference is removed.

PNML.enabled — Method

enabled(::AbstractPetriNet, ::LVector) -> LVector

Returns labelled vector of id=>boolean where true means transitionid is enabled at marking.

PNML.equals — Method

equals(a::PNML.AbstractSort, b::PNML.AbstractSort) -> Any

Are the sorts eltype the same? First the must have the same type. Then any contents of the sorts are compared.

PNML.find_net — Function

Return PnmlNet having id or nothing`.

PNML.find_nets — Function

Return nets matching pntd type given as symbol or pnmltype singleton.

PNML.fire! — Method

Return the marking after firing transition. M1 = M0 + Cf. M0 is the initial marking vector, f is the firing vector, i.e. which transition is to fire and C is the incidence matrix.

PNML.first_net — Method

first_net(model) -> PNML.PnmlNet

Return first net contained by doc.

PNML.firstchild — Function

firstchild(
    tag::AbstractString,
    node::EzXML.Node
) -> Union{Nothing, EzXML.Node}
firstchild(
    tag::AbstractString,
    node::EzXML.Node,
    ns::AbstractString
) -> Union{Nothing, EzXML.Node}

Return up to 1 immediate child of el that is a tag.

PNML.firstpage — Method

Usually the only interesting page.

PNML.flatten_pages! — Function

flatten_pages!(net::PnmlNet[; options])
flatten_pages!(model::PnmlModel[; options])

Merge page content into the 1st page of the net or all nets of a model.

Options trim::Bool Remove refrence nodes (default false). See deref!. verbose::Bool Print breadcrumbs (default false).

PNML.get_label — Function

get_label(x, tag::Symbol) -> PnmlLabel

Return first label of x with a matching tagvalue.

PNML.get_toolinfo — Function

Return first toolinfo having a matching toolname and version.

PNML.getfirst — Method

Return first true f of v or nothing.

PNML.has_arc — Function

Return true if any arc has id.

PNML.has_label — Function

has_label(x, tag::Symbol) -> Bool

Does x have any label with a matching tag.

PNML.has_labels — Function

has_labels(x) -> Bool

Does x have any labels.

PNML.has_name — Function

has_name(x) -> Bool

Return true if there is a name.

PNML.has_place — Function

Return true if there is any place with id?

PNML.has_refplace — Function

PNML.has_reftransition — Function

PNML.has_toolinfo — Function

has_toolinfo(infos, toolname[, version]) -> Bool

Does any toolinfo in iteratable infos have a matching toolname, and a matching version (if it is provided). toolname and version will be turned into Regexs to match against each ToolInfo in the infos collection.

PNML.has_transition — Function

Is there a transition with id?

PNML.hastag — Method

Use with Fix2 to filter anything with tag accessor.

PNML.in_inscriptions — Method

Iterate over preset of transition, returning source place id => inscription value pairs.

PNML.in_out — Function

Return tuple of input, output labelled vectors with key of place ids and value of arc inscription's value for use as a transition function. #TODO When do these get called "pre" and "post"?

PNML.inc_indent — Method

Increment the :indent value by indent_width.

PNML.incidence_matrix — Method

incidence_matrix(petrinet) -> LArray

C[transition,place] = inscription(transition,place) - inscription(place,transition)

PNML.indent — Method

Return string of current indent size in io.

PNML.infos — Method

Content of a ToolInfo.

PNML.initial_marking — Function

Return the initial marking of a place.

PNML.initial_markings — Method

initial_markings(petrinet) -> LVector{marking_value_type(pntd)}

LVector labelled with place id and holding initial marking's value.

PNML.input_matrix — Method

Input matrix

PNML.ins — Method

ins(net, transition_id) -> LVector

Return vector of inscription values labeled with source place id for arcs with `transition_id` as the target id.

PNML.inscription — Function

Return incription value of arc.

PNML.inscription_type — Function

inscription_type(pntd) -> Inscription{inscription_value_type(pntd)}.

PNML.inscription_value_type — Function

Return value type.

PNML.ispid — Method

ispid(x::Symbol) -> Base.Fix2{typeof(===), Symbol}

Return function to be used like: any(ispid(sym), iteratewithpid)

PNML.labels — Function

labels(x) -> Iterateable

Return iterator of labels attached to x.

PNML.marking_type — Function

marking_type(pntd) -> Marking{marking_value_type(pntd)}.

PNML.marking_value_type — Function

Return value type.

PNML.metagraph — Method

metagraph(::AbstractPetriNet) -> MetaGraph

Return MetaGraph instance based on a SimpleDiGraph with PNML nodes attached to vertices and edges.

PNML.name — Function

name(x) -> String

Return name String. Default to empty string.

PNML.name — Method

Name of tool to for this tool specific information element.

PNML.nets — Method

nets(model::PNML.PnmlModel) -> Tuple{Vararg{PNML.PnmlNet}}

Return all nets of model.

PNML.nettype — Function

Return the PnmlType subtype representing the flavor (or pntd) of this kind of Petri Net Graph.

See also pnmltype

PNML.number_value — Method

number_value(::Type{T}, s) -> T

Parse string as a type T <: Number.

PNML.out_inscriptions — Method

Iterate over postset of transition, returning target place id => inscription value pairs.

PNML.output_matrix — Method

Output Matrix

PNML.outs — Method

outs(net, transition_id) -> LVector

Return vector of inscription values labeled with target place id for arcs with transition_id as the source id.

PNML.page_type — Function

Type of Page.

PNML.pages — Function

pages(net::PnmlLabel|page::Page) -> iterator

Return iterator of pages directly owned by that object. See allpages for an iterator over all pages in the PNML network model. Maintains order (insertion order).

PNML.pages — Method

Iterator of Pages directly owned by net.

PNML.parse_and — Method

parse_and(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

PNML.parse_arc — Method

parse_arc(node::XMLNode, pntd::PnmlType, idregistry) -> Arc{typeof(pntd), typeof(inscription)}

Construct an Arc with labels specialized for the PnmlType.

PNML.parse_booleanconstant — Method

parse_booleanconstant(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

PNML.parse_condition — Method

parse_condition(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Condition

Label of transition nodes.

Details

Condition is defined by the ISO Specification as a High-level Annotation, meaning it has <text> and <structure> elements. With all meaning in the element that the <structure> holds evaluating to a boolean value. We extend this to anything that evaluates to a boolean value when treated as a functor.

A Condition should evaluate to a boolean. See AbstractTerm.

PNML.parse_condition_term — Method

parse_condition_term(
    cnode::EzXML.Node,
    pntd::PnmlType,
    idregistry
) -> PNML.Term

PNML.parse_decl — Method

The body has only a :declaration, return its value as a string.

PNML.parse_declaration — Method

parse_declaration(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Declaration

Return Declaration label of 'net' and 'page' nodes. Assume behavior of a High-level Net label in that the meaning is in a <structure>.

Expected format: <declaration> <structure> <declarations> <namedsort/> <namedsort/> ...

PNML.parse_equality — Method

parse_equality(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

PNML.parse_file — Method

parse_file(fname::AbstractString) -> PNML.PnmlModel

Build a PnmlModel from a file containing XML. See parse_str and parse_pnml.

PNML.parse_graphics — Method

parse_graphics(
    node,
    pntd,
    reg
) -> PNML.Graphics{DecFP.Dec32}

Arcs, Annotations and Nodes have different graphics semantics. Return a Graphics holding the union of possibilities.

PNML.parse_graphics_coordinate — Method

parse_graphics_coordinate(
    node,
    pntd,
    reg
) -> PNML.Coordinate{DecFP.Dec32}

Return Coordinate. Specification seems to only use integers, we also allow real numbers.

PNML.parse_graphics_fill — Method

parse_graphics_fill(node, pntd, reg) -> PNML.Fill

Return Fill

PNML.parse_graphics_font — Method

parse_graphics_font(node, pntd, reg) -> PNML.Font

Return Font.

PNML.parse_graphics_line — Method

parse_graphics_line(node, pntd, reg) -> PNML.Line

Return Line.

PNML.parse_hlinitialMarking — Method

parse_hlinitialMarking(
    node::EzXML.Node,
    pntd::AbstractHLCore,
    idregistry::PnmlIDRegistry
) -> PNML.HLMarking{PNML.Term}

High-level initial marking labels are expected to have a Term in the <structure> child. We extend the pnml standard by allowing node content to be numeric: parsed to Int and Float64.

PNML.parse_hlinscription — Method

parse_hlinscription(
    node::EzXML.Node,
    pntd::AbstractHLCore,
    idregistry::PnmlIDRegistry
) -> PNML.HLInscription{PNML.Term}

hlinscriptions are expressions.

PNML.parse_imply — Method

parse_imply(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

PNML.parse_inequality — Method

parse_inequality(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

PNML.parse_initialMarking — Method

parse_initialMarking(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> Union{PNML.Marking{Float64}, PNML.Marking{Int64}}

PNML.parse_inscription — Method

parse_inscription(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Inscription

PNML.parse_inscription_term — Method

parse_inscription_term(inode, pntd, idregistry) -> PNML.Term

PNML.parse_label — Method

parse_label(
    node::EzXML.Node,
    _::PnmlType,
    _::PnmlIDRegistry
) -> NamedTuple{(:tag, :xml), Tuple{Symbol, EzXML.Node}}

Should not often have a '<label>' tag, this will bark if one is found. Return NamedTuple (tag,node), to defer parsing the xml.

PNML.parse_marking_term — Method

parse_marking_term(marknode, pntd, idregistry) -> PNML.Term

PNML.parse_name — Method

parse_name(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Name

Return Name label holding text value and optional tool & GUI information.

PNML.parse_namedoperator — Method

parse_namedoperator(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.NamedOperator{PNML.VariableDeclaration, PNML.Term}

PNML.parse_namedsort — Method

parse_namedsort(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.NamedSort

PNML.parse_net — Function

parse_net(
    node::EzXML.Node,
    idregistry::PnmlIDRegistry
) -> PNML.PnmlNet
parse_net(
    node::EzXML.Node,
    idregistry::PnmlIDRegistry,
    pntd_override::Union{Nothing, PnmlType}
) -> PNML.PnmlNet

Return a PnmlNet`.

PNML.parse_net_1 — Method

Parse net with a defined PnmlType. The PNTD is used to set the marking, inscription, condition and sort type parameters.

PNML.parse_not — Method

parse_not(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

PNML.parse_or — Method

parse_or(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

PNML.parse_page! — Method

parse_page!(tup, node, pntd, idregistry) -> Page

Place Page in pagedict using id as the key.

PNML.parse_place — Method

parse_place(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Place{_A, _B, PNML.SortType} where {_A, _B}

PNML.parse_pnml — Method

parse_pnml(xmlnode, idregistry) -> PnmlModel

Start parse from the root node of a well formed pnml XML document. Return a PnmlModel holding one or more PnmlNet.

PNML.parse_refPlace — Method

parse_refPlace(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.RefPlace

PNML.parse_refTransition — Method

parse_refTransition(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.RefTransition

PNML.parse_sort — Method

parse_sort(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> Any

Sorts are found within a <structure> element.

PNML.parse_sorttype_term — Method

parse_sorttype_term(typenode, pntd, idregistry) -> Any

#TODO where does this belong? A concrete subtype of AbstractSort. Built from many different elements that contain a Sort: type, namedsort, variabledecl, multisetsort, productsort, numberconstant, partition...

Sort = BuiltInSort | MultisetSort | ProductSort | UserSort

PNML.parse_str — Method

parse_str(str::AbstractString) -> PNML.PnmlModel

Build a PnmlModel from a string containing XML. See parse_file and parse_pnml.

PNML.parse_structure — Method

parse_structure(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Structure

Return Structure holding a <structure>. Should be inside of an label. A "claimed" label usually elids the <structure> level (does not call this method).

PNML.parse_subterm — Method

parse_subterm(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

Used to construct the syntax tree of multi-sorted algebra.

PNML.parse_term — Method

parse_term(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

There will be no XML node 'term'. Instead it is the interpertation of the child of some 'structure' or def elements. Terms kinds are Variable and Operator.

PNML.parse_text — Method

parse_text(
    node::EzXML.Node,
    _::PnmlType,
    _::PnmlIDRegistry
) -> String

Return the stripped string of node's content.

PNML.parse_tokengraphics — Method

parse_tokengraphics(
    node::EzXML.Node,
    pntd::PnmlType,
    reg
) -> PNML.TokenGraphics{DecFP.Dec32}

Parse high-level place-transition net's (HL-PTNet) toolspecific structure defined for token graphics. See TokenGraphics and parse_tokenposition.

PNML.parse_tokenposition — Method

parse_tokenposition(
    node,
    pntd,
    reg
) -> PNML.Coordinate{DecFP.Dec32}

Return Cartesian Coordinate relative to containing element.

PNML.parse_toolspecific — Method

parse_toolspecific(
    node,
    pntd,
    reg
) -> Union{PNML.ToolInfo{Vector{PNML.AnyElement}}, PNML.ToolInfo{PNML.TokenGraphics{DecFP.Dec32}}}

Return ToolInfo with tool & version attributes and content.

The content can be one or more well-formed xml elements. Each are wrapped in a PnmlLabel.

PNML.parse_transition — Method

parse_transition(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.Transition{_A, PNML.Condition} where _A

PNML.parse_tuple — Method

parse_tuple(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

PNML.parse_type — Method

parse_type(
    node::EzXML.Node,
    pntd::PnmlType,
    idregistry::PnmlIDRegistry
) -> PNML.SortType

Defines the "sort" of tokens held by the place and semantics of the marking. NB: The "type" of a place from many-sorted algebra is different from the Petri Net "type" of a net or "pntd". Neither is directly a julia type.

Allow all pntd's places to have a <type> label. Non high-level are expecting a numeric sort: eltype(sort) <: Number. See default_sort`

PNML.parse_unparsed — Method

parse_unparsed(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Term

PNML.parse_variable — Method

parse_variable(
    node::EzXML.Node,
    pntd::PnmlType,
    reg::PnmlIDRegistry
) -> PNML.Variable

PNML.pid — Function

pid(x) -> Symbol

Return pnml id symbol. An id's value is identity/unique. A tag may have multiple of same value.

PNML.place — Function

Return the place with id.

PNML.place_idset — Function

Return iterator of all place IDs.

PNML.place_type — Function

Type of Place.

PNML.places — Function

Return iterator of all places.

PNML.pnml_namespace — Method

pnml_namespace(
    node::EzXML.Node;
    missing_ns_fatal,
    default_ns
) -> String

Return namespace. When node does not have a namespace return default value pnml_ns and warn or throw an error.

PNML.pnmlnet_type — Function

PNML.post_flat_verify — Method

Verify a PnmlNet after it has been flattened or is otherwise expected to be a single-page net.

PNML.postset — Method

Iterate ids of output (arc target)) for source transition or place id.

See out_inscriptions and transition_function.

PNML.preset — Method

Iterate ids of input (arc source)) for output transition or place id.

See in_inscriptions and transition_function.

PNML.rate — Method

rate(transition) -> Float64

Return rate value of transition. Missing rate labels are defaulted to zero, or no delay.

PNML.rate_value_type — Function

Return rate value type of argument.

PNML.rates — Method

rates(petrinet::AbstractPetriNet) -> LVector

Return a transition-id labelled vector of rate values.

We allow all PNML nets to be stochastic Petri nets. See rate.

PNML.reachability_graph — Method

PNML.refid — Method

Return the id of the referenced node.

PNML.refplace — Function

Return reference place matching id.

PNML.refplace_idset — Function

refplace_idset(x) -> Set{Symbol} #TODO iterator?

Return reference place pnml ids.

PNML.refplace_type — Function

Tyoe of RefPlace.

PNML.refplaces — Function

Return vector of all reference places.

PNML.reftransition — Function

Return reference transition matching id.

PNML.reftransition_idset — Function

reftransition_idset(x) -> Set{Symbol} #TODO iterator?

Return reference transition pnml ids.

PNML.reftransition_type — Function

Type of RefTransition.

PNML.reftransitions — Function

Return vector of all reference transitions.

PNML.save_config! — Function

save_config!(config::PnmlConfig=CONFIG)

Save a PNML.jl configuration config (by default, PNML) to your LocalPreferences.toml file using Preferences.jl.

The saved preferences will be automatically loaded next time you using PNML

Examples

julia> using PNML

julia> PNML.CONFIG.verbose = true;

julia> PNML.CONFIG.warn_on_unclaimed = true;     # Customize some defaults

julia> PNML.save_config!(PNML.CONFIG); # Will be automatically read next time you `using PNML`

PNML.source — Method

source(arc) -> Symbol

Return identity symbol of source of arc.

PNML.src_arcs — Function

Return arcs that have a source of transition id.

See also all_arcs, tgt_arcs.

PNML.tag — Function

tag(x) -> Symbol

Return tag symbol. A tag may have multiple of same value. An id's value is identity/unique.

PNML.target — Method

target(arc) -> Symbol

Return identity symbol of target of arc.

PNML.term_value_type — Function

Return value type.

PNML.text — Method

Return text field. All labels must have one that may be nothing.

PNML.text_content — Function

Find first :text in vx and return its :content as string.

PNML.tgt_arcs — Function

Return arcs that have a target of transition id.

See also all_arcs, src_arcs.

PNML.transition — Function

PNML.transition_function — Function

transition_function(petrinet::AbstractPetriNet) -> LVector{Symbol, Tuple{LVector,LVector}

Transition function of a Petri Net. Each transition has an input vector and an output vector. Each labelled vector is indexed by the place on the other end of the arc. Values are inscriptions of the arc.

keys are transition ids

values are tuple of input, output labeled vectors,

label is source or target place id - inscription (integer?)

tfun = LVector(
    birth=(LVector(rabbits=1.0), LVector(rabbits=2.0)),
    predation=(LVector(wolves=1.0, rabbits=1.0), LVector(wolves=2.0)),
    death=(LVector(wolves=1.0), LVector()),
)

Vector{Tuple{Dict{Symbol, Number},Dict{Symbol, Number}}

Δ = 3-element LabelledArrays.LArray{Tuple{LabelledArrays.LArray{Float64, 1, Vector{Float64}},
                                          LabelledArrays.LArray{T, 1, D} where {T, D<:AbstractVector{T}}},
                                    1,
                 Vector{Tuple{LabelledArrays.LArray{Float64, 1, Vector{Float64}},
                              LabelledArrays.LArray{T, 1, D}  where {T, D<:AbstractVector{T}}}},
(:birth, :predation, :death)}:

[tid => ([src=>inscription], [tgt=>inscription])]

PNML.transition_idset — Function

PNML.transition_type — Function

Type of TYransition.

PNML.transitions — Function

Return iterator of all transitions.

PNML.type — Method

type(::SortType) -> AbstractSort

Return type of sort object of a Place.

PNML.unparsed_tag — Function

unparsed_tag(
    node::EzXML.Node,
    pntd::PnmlType
) -> Pair{Symbol, Vector{PNML.AnyXmlNode}}
unparsed_tag(
    node::EzXML.Node,
    pntd::PnmlType,
    _::Union{Nothing, PnmlIDRegistry}
) -> Pair{Symbol, Vector{PNML.AnyXmlNode}}

Return tag => AnyXmlNode holding well formed XML tree/forest.

The main use-case is to be wrapped in a PnmlLabel, AnyElement, Term or other.

PNML.value — Function

value(x)

Return value of x. Can be a wrapped value or a derived value.

PNML.value — Method

value(axn::AnyXmlNode) -> Union{Vector{AnyXmlNode}, String, Substring}

Return vector of children or content of the node.

PNML.value — Method

value(m::Marking) -> Union{Int,Float64}

PNML.version — Method

Version of tool for this tool specific information element.

PNML.xmlroot — Method

xmlroot(s::AbstractString) -> EzXML.Node

Parse string s into EzXML node.

PNML.@xml_str — Macro

Parse string into EzXML node.

See xmlroot.

PnmlTypeDefs Reference

PNML.PnmlTypeDefs — Module

Petri Net Type Definition (pntd) URI mapped to PnmlType subtype singleton.

PNML.PnmlTypeDefs.default_pntd_map — Constant

Map from Petri Net Type Definition (pntd) URI to Symbol. Allows multiple strings to map to the same pntd.

There is a companion map pnmltype_map that takes the symbol to a type object.

The URI is a string and may be the full URL of a pntd schema, just the schema file name, or a placeholder for a future schema.

For readability, the 'pntd symbol' should match the name used in the URI with inconvinient characters removed or replaced. For example, '-' is replaced by '_'.

PNML.PnmlTypeDefs.pnmltype_map — Constant

The key Symbols are the supported kinds of Petri Nets.

PNML.PnmlTypeDefs.AbstractContinuousNet — Type

abstract type AbstractContinuousNet <: PnmlType

Uses floating point numbers for markings, inscriptions. Most of the functionality is shared with AbstractPnmlCore. This seperates the

PNML.PnmlTypeDefs.AbstractHLCore — Type

abstract type AbstractHLCore <: PnmlType

Base of High Level Petri Net pntds which add large extensions to PNML core. hlinitialMarking, hlinscription, and defined label structures.

See HLCoreNet, SymmetricNet, PT_HLPNG and others.

PNML.PnmlTypeDefs.AbstractPnmlCore — Type

abstract type AbstractPnmlCore <: PnmlType

Base of token/integer-based Petri Net pntds.

See PnmlCoreNet, PTNet and others.

PNML.PnmlTypeDefs.ContinuousNet — Type

struct ContinuousNet <: AbstractContinuousNet

TODO: Continuous Petri Net

PNML.PnmlTypeDefs.HLCoreNet — Type

struct HLCoreNet <: AbstractHLCore

HLCoreNet can be used for generic high-level nets. We try to implement and test all function at `PnmlCoreNet level, but expect to find use for a concrete type at this level for testing high-level extensions.

PNML.PnmlTypeDefs.HLPNG — Type

struct HLPNG <: AbstractHLCore

High-Level Petri Net Graphs (HLPNGs) are the most intricate High-Level Petri Net schema. It extends SymmetricNet, including with

declarations for sorts and functions (ArbitraryDeclarations)

sorts for Integer, String, and List

PNML.PnmlTypeDefs.PTNet — Type

struct PTNet <: AbstractPnmlCore

Place-Transition Petri Nets add small extensions to core PNML. Integer-valued initialMarking and inscription.

The grammer file is ptnet.pnml so we name it PTNet. Note that 'PT' is often the prefix for XML tags specialized for this net type.

PNML.PnmlTypeDefs.PT_HLPNG — Type

struct PT_HLPNG <: AbstractHLCore

Place-Transition Net in HLCoreNet notation.

PNML.PnmlTypeDefs.PnmlCoreNet — Type

struct PnmlCoreNet <: AbstractPnmlCore

The most minimal concrete Petri Net.

Used to implement and test the core PNML support. Covers the complete graph infrastructure including labels attached to nodes and arcs.

PNML.PnmlTypeDefs.PnmlType — Type

abstract type PnmlType

Abstract root of a dispatch type based on Petri Net Type Definition (pntd).

Each Petri Net Markup Language (PNML) network element will have a single pntd URI as a required 'type' XML attribute. That URI should refer to a RelaxNG schema defining the syntax and semantics of the XML model.

Selected abbreviations, URIs that do not resolve to a valid schema file, are also allowed.

Refer to pntd_symbol and pnmltype for how to get from the URI to a singleton.

PNML.PnmlTypeDefs.SymmetricNet — Type

struct SymmetricNet <: AbstractHLCore

Symmetric Petri Net is the best-worked use case in the primer and ISO specification part 2.

PNML.PnmlTypeDefs.add_nettype! — Method

add_nettype!(
    elements::AbstractDict,
    s::Symbol,
    pntd::PnmlType
)

Add or replace mapping from s to PnmlType singleton pntd.

PNML.PnmlTypeDefs.all_nettypes — Method

Return iterator over PnmlType singletons.

PNML.PnmlTypeDefs.all_nettypes — Method

Return iterator over PnmlType singletons.

PNML.PnmlTypeDefs.iscontinuous — Function

Values are floating point.

PNML.PnmlTypeDefs.isdiscrete — Function

Values are integers.

PNML.PnmlTypeDefs.ishighlevel — Function

Values are many-sorted.

PNML.PnmlTypeDefs.pnmltype — Function

pnmltype(pntd::T, reg)
pnmltype(uri::AbstractString, reg)
pnmltype(s::Symbol; pnmltype_map=pnmltype_map, reg)

Map either a text string or a symbol to a dispatch type singlton.

While that string may be a URI for a pntd, we treat it as a simple string without parsing. The PnmlTypeDefs.pnmltype_map and PnmlTypeDefs.default_pntd_map are both assumed to be correct here.

Unknown or empty uri will map to symbol :pnmlcore. Unknown symbol throws a DomainError exception.

Examples

julia> pnmltype(PnmlCoreNet())
PnmlCoreNet()

julia> pnmltype("nonstandard")
PnmlCoreNet()

julia> pnmltype(:symmetric)
SymmetricNet()

PNML.PnmlTypeDefs.pntd_symbol — Method

pntd_symbol(s::AbstractString) -> Symbol

Map string s to a pntd symbol using default_pntd_map. Any unknown s is mapped to :pnmlcore. Returned symbol is suitable for pnmltype to use to index into pnmltype_map.

Examples

julia> PnmlTypeDefs.pntd_symbol("foo")
:pnmlcore

PnmlIDRegistrys Reference

PNML.PnmlIDRegistrys — Module

Petri Net Markup Language identifier registry.

PNML.PnmlIDRegistrys.PnmlIDRegistry — Type

Holds a set of pnml id symbols and a lock to allow safe reentrancy.

struct PnmlIDRegistry{L}

ids::Set{Symbol}

lk::Any

Base.isempty — Method

isempty(
    idregistry::PnmlIDRegistry{L<:Base.AbstractLock}
) -> Bool

Is the set of id symbols empty?

PNML.PnmlIDRegistrys.isregistered — Method

isregistered(reg::PnmlIDRegistry, s::AbstractString) -> Bool

Return true if s is registered in reg.

PNML.PnmlIDRegistrys.register_id! — Method

register_id!(
    idregistry::PnmlIDRegistry,
    s::AbstractString
) -> Symbol

Register id symbol and return the symbol.

PNML.PnmlIDRegistrys.registry — Function

registry([lock]) -> PnmlIDRegistry

Construct a PNML ID registry using the supplied lock or a ReentrantLock`.

PNML.PnmlIDRegistrys.reset! — Method

reset!(
    idregistry::PnmlIDRegistry{L<:Base.AbstractLock}
) -> Set{Symbol}

Empty the set of id symbols. Use case is unit tests. In normal use it should never be needed.